### 2020-12-18 #2920 @whedon
[PRE REVIEW]: VAST: The Void Analysis Software Toolkit
**Submitting author:** @kadglass (<a href="http://orcid.org/0000-0002-9540-546X">Kelly Douglass</a>)
**Repository:** <a href="https://github.com/DESI-UR/VAST" target ="_blank">https://github.com/DESI-UR/VAST</a>
**Version:** v1.0.0-alpha
