### 2020-12-18 #2920 @kthyng
Hi @kadglass and thanks for your submission. It looks to me like your submission is in scope for JOSS, but light on some of our packaging requirements, like clear documentation for how to use each of the two subpackages. The readme's are a start, and I see references to example scripts, but it is likely that under review these would need to be expanded. The reviewers will use your software and therefore need to be able to understand how to use it, and with as little effort as possible since we are asking them to spend their time reviewing your software.

In the meantime, I will work on recruiting a handling editor, however, I don't expect there to be much movement on reviews until the new year.
