### 2021-01-25 #2920 @dfm
@kadglass Following up on @kthyng's comment above (https://github.com/openjournals/joss-reviews/issues/2920#issuecomment-748178437): this package still needs substantial work on the documentation, packaging, and testing to bring it up to the standards expected for JOSS (see the [JOSS docs](https://joss.readthedocs.io/en/latest/review_criteria.html) for more info). Can you give us a timeline for when you expect to be able to get to this? Thanks!
