### 2021-01-26 #2920 @kadglass
@dfm Absolutely!  We have looked at the JOSS docs, and we thought that we had checked all of the boxes.  Could you give us more detail on what you would like to see?  Thank you!
