### 2021-01-27 #2920 @dfm
@kadglass: The main issues are _documentation_ and _testing_.

1. *Documentation*: As far as I can tell (and as noted by @kthyng) there seems to be no documentation beyond the two readmes in the subpackages. The argument could perhaps be made that this nearly reaches the bare minimum required to check the "Installation instructions" and "Example usage" boxes, but I don't find it to be sufficient. For example, there should be more discussion of what the examples are trying to convey. The "Functionality documentation" component is not satisfied. For a Python package, we would expect documentation generated using something like Sphinx, and we would expect these docs to include installation instructions, tutorials, and (at least automatically generated) API documentation.
2. *Testing*: While, again, there is an extremely limited test (it appears to just check compilation and that the package imports) this is not sufficient to satisfy the testing requirements of a JOSS submission.

Then other components are such as "Community guidelines" and "Functionality/Performance" are also not fully satisfied, but these are less of a major issue at this stage (but they should be addressed eventually!).

If I've missed something in the repo, please point me to it. Otherwise, please take your time adding these components and I'd be happy to answer questions as you go!
