### 2021-01-28 aea2204
Add mask input option, galaxy ID tracking
