### 2021-01-28 1051cee
Fix bug allowing zero-volume cells in zones
