+++ Vsquared/vast/vsquared/classes.py
@@ -126,8 +126,8 @@ class Tesselation:
         self.volumes = vol
         if viz:
             self.vertIDs = reg
-            vecut = np.zeros(len(vol),dtype=bool)
-            vecut[cut] = True
+            vecut = np.zeros(len(vol))
+            vecut[cut] = 1.
             self.vecut = vecut
             self.verts = ver
         print("Triangulating...")
@@ -174,12 +174,17 @@ class Zones:
         lut   = np.zeros(len(vol), dtype=int)
         depth = np.zeros(len(vol), dtype=int)
 
-        zvols = []
-        zcell = []
+        zvols = [0.]
+        zcell = [[]]
 
         print("Building zones...")
 
         for i in range(len(vol)):
+
+            if vol2[i] == 0.:
+                lut[srt[i]] = -1
+                zcell[-1].append(srt[i])
+                continue
 
             ns = nei2[i]
             vs = vol[ns]
@@ -187,9 +192,9 @@ class Zones:
 
             if n == srt[i]:
                 # This cell has the largest volume of its neighbors
-                lut[n] = len(zvols)
-                zcell.append([n])
-                zvols.append(vol[n])
+                lut[n] = len(zvols) - 1
+                zcell.insert(-1,[n])
+                zvols.insert(-1,vol[n])
             else:
                 # This cell is put into its least-dense neighbor's zone
                 lut[srt[i]]   = lut[n]
@@ -212,8 +217,12 @@ class Zones:
         for i in range(len(vol)):
             ns = nei[i]
             z1 = lut[i]
+            if z1 == -1:
+                continue
             for n in ns:
                 z2 = lut[n]
+                if z2 == -1:
+                    continue
                 if z1 != z2:
                     # This neighboring cell is in a different zone
                     if z2 not in zlinks[0][z1]:
@@ -235,8 +244,12 @@ class Zones:
                         vts = vts[[len(vts[vts==v])==2 for v in vts]]
                         #vts = np.unique(vts[vts!=-1])
                         if len(vts)>2:
-                            vcs = (tess.verts[vts].T[0:2]).T
-                            zverts[z1].append((vts[ConvexHull(vcs).vertices]).tolist())
+                            try:
+                                vcs = (tess.verts[vts].T[0:2]).T
+                                zverts[z1].append((vts[ConvexHull(vcs).vertices]).tolist())
+                            except:
+                                vcs = (tess.verts[vts].T[1:3]).T
+                                zverts[z1].append((vts[ConvexHull(vcs).vertices]).tolist())
                             znorms[z1].append([i,n])
         self.zlinks = zlinks
         if viz:
