### 2021-01-31 c0ec9bf
Starting to make variable names consistent
