+++ VoidFinder/vast/voidfinder/_voidfinder.py
@@ -33,7 +33,7 @@ from ._voidfinder_cython import main_algorithm, \
 
 from ._voidfinder_cython_find_next import GalaxyMap, \
                                           Cell_ID_Memory, \
-                                          GalaxyMapCustomDict, \
+                                          ParticleMapCustomDict, \
                                           HoleGridCustomDict, \
                                           NeighborMemory
 
@@ -46,16 +46,6 @@ from ctypes import c_int64, c_double, c_float
 from queue import Empty
 
 from copy import deepcopy
-
-import pickle
-
-from astropy.table import Table
-
-from .table_functions import to_array
-
-import matplotlib
-#matplotlib.use('TkAgg')
-import matplotlib.pyplot as plt
 
 
 
@@ -103,7 +93,6 @@ def _hole_finder(hole_grid_shape,
     hole_center_iter_dist : scalar float
         distance to shift hole centers during iterative void hole growing in 
         Mpc/h
-        
         
     galaxy_map_grid_edge_length : float or None
         edge length in Mpc/h for the secondary grid for finding nearest neighbor 
@@ -159,7 +148,6 @@ def _hole_finder(hole_grid_shape,
         might actually get a checkpoint at say 1,030,000.  If None, disables 
         saving the checkpoint file
         
-    
     use_start_checkpoint : bool
         Whether to attempt looking for a  VoidFinderCheckpoint.h5 file which can 
         be used to restart the VF run
@@ -281,7 +269,6 @@ def _hole_finder(hole_grid_shape,
                                                             max_dist,
                                                             galaxy_coords,
                                                             survey_name,
-                                                            #hole_radial_mask_check_dist,
                                                             save_after=save_after,
                                                             use_start_checkpoint=use_start_checkpoint,
                                                             batch_size=batch_size,
@@ -304,8 +291,8 @@ class CellIDGenerator(object):
     This class generates an array of 'batch_size' (i,j,k) cell IDs from the
     VoidFinder search grid, for the main VoidFinder algorithm to run on.
     
-    To generate the cell IDs, this class stores the shape of the search grid, and 
-    a dictionary of the grid cells which do not need to be searched.
+    To generate the cell IDs, this class stores the shape of the search grid, 
+    and a dictionary of the grid cells which do not need to be searched.
     """
     
     def __init__(self, 
@@ -317,56 +304,60 @@ class CellIDGenerator(object):
         
         self.data = np.empty(4, dtype=np.int64)
         
-        ################################################################################
-        # Store off the shape of the 3D search grid, but also store a 4th value so 
-        # we don't keep having to recalculate the modulus number used for finding the
-        # i value.
-        ################################################################################
+        ########################################################################
+        # Store off the shape of the 3D search grid, but also store a 4th value 
+        # so we do not keep having to recalculate the modulus number used for 
+        # finding the i value.
+        ########################################################################
         self.data[0] = grid_dim_i
         self.data[1] = grid_dim_j
-        self.data[2] = grid_dim_k #also modulus for finding j
-        self.data[3] = grid_dim_j*grid_dim_k #modulus for finding i
+        self.data[2] = grid_dim_k # also modulus for finding j
+        self.data[3] = grid_dim_j*grid_dim_k # modulus for finding i
         
         self.cell_ID_dict = cell_ID_dict
         
+
     def gen_cell_ID_batch(self, start_idx, batch_size, output_array):
         """
         Description
         ===========
         
-        Given the user-provided 'output_array', fill in at most 'batch_size' cell IDs into
-        'output_array'. 
+        Given the user-provided 'output_array', fill in at most 'batch_size' 
+        cell IDs into 'output_array'. 
         
[…]
