### 2021-02-03 64da5b5
Add third ZOBOV pruning method
