### 2021-02-03 9124d38
Fix inclusion of isolated zones
