+++ Documentation/source/V2.rst
@@ -1,0 +1,28 @@
+V\ :sup:`2`
+===========
+
+`vast.Vsquared` is a software package for finding voids based on the ZOBOV 
+(ZOnes Bordering On Voidness) algorithm [@Neyrinck:2007], which grows voids from 
+local minima in the density field.
+
+
+**Voronoi tesselation**
+
+V\ :sup:`2` first produces a Voronoi tessellation of the catalog of large-scale 
+tracers, and the volumes of the Voronoi cells are used to identify local density 
+minima.
+
+**Zones**
+
+Zones are then built from density minima in the distribution of the Voronoi 
+cells using a watershed transform, where each cell is linked to its least dense 
+neighbor.
+
+**Defining the voids**
+
+Finally, voids are formed from these by identifying low-density boundaries 
+between adjacent zones and using them to grow unions of weakly divided zones.  
+This list of voids is then typically pruned to remove void candidates unlikely 
+to be true voids.  `Vsquared` includes several of the different void-pruning 
+methods that exist, including methods from other ZOBOV implementations such as 
+VIDE [@VIDE:2012] and REVOLVER [@REVOLVER:2018].
