### 2021-03-04 68bdbde
Updated VF ReadTheDocs documentation and made necessary adjustments to docstrings
