+++ VoidFinder/scripts/SDSS_VoidFinder_dr7.py
@@ -146,6 +146,7 @@ temp_infile.close()
 
 wall_coords_xyz, field_coords_xyz, hole_grid_shape, coords_min = filter_galaxies(galaxy_data_table,
                                                                                  survey_name,
+                                                                                 out_directory,
                                                                                  dist_limits=dist_limits,
                                                                                  #mag_cut_flag=mag_cut,
                                                                                  #rm_isolated_flag=rm_isolated,
@@ -190,7 +191,7 @@ find_voids(wall_coords_xyz,
            #hole_center_iter_dist=1.0,
            maximal_spheres_filename=out1_filename,
            void_table_filename=out2_filename,
-           potential_voids_filename=survey_name+'potential_voids_list.txt',
+           potential_voids_filename=out_directory + survey_name + 'potential_voids_list.txt',
            num_cpus=num_cpus,
            batch_size=10000,
            verbose=0,
