+++ VoidFinder/vast/voidfinder/voidfinder.py
@@ -46,12 +46,13 @@ RtoD = 180./np.pi
 
 def filter_galaxies(galaxy_table,
                     survey_name, 
-                    mag_cut_flag=True, 
+                    out_directory,
+                    mag_cut=True, 
                     dist_limits=None,
-                    rm_isolated_flag=True,
+                    rm_isolated=True,
                     write_table=True,
                     sep_neighbor=3,
-                    distance_metric='comoving', 
+                    dist_metric='comoving', 
                     h=1.0,
                     hole_grid_edge_length=5.0,
                     magnitude_limit=-20.09,
@@ -60,12 +61,12 @@ def filter_galaxies(galaxy_table,
     A hodge podge of miscellaneous tasks which need to be done to format the data into
     something the main find_voids() function can use.
     
-    1). Optional magnitude cut
-    2). Convert from ra-dec-redshift space into xyz space
-    3). Calculate the hole search grid shape
-    4). Optional remove isolated galaxies by partitioning them into wall (non-isolated)
-            and field (isolated) groups
-    5). Optionally write out the wall and field galaxies to disk
+    1) Optional magnitude cut
+    2) Convert from ra-dec-redshift space into xyz space
+    3) Calculate the hole search grid shape
+    4) Optional remove isolated galaxies by partitioning them into wall (non-isolated)
+       and field (isolated) groups
+    5) Optionally write out the wall and field galaxies to disk
     
     
     Parameters
@@ -79,8 +80,11 @@ def filter_galaxies(galaxy_table,
         
     survey_name : str
         Name of the galxy catalog, string value to prepend or append to output names
+
+    out_directory : string
+        Directory path for output files
         
-    mag_cut_flag : bool
+    mag_cut : bool
         whether or not to cut on magnitude, removing galaxies less than
         magnitude_limit
 
@@ -90,7 +94,7 @@ def filter_galaxies(galaxy_table,
     magnitude_limit : float
         value at which to perform magnitude cut
         
-    rm_isolated_flag : bool
+    rm_isolated : bool
         whether or not to perform Nth neighbor distance calculation, and use it
         to partition the input galaxies into wall and field galaxies
         
@@ -101,7 +105,7 @@ def filter_galaxies(galaxy_table,
     sep_neighbor : int, positive
         if rm_isolated_flag is true, find the Nth galaxy neighbors based on this value
         
-    distance_metric : str
+    dist_metric : str
         Distance metric to use in calculations.  Options are 'comoving' 
         (default; distance dependent on cosmology) and 'redshift' (distance 
         independent of cosmology).
@@ -143,14 +147,14 @@ def filter_galaxies(galaxy_table,
     # Filter based on magnitude and convert from redshift to radius if necessary
     #---------------------------------------------------------------------------
     # Remove faint galaxies
-    if mag_cut_flag:
+    if mag_cut:
         
         galaxy_table = galaxy_table[galaxy_table['rabsmag'] <= magnitude_limit]
 
     # Remove galaxies outside redshift range
     if dist_limits is not None:
 
-        if distance_metric == 'comoving':
+        if dist_metric == 'comoving':
 
             distance_boolean = np.logical_and(galaxy_table['Rgal'] >= dist_limits[0], 
                                               galaxy_table['Rgal'] <= dist_limits[1])
@@ -165,9 +169,7 @@ def filter_galaxies(galaxy_table,
 
 
     # Convert galaxy coordinates to Cartesian
-    coords_xyz = ra_dec_to_xyz(galaxy_table,
-                               distance_metric,
-                               h)
+    coords_xyz = ra_dec_to_xyz(galaxy_table, dist_metric, h)
     ############################################################################
 
     
@@ -184,7 +186,7 @@ def filter_galaxies(galaxy_table,
[…]
