+++ Documentation/source/VoidRender_examples.rst
@@ -1,0 +1,12 @@
+################
+Using VoidRender
+################
+
+
+
+Example script
+==============
+
+Included in the `VAST/VoidFinder` repository (`VAST/VoidFinder/scripts/`) is 
+``visualize_voids.py``, a working example script for how to run `VoidRender` to 
+visualize the voids found with `VoidFinder` in a given catalog.
