### 2021-03-04 b65234f
Removed unusued package imports from example scripts
