+++ Vsquared/scripts/vsquared.py
@@ -1,0 +1,1311 @@
+
+
+
+
+
+
+<!DOCTYPE html>
+<html lang="en" >
+  <head>
+    <meta charset="utf-8">
+  <link rel="dns-prefetch" href="https://github.githubassets.com">
+  <link rel="dns-prefetch" href="https://avatars.githubusercontent.com">
+  <link rel="dns-prefetch" href="https://github-cloud.s3.amazonaws.com">
+  <link rel="dns-prefetch" href="https://user-images.githubusercontent.com/">
+
+
+
+  <link crossorigin="anonymous" media="all" integrity="sha512-/n+kg/Y7wLPWswRTnQ7p4klgJiOIgNxoS9UsKsJGw2mwof0aUh756y7KdN1gMfbOTPt+o1ucHMxxQt6T+/iAwg==" rel="stylesheet" href="https://github.githubassets.com/assets/frameworks-fe7fa483f63bc0b3d6b304539d0ee9e2.css" />
+  <link crossorigin="anonymous" media="all" integrity="sha512-+k1b7b/MsUroDiQAcg/brRFWwLahsFpuoDCkdBhq8Vjlmzcq5HyiOhczzY+jQfMnf/i7qQKS1e7foF91R6dh4A==" rel="stylesheet" href="https://github.githubassets.com/assets/site-fa4d5bedbfccb14ae80e2400720fdbad.css" />
+    <link crossorigin="anonymous" media="all" integrity="sha512-z10XGPflnXYB6UOcdRwlX44Wlf/N3KPFQ3EGbql7fxBn7S8fyjyY1OpfD446bs+gHPWXUWVA957mXDmL0xt30A==" rel="stylesheet" href="https://github.githubassets.com/assets/behaviors-cf5d1718f7e59d7601e9439c751c255f.css" />
+    
+    
+    
+    <link crossorigin="anonymous" media="all" integrity="sha512-tSD1afBYU3wIwXyYDGk9XEHpzh9KNUpufc0Eu+s+8cfiP6TVZUUxEiZci/allIvQfThbqi/CJapWsCf0etmWyA==" rel="stylesheet" href="https://github.githubassets.com/assets/github-b520f569f058537c08c17c980c693d5c.css" />
+
+  <script crossorigin="anonymous" defer="defer" integrity="sha512-8K2vvwbW+6H27Nad5ydg8PA2/aMD/LKq+EiK9s0U0hhVZxCI2tWBsYk9beAtisRw2j+Or5k2/F+6dk02nmj/PA==" type="application/javascript" src="https://github.githubassets.com/assets/environment-f0adafbf.js"></script>
+    <script crossorigin="anonymous" defer="defer" integrity="sha512-WNhznLW8V61RO/6AS3mIcQ0Ik9+Yfd3gQCqpOCmJCWtjDKt2M2tiYbK0fo6UqPme0uzwu0C1tD67gzwNBtPB8Q==" type="application/javascript" src="https://github.githubassets.com/assets/chunk-frameworks-58d8739c.js"></script>
+    <script crossorigin="anonymous" defer="defer" integrity="sha512-fZkql5oygGKJNtU8SUVtnpFjUAzTeMaCA/i8zTkH9nuyMDrLtvsDYDYxlFWegsoftVIciba7HsFSOUJQpb4Y+g==" type="application/javascript" src="https://github.githubassets.com/assets/chunk-vendor-7d992a97.js"></script>
+  
[…]
