+++ Documentation/source/Vsquared_examples.rst
@@ -53,6 +53,125 @@ a finite selection of example configuration files:
    main galaxy sample.  A volume-limited version of this galaxy catalog is 
    provided with the package 
    (``VAST/Vsquared/data/vollim_dr7_cbp_102709.fits``).
+
+See :ref:`V2-config` for details on the configuration file options.
+
+
+
+
+Finding voids
+=============
+
+
+Script
+------
+
+The easiest way to use `VAST.Vsquared` is to use the ``vsquared.py`` script, 
+located in ``VAST/Vsquared/scripts/``:
+
+.. argparse::
+   :filename: ../../Vsquared/scripts/vsquared.py
+   :func: p
+   :prog: python vsquared.py
+
+
+In a Python Shell
+-----------------
+
+Finding voids can also be done in a Python shell, using the 
+``vast.vsquared.zobov.Zobov`` class and its methods:
+
+1. Create a `Zobov` object using the desired configuration file and additional 
+   input parameters::
+
+       newZobov = Zobov("DR7_config.ini")
+   
+   See :ref:`V2-init` for details on the initialization method's arguments.
+2. Apply a void-pruning method to the voids found::
+
+       newZobov.sortVoids()
+   
+   See :ref:`V2-sort` for details on this method's arguments.
+3. Save the results to disk (these methods take no additional arguments)::
+
+       newZobov.saveVoids()
+       newZobov.saveZones()
+       newZobov.preViz() #if intending to visualize results
+
+
+.. _V2-config:
+
+Configuration File Options
+--------------------------
+
+Using `VAST.Vsquared` requires a configuration file with the following options:
+
+.. list-table:: Configuration file options
+   :width: 100%
+   :widths: 25 25 25 25 50
+   :header-rows: 1
+
+   * - Key
+     - Section
+     - Data type
+     - Unit
+     - Comment
+   * - Input Catalog
+     - Paths
+     - string
+     - 
+     - Path to the input data catalog
+   * - Survey Name
+     - Paths
+     - string
+     - 
+     - Survey identifier to use in output file names
+   * - Output Directory
+     - Paths
+     - string
+     - 
+     - Path to the directory where output files will be saved
+   * - H_0
+     - Cosmology
+     - float
+     - (km/s)/Mpc
+     - Hubble constant of the desired cosmology
+   * - Omega_m
+     - Cosmology
+     - float
+     - 
+     - Dimensionless matter density parameter of the desired cosmology
+   * - redshift_min
+     - Settings
+     - float
+     - 
+     - The redshift above which void-finding will be applied
+   * - redshift_max
+     - Settings
+     - float
+     - 
+     - The redshift below which void-finding will be applied
+   * - rabsmag_min
+     - Settings
+     - float
+     - 
+     - The minimum magnitude for a galaxy to be used for void-finding
+   * - radius_min
+     - Settings
+     - float
+     - Mpc/h
+     - The minimum radius for a void candidate to be considered a true void
+   * - nside
+     - Settings
+     - integer
+     - 
+     - The NSIDE parameter used in the HEALPix pixelization of the survey mask; 
+       must be a power of 2
+   * - redshift_step
+     - Settings
+     - float
+     - 
+     - The step size used to create a comoving-distance-to-redshift lookup table 
    
    
    
@@ -108,17 +227,10 @@ Data columns
      - Data type
      - Unit
      - Comment
-   * - Rgal
-     - float
-     - Mpc/h
-     - Comoving distance.  Only used if ``dist_metric`` is set to ``comoving``.  
-       If this column is not provided, and the distance metric is set to 
-       ``comoving``, then the comoving distances will be calculated based on the 
[…]
