### 2021-03-29 cdc4b4d
Fix bugs in documentation
