+++ Vsquared/scripts/vsquared.py
@@ -1,1311 +1,36 @@
+#!/usr/bin/env python
+"""Driver program for Voronoi Voids (V^2) void finder module using the ZOBOV
+algorithm.
+"""
 
+from argparse import ArgumentParser, ArgumentDefaultsHelpFormatter
+from vast.vsquared.zobov import Zobov
 
+p = ArgumentParser(description='Voronoi Voids (V^2) void finder.',
+                   formatter_class=ArgumentDefaultsHelpFormatter)
 
+p.add_argument('-m', '--method', type=int, default=0,
+               help='Void-finding method (0,1,2,3,...)')
+p.add_argument('-v', '--visualize', action='store_true', default=False,
+               help='Enable void visualization.')
+p.add_argument('-w', '--save_intermediate', action='store_true', default=False,
+               help='Save intermediate files in void calculation.')
 
+req = p.add_argument_group('required named arguments')
+req.add_argument('-c', '--config', dest='config_file', required=True, default="DR7_config.ini",
+                 help='V^2 config file (INI format).')
 
+args = p.parse_args()
 
-<!DOCTYPE html>
-<html lang="en" >
-  <head>
-    <meta charset="utf-8">
-  <link rel="dns-prefetch" href="https://github.githubassets.com">
-  <link rel="dns-prefetch" href="https://avatars.githubusercontent.com">
-  <link rel="dns-prefetch" href="https://github-cloud.s3.amazonaws.com">
-  <link rel="dns-prefetch" href="https://user-images.githubusercontent.com/">
+newZobov = Zobov(args.config_file, args.method, 3,
+                 save_intermediate=args.save_intermediate,
+                 visualize=args.visualize)
 
+newZobov.sortVoids(method=args.method)
 
+newZobov.saveVoids()
 
-  <link crossorigin="anonymous" media="all" integrity="sha512-/n+kg/Y7wLPWswRTnQ7p4klgJiOIgNxoS9UsKsJGw2mwof0aUh756y7KdN1gMfbOTPt+o1ucHMxxQt6T+/iAwg==" rel="stylesheet" href="https://github.githubassets.com/assets/frameworks-fe7fa483f63bc0b3d6b304539d0ee9e2.css" />
-  <link crossorigin="anonymous" media="all" integrity="sha512-+k1b7b/MsUroDiQAcg/brRFWwLahsFpuoDCkdBhq8Vjlmzcq5HyiOhczzY+jQfMnf/i7qQKS1e7foF91R6dh4A==" rel="stylesheet" href="https://github.githubassets.com/assets/site-fa4d5bedbfccb14ae80e2400720fdbad.css" />
-    <link crossorigin="anonymous" media="all" integrity="sha512-z10XGPflnXYB6UOcdRwlX44Wlf/N3KPFQ3EGbql7fxBn7S8fyjyY1OpfD446bs+gHPWXUWVA957mXDmL0xt30A==" rel="stylesheet" href="https://github.githubassets.com/assets/behaviors-cf5d1718f7e59d7601e9439c751c255f.css" />
-    
-    
-    
-    <link crossorigin="anonymous" media="all" integrity="sha512-tSD1afBYU3wIwXyYDGk9XEHpzh9KNUpufc0Eu+s+8cfiP6TVZUUxEiZci/allIvQfThbqi/CJapWsCf0etmWyA==" rel="stylesheet" href="https://github.githubassets.com/assets/github-b520f569f058537c08c17c980c693d5c.css" />
+newZobov.saveZones()
 
[…]
