+++ VoidFinder/python/build/lib.linux-x86_64-3.7/voidfinder/hole_combine.py
@@ -1,0 +1,775 @@
+from astropy.table import Table
+
+import numpy as np
+
+from .table_functions import to_array,to_vector
+
+import time
+
+#import warnings
+#warnings.simplefilter('error')
+
+from ._hole_combine_cython import remove_duplicates_2, find_maximals_2, find_holes_2
+
+
+################################################################################
+################################################################################
+
+
+def spherical_cap_volume(radius, height):
+    '''Calculate the volume of a spherical cap'''
+
+    volume = np.pi*(height**2)*(3*radius - height)/3.
+
+    return volume
+
+
+################################################################################
+################################################################################
+
+
+def cap_height(R, r, d):
+    '''Calculate the height of a spherical cap.
+
+    Parameters
+    __________
+
+    R : radius of sphere
+
+    r : radius of other sphere
+
+    d : distance between sphere centers
+
+
+    Output
+    ______
+
+    h : height of cap
+    '''
+    '''for elem in d:
+        if np.isnan(elem) or elem == 0:
+            print(d)
+            print(':(')'''
+    h = (r - R + d)*(r + R - d)/(2*d)
+    
+
+    return h
+
+
+################################################################################
+################################################################################
+
+
+def combine_holes(spheres_table, frac=0.1):
+    '''
+    Combines the potential void spheres into voids.
+
+
+    Parameters:
+    ===========
+
+    spheres_table : astropy table of length N
+        Table of holes found, sorted by radius.  Required columns: radius, x, y, 
+        z (in units of Mpc/h)
+
+    frac : float
+        Fraction of hole volume.  If a hole overlaps a maximal sphere by more 
+        than this fraction, it is not part of a unique void.  Default is 10%.
+
+
+    Returns:
+    ========
+
+    maximal_spheres_table : astropy table of length M
+        Table of maximal spheres (largest hole in each void).  Minimum radius is 
+        10 Mpc/h.  Columns: x, y, z, radius (all in units of Mpc/h), flag 
+        (unique void number).
+
+    holes_table : astropy table of length P
+        Table of all holes that belong to voids.  Columns: x, y, z, radius (all 
+        in units of Mpc/h), flag (void number identifier).
+    '''
+
+
+
+    time_start = time.time()
+
+    ############################################################################
+    # Remove duplicate holes
+    #---------------------------------------------------------------------------
+    unique_spheres_table = remove_duplicates(spheres_table)
+    ############################################################################
+
+    print("remove_duplicates: ", time.time() - time_start)
+
+
+    ############################################################################
+    # Maximal spheres
+    #---------------------------------------------------------------------------
+    time_start = time.time()
+    
+    maximal_spheres_table, maximal_spheres_indices = find_maximals(unique_spheres_table, frac)
+
+    print("find_maximals: ", time.time() - time_start)
+
+    # Array of coordinates for maximal spheres
+    maximal_spheres_coordinates = to_array(maximal_spheres_table)
+
+    # Array of radii for maximal spheres
+    maximal_spheres_radii = np.array(maximal_spheres_table['radius'])
+
+    print('Maximal spheres identified')
+    ############################################################################
+
+
+
+    ############################################################################
+    # Assign spheres to voids
+    #---------------------------------------------------------------------------
+    
+    time_start = time.time()
+    holes_table = find_holes(unique_spheres_table, maximal_spheres_table, maximal_spheres_indices)
[…]
