+++ VoidFinder/outputs/delta_runs/VF_deltas_minradius20.sh
@@ -3,11 +3,11 @@
 #SBATCH --job-name=deltafields_beforenames                                                                                                                                                                 
 #SBATCH --time=01:00:00                                                                                                                                                                                  
 #SBATCH --mail-type=ALL                                                                                                                                                                                    
-#SBATCH -o /scratch/ierez/IGMCosmo/VoidFinder/outputs/delta_runs/before_names.log                                                                                                                          
-#SBATCH -e /scratch/ierez/IGMCosmo/VoidFinder/outputs/delta_runs/before_names.err                                
+#SBATCH -o /scratch/ierez/IGMCosmo/VoidFinder/outputs/delta_runs/before_names_minradius20.log                                                                                                            
+#SBATCH -e /scratch/ierez/IGMCosmo/VoidFinder/outputs/delta_runs/before_names_minradius20.err                                
                                                                                                                               
 
-echo 'Run VoidFinder on delta fields without filter and before name changes.'
+echo 'Run VoidFinder on delta fields without filter and before name changes with minimum maximal sphere raius set to 20 Mpc/h.'
 echo '200g 1d'
 if [ X"$SLURM_STEP_ID" = "X" -a X"$SLURM_PROCID" = "X"0 ]
 then
@@ -20,7 +20,7 @@ hostname
 now=$(date)
 echo "Starting date: $now"
 module load anaconda3
-python /scratch/ierez/IGMCosmo/VoidFinder/python/scripts/VoidFinder_DR16_deltafields_fits.py
+python /scratch/ierez/IGMCosmo/VoidFinder/python/scripts/VoidFinder_DR16_deltafields_fits_minradius20.py
 echo 'Done :)'
 now=$(date)
 echo "Ending date: $now"
