+++ VoidFinder/python/build/lib.linux-x86_64-3.7/voidfinder/preprocessing.py
@@ -1,0 +1,296 @@
+
+import numpy as np
+
+from astropy.table import Table
+
+from astropy.io import ascii
+
+from voidfinder.distance import z_to_comoving_dist
+
+import time
+
+import h5py
+
+from voidfinder.constants import c #speed of light
+
+import os
+
+from astropy.io import fits
+
+def load_data_to_Table(input_filepath):
+    """
+    Description
+    ===========
+    
+    Load a .dat or .h5 file representing a table of numerical values.  For HDF5 (.h5)
+    files, assumes the "columns" are stored as individual datasets on the root h5py.File
+    object.
+    
+    If the input filepath ends with a .h5 extension, this will attempt to read it in 
+    as the HDF5 format, otherwise for all other extensions it will attempt the normal
+    Table.read()
+    
+    Parameters
+    ==========
+    
+    input_filepath : str
+        path to the input .dat or .h5 file
+        
+    Returns
+    =======
+    
+    astropy.table.Table
+    """
+    if input_filepath.endswith(".fits"):
+        
+        data = fits.open(input_filepath)
+        data=data[1].data
+        data_table = Table()
+        ''' 
+        print('Inside load table before columns')
+        RA=Table.Column(data['RA'][0:1000], name='ra')
+        DEC=Table.Column(data['DEC'][0:1000], name='dec')
+        Z=Table.Column(data['Z'][0:1000], name='redshift')
+        DELTA=Table.Column(data['deltas'][0:1000], name='deltas')
+        '''
+        print('Inside load table before columns')                                                     
+        RA=Table.Column(data['ra'], name='ra')                                                
+        DEC=Table.Column(data['dec'], name='dec')                                             
+        Z=Table.Column(data['z'], name='redshift')                                          
+        DELTA=Table.Column(data['delta'], name='delta') 
+        
+        print('Inside load table before adding columns')
+        data_table.add_column(RA)
+        data_table.add_column(DEC)
+        data_table.add_column(Z)
+        data_table.add_column(DELTA)
+        ''' 
+        print('Length of data before removing 0s:')
+        print(len(data_table))
+        data_table=data_table[data_table['deltas']!=0]
+        print('Length of data after removing 0s:')
+        print(len(data_table))
+        '''
+        print('fits to table done:)')
+        
+
+    
+    else:    
+        if input_filepath.endswith(".h5"):
+        
+            infile = h5py.File(input_filepath, 'r')
+        
+            data_table = Table()
+        
+            for col in list(infile.keys()):
+            
+                data_table[col] = infile[col][()]
+            
+                infile.close()
+        
+        else:
+        
+            if os.path.getsize(input_filepath) < 5e9:
+            
+                data_table = Table.read(input_filepath, format='ascii.commented_header')
+
+            else:
+        
+                # Import header line
+                data_table_fobj = open(input_filepath, 'r')
+
+                header_line = data_table_fobj.readline()
+
+                data_table_fobj.close()
+
+                # Parse header line
+                col_names = header_line.split(' ')
+
+                # Read in the data in 100 Mb chunks
+                data_table = ascii.read(input_filepath, 
+                                    format='no_header', 
+                                    names=col_names[1:], 
+                                    guess=False, 
+                                    fast_reader={'chunk_size': 100*1000000})
+    
+    
+    return data_table
+    
+    
+    
+
+
+def file_preprocess(galaxies_filename, 
+                    in_directory, 
+                    out_directory, 
+                    mag_cut=True,
+                    flux_cut=None,
+                    rm_isolated=True,
+                    dist_metric='comoving', 
+                    min_z=None,
+                    max_z=None,
+                    Omega_M=0.3,
[…]
