### 2021-04-08 64fc28e
Added enhancement #41 Cartesian input for VoidFinder ability to utilize xyz-based mask
