+++ VoidFinder/python/voidfinder/voidfinder.py
@@ -13,12 +13,12 @@ import matplotlib.pyplot as plt
 
 from .hole_combine import combine_holes, combine_holes_2
 
-from .voidfinder_functions import mesh_galaxies, \
+from .voidfinder_functions import mesh_data, \
                                   in_mask, \
                                   not_in_mask, \
                                   in_survey, \
                                   save_maximals, \
-                                  mesh_galaxies_dict
+                                  mesh_data_dict
                                   #build_mask, \
 
 from .table_functions import add_row, \
@@ -32,7 +32,7 @@ from .volume_cut import volume_cut, check_hole_bounds
 
 from .avsepcalc import av_sep_calc
 
-from .mag_cutoff_function import mag_cut, field_gal_cut
+from .mag_cutoff_function import mag_cut, field_data_cut
 
 from ._voidfinder import _hole_finder
 
@@ -42,7 +42,7 @@ from ._voidfinder_cython import check_mask_overlap
 
 
 ################################################################################
-# I made these constants into default parameters in filter_galaxies() and 
+# I made these constants into default parameters in filter_data() and 
 # find_voids() because a user may wish to try different grid spacings and such.  
 # I left DtoR and RtoD because they will never ever change since they're based 
 # on pi.
@@ -56,7 +56,7 @@ RtoD = 180./np.pi
 
 
 
-def filter_data(galaxy_table,
+def filter_data(data_table,
                     survey_name, 
                     mag_cut=True,
                     flux_cut=None, 
@@ -78,17 +78,17 @@ def filter_data(galaxy_table,
     1). Optional magnitude cut
     2). Convert from ra-dec-redshift space into xyz space
     3). Calculate the hole search grid shape
-    4). Optional remove isolated galaxies by partitioning them into wall (non-isolated)
+    4). Optional remove isolated data by partitioning them into wall (non-isolated)
             and field (isolated) groups
-    5). Optionally write out the wall and field galaxies to disk
+    5). Optionally write out the wall and field data to disk
     
     
     Parameters
     ==========
     
-    galaxy_table : astropy.table of shape (N,?)
+    data_table : astropy.table of shape (N,?)
         variable number of required columns.  If doing magnitude cut, must include
-        'rabsmag' column. If distance metric is 'comoving', must include 'Rgal'
+        'rabsmag' column. If distance metric is 'comoving', must include 'Rdist'
         column, otherwise must include 'redshift'.  Also must always include 'ra' 
         and 'dec'
         
@@ -106,14 +106,14 @@ def filter_data(galaxy_table,
         
     rm_isolated_flag : bool
         whether or not to perform Nth neighbor distance calculation, and use it
-        to partition the input galaxies into wall and field galaxies
+        to partition the input data into wall and field data
         
     write_table : bool
-        use astropy.table.Table.write to write out the wall and field galaxies
+        use astropy.table.Table.write to write out the wall and field data
         to file
         
     sep_neighbor : int, positive
-        if rm_isolated_flag is true, find the Nth galaxy neighbors based on this value
+        if rm_isolated_flag is true, find the Nth data neighbors based on this value
         
     distance_metric : str
         Distance metric to use in calculations.  Options are 'comoving' 
@@ -134,11 +134,11 @@ def filter_data(galaxy_table,
     Returns
     =======
     
-    wall_gals_xyz : numpy.ndarray of shape (K,3)
-        the galaxies which were designated not to be isolated
[…]
