+++ VoidFinder/python/scripts/VoidFinder_deltafields_DR16S82.py
@@ -1,0 +1,225 @@
+'''VoidFinder - Hoyle & Vogeley (2002)'''
+
+
+
+################################################################################
+#
+# If you have control over your python environment, voidfinder can be installed
+# as a normal python package via 'python setup.py install', in which case the 
+# below import of 'sys' and 'sys.path.insert(0, '/abspath/to/VoidFinder/python'
+# is unnecessary.  If you aren't able to install the voidfinder package,
+# you can use the sys.path.insert to add it to the list of available packages
+# in your python environment.
+#
+# Alternately, "python setup.py develop" will 'install' some symlinks which
+# point back to the current directory and you can run off the same voidfinder
+# repository that you're working on as if it was installed
+#
+################################################################################
+
+
+#import sys
+#sys.path.insert(1, '/home/oneills2/VoidFinder/python/')
+#sys.path.insert(1, '/Users/kellydouglass/Documents/Research/VoidFinder/python/')
+import sys
+sys.path.insert(0, "/scratch/ierez/IGMCosmo/VoidFinder/python/")
+################################################################################
+#
+#   IMPORT MODULES
+#
+################################################################################
+
+from voidfinder import find_voids, filter_data
+
+from voidfinder.multizmask import generate_mask
+from voidfinder.preprocessing import file_preprocess
+from voidfinder.table_functions import to_vector, to_array
+
+from astropy.table import Table
+import pickle
+import numpy as np
+
+
+
+
+################################################################################
+#
+#   USER INPUTS
+#
+################################################################################
+
+
+# Number of CPUs available for analysis.
+# A value of None will use one less than all available CPUs.
+num_cpus = 1
+
+#-------------------------------------------------------------------------------
+survey_name = 'DR16S82_deltafields'
+
+# File header
+#in_directory = '/Users/kellydouglass/Documents/Research/Voids/VoidFinder/data/SDSS/'
+#out_directory = '/Users/kellydouglass/Documents/Research/Voids/VoidFinder/data/SDSS/'
+in_directory = '/scratch/ierez/IGMCosmo/VoidFinder/data/DR16S82_H/delta_fields/'
+out_directory = '/scratch/ierez/IGMCosmo/VoidFinder/outputs/delta_runs/before_merge/'
+
+# Input file name
+#data_filename = 'data.dat'  # File format: RA, dec, redshift, comoving distance, absolute magnitude
+#data_filename = 'mini_data_reconstructed_removed.dat' 
+data_filename = 'deltafields_RAadded90.fits'
+
+#data_type='delta_fields'
+#data_type='galaxies' or 'tomographic_maps'
+#deltas_filename = 'vollim_dr7_cbp_102709.dat'  # File format: RA, dec, redshift, comoving distance, absolute magnitude  
+#-------------------------------------------------------------------------------
+
+#-------------------------------------------------------------------------------
+# Survey parameters
+# Note: These can be set to None, in which case VoidFinder will use the limits 
+# of the data catalog.
+min_z = 2.1
+max_z = 3.2
+
+# Cosmology (uncomment and change values to change cosmology)
+# Need to also uncomment relevent inputs in function calls below
+#Omega_M = 0.3
+Omega_M= 0.3147
+#h = 1
+
+# Uncomment if you do NOT want to use comoving distances
+# Need to also uncomment relevant inputs in function calls below
+#dist_metric = 'redshift'
+#-------------------------------------------------------------------------------
+
+#-------------------------------------------------------------------------------
+# Uncomment if you do NOT want to remove data with Mr > -20
+# Need to also uncomment relevent input in function call below
+mag_cut = False #for galaxies
+flux_cut = None #for delta fields or tomographic maps
+
+# Uncomment if you do NOT want to remove isolated data
+# Need to also uncomment relevent input in function call below
[…]
