### 2021-06-26 1adb24c
Changed VF mask generation to depend on maximum redshift of volume-limited sample
