### 2021-07-21 a70a066
Add Vsquared visualization comments
