### 2021-08-01 cb83acf
Added switch to turn on/off non-empty hole checking; commented out 'install_requires=open('requirements.txt') in setup.py until we can establish >= semantics for package requirement versions instead of == semantics because it can destroy existing good environments
