### 2021-08-11 ac39409
Update config to static/templates folders.
