### 2021-08-11 04c9b8f
Adding .readthedocs.yaml file for configuration
