+++ .readthedocs.yaml
@@ -7,4 +7,4 @@ version: 2
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
-    configuration: docs/config.py
+    configuration: Documentation/source/config.py
