+++ Documentation/source/requirements.txt
@@ -1,0 +1,2 @@
+numpydoc
+sphinx-argparse
