+++ .readthedocs.yaml
@@ -8,3 +8,8 @@ version: 2
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
     configuration: Documentation/source/config.py
+    
+# Python version and requirements
+python:
+    install:
+    - requirements: Documentation/source/requirements.txt
