### 2021-08-11 e21219d
Switched back to sphinx_rtd_theme theme for documentation
