### 2021-08-13 c2c0f00
Update V2 visualization documentation
