### 2021-08-13 a48b7ea
Updated sample VF script to match changes in functions
