### 2021-08-13 6e30950
Added variable redshift column name
