+++ Documentation/source/Vsquared_examples.rst
@@ -26,7 +26,9 @@ A summary checklist for installing and running `VAST.Vsquared`.
    * etc.
 
  * Run ``vsquared.py`` from the ``VAST/Vsquared/scripts`` directory on your 
-   machine or using a cluster.
+   machine or using a cluster::
+   
+   python vsquared.py -c DR7_config.ini
 
 The output files will be located in the directory specified by the Output 
 directory.
