### 2021-08-13 2f13f84
Added xyz loading to load_galaxy_data in voidrender code and made default wall color black
