+++ Documentation/source/Vsquared_examples.rst
@@ -29,6 +29,9 @@ A summary checklist for installing and running `VAST.Vsquared`.
    machine or using a cluster::
    
     python vsquared.py -c DR7_config.ini
+    
+   Note: Include the ``-v`` option to produce the output required for 
+   ``VAST.VoidRender``.
 
 The output files will be located in the directory specified by the Output 
 directory.
