+++ VoidFinder/scripts/VoidFinder_DR16S82_reconstructed_numcpus3.py
@@ -1,0 +1,214 @@
+'''VoidFinder - Hoyle & Vogeley (2002)'''
+
+
+
+################################################################################
+#
+# If you have control over your python environment, voidfinder can be installed
+# as a normal python package via 'python setup.py install', in which case the 
+# below import of 'sys' and 'sys.path.insert(0, '/abspath/to/VoidFinder/python'
+# is unnecessary.  If you aren't able to install the voidfinder package,
+# you can use the sys.path.insert to add it to the list of available packages
+# in your python environment.
+#
+# Alternately, "python setup.py develop" will 'install' some symlinks which
+# point back to the current directory and you can run off the same voidfinder
+# repository that you're working on as if it was installed
+#
+################################################################################
+
+
+import sys
+#sys.path.insert(1, '/home/oneills2/VoidFinder/python/')
+#sys.path.insert(1, '/Users/kellydouglass/Documents/Research/VoidFinder/python/')
+#sys.path.insert(1, '/scratch/ierez/IGMCosmo/VoidFinder/python/')
+
+################################################################################
+#
+#   IMPORT MODULES
+#
+################################################################################
+
+from vast.voidfinder import find_voids, filter_data
+
+from vast.voidfinder.multizmask import generate_mask
+from vast.voidfinder.preprocessing import file_preprocess
+from vast.voidfinder.table_functions import to_vector, to_array
+
+import pickle
+
+
+
+
+
+################################################################################
+#
+#   USER INPUTS
+#
+################################################################################
+
+
+# Number of CPUs available for analysis.
+# A value of None will use one less than all available CPUs.
+num_cpus = 3
+
+#-------------------------------------------------------------------------------
+survey_name = 'SDSS_DR16S82_reconstructedmaps_'
+
+# File header
+# Change these directory paths to where your data is stored, and where you want                                                                                                                                                             
+# the output to be saved.
+in_directory = '/scratch/sbenzvi_lab/boss/dr16/reconstructed_maps/'
+out_directory = '/scratch/ierez/IGMCosmo/VoidFinder/outputs/recons_runs/'
+
+# Input file name
+data_filename = 'map_reconstructed.fits'  # File format: RA, dec, redshift, comoving distance, absolute magnitude
+#-------------------------------------------------------------------------------
+
+#-------------------------------------------------------------------------------
+# Survey parameters
+# Note: These can be set to None, in which case VoidFinder will use the limits 
+# of the galaxy catalog.
+min_z = 1.95
+max_z = 3.15
+
+# Cosmology (uncomment and change values to change cosmology)
+# Need to also uncomment relevent inputs in function calls below
+Omega_M = 0.3147
+#h = 1
+
+# Uncomment if you do NOT want to use comoving distances
+# Need to also uncomment relevent inputs in function calls below
+dist_metric = 'comoving'
+#-------------------------------------------------------------------------------
+
+#-------------------------------------------------------------------------------
+# Uncomment if you do NOT want to remove galaxies with Mr > -20
+# Need to also uncomment relevent input in function calls below
+#mag_cut = False
+
+flux_cut=None
+mag_cut=False
+# Uncomment if you do NOT want to remove isolated galaxies
+# Need to also uncomment relevent input in function calls below
+#rm_isolated = False
+#-------------------------------------------------------------------------------
+
+
+
+
+################################################################################
+#
+#   PREPROCESS DATA
+#
+################################################################################
+
+data_table, dist_limits, out1_filename, out2_filename = file_preprocess(data_filename, 
+                                                                               in_directory, 
+                                                                               out_directory, 
+                                                                               #mag_cut=mag_cut,
+                                                                               flux_cut=flux_cut,
+                                                                               #rm_isolated=rm_isolated,
[…]
