+++ VoidFinder/scripts/data.txt
@@ -1,0 +1,6 @@
+x y z delta
+0 0 0 1
+1 2 2 2
+0 0 1 1
+2 3 1 1
+0 1 0 3

[…]
