+++ Documentation/source/VoidFinder_functions.rst
@@ -7,26 +7,26 @@ VoidFinder API
 ``file_preprocess``
 ===================
 
-.. autofunction:: VAST.VoidFinder.vast.voidfinder.preprocessing.file_preprocess
+.. autofunction:: vast.voidfinder.preprocessing.file_preprocess
 
 
 
 ``filter_galaxies``
 ===================
 
-.. autofunction:: VAST.VoidFinder.vast.voidfinder.filter_galaxies
+.. autofunction:: vast.voidfinder.filter_galaxies
 
 
 ``ra_dec_to_xyz``
 =================
 
-.. autofunction:: VAST.VoidFinder.vast.voidfinder.ra_dec_to_xyz
+.. autofunction:: vast.voidfinder.ra_dec_to_xyz
 
 
 ``z_to_comoving_dist``
 ======================
 
-.. autofunction:: VAST.VoidFinder.vast.voidfinder.distance.z_to_comoving_dist
+.. autofunction:: vast.voidfinder.distance.z_to_comoving_dist
 
 
 
@@ -35,11 +35,11 @@ VoidFinder API
 ``generate_mask``
 =================
 
-.. autofunction:: VAST.VoidFinder.vast.voidfinder.multizmask.generate_mask
+.. autofunction:: vast.voidfinder.multizmask.generate_mask
 
 
 
 ``find_voids``
 ==============
 
-.. autofunction:: VAST.VoidFinder.vast.voidfinder.find_voids
+.. autofunction:: vast.voidfinder.find_voids
