### 2021-08-16 a927f4e
Fixed auto-capitalization of cython in requirements.txt files
