### 2021-08-16 659b8b4
Removed sys.path.insert lines for building docs
