+++ Documentation/source/conf.py
@@ -10,10 +10,10 @@
 # add these directories to sys.path here.  If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 
-import os
-import sys
-sys.path.insert(0, os.path.abspath('../../VoidFinder/'))
-sys.path.insert(1, os.path.abspath('../../Vsquared/'))
+#import os
+#import sys
+#sys.path.insert(0, os.path.abspath('../../VoidFinder/'))
+#sys.path.insert(1, os.path.abspath('../../Vsquared/'))
 
 
 # -- Project information -----------------------------------------------------
