### 2021-08-16 9f72fd8
Readding sys.path.insert lines for building docs
