### 2021-08-16 1e6de3c
Changing sys.path.insert path to point into the package
