+++ Documentation/source/conf.py
@@ -17,8 +17,8 @@ sys.path.insert(1, os.path.abspath('../../Vsquared/'))
 #sys.path.insert(0, os.path.abspath('../../VoidFinder/vast/voidfinder/'))
 #print(sys.path)
 
-import vast.voidfinder as VF
-print(VF.__file__)
+import vast
+print(vast.__file__)
 
 # -- Project information -----------------------------------------------------
 
