### 2021-08-17 3bd4479
Debugging docs - printing everything inside compiled vast.voidfinder file
