### 2021-08-17 0693925
Debugging docs - Fixing typo in directory acquisition
