+++ Documentation/source/index.rst
@@ -2,13 +2,17 @@
    sphinx-quickstart on Sun Jan 31 18:17:57 2021.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
+   
+.. role:: raw-html(raw)
+    :format: html
+
+
 
 Welcome to VAST's documentation!
 ================================
 
-The **Void Analysis Software Toolkit**, or VAST, provides pure Python 
-implementations of two popular classes of void-finding algorithms in galaxy 
-catalogs:
+The **Void Analysis Software Toolkit**, or VAST, provides Python implementations 
+of two popular classes of void-finding algorithms in galaxy catalogs:
 1. Void identification by growing spherical voids.
 2. Void identification using watershed algorithms.
 
@@ -22,7 +26,7 @@ formulating a concrete definition to use in a void-finding algorithm is not
 trivial.  As a result, several different algorithms exist to identify these 
 cosmic underdensities.  The Void Analysis Software Toolkit, or VAST, provides 
 Python 3 implementations of two such algorithms: **VoidFinder** and 
-**V\ :sup:`2`**.
+:raw-html:`<strong>V<sup>2</sup></strong>`.
 
 .. toctree::
    :maxdepth: 2
@@ -34,7 +38,7 @@ Python 3 implementations of two such algorithms: **VoidFinder** and
 
 .. toctree::
    :maxdepth: 2
-   :caption: V\ :sup:`2`
+   :caption: :raw-html:`<strong>V<sup>2</sup></strong>`
    
    Vsquared_intro
    Vsquared_examples
