### 2021-08-17 91502dc
Bolding all references of algorithms
