+++ Documentation/source/VoidRender_examples.rst
@@ -1,3 +1,7 @@
+.. role:: raw-html(raw)
+    :format: html
+
+
 ################
 Using VoidRender
 ################
@@ -6,19 +10,19 @@ Using VoidRender
 Quick start
 ===========
 
-A summary checklist for installing and running `VAST.VoidRender`.
+A summary checklist for installing and running **VoidRender**.
 
  * Clone the `GitHub repository <https://github.com/DESI-UR/VAST>`_
- * Navigate to ``VAST/VoidFender`` (for running `VoidRender` on output from 
-   `VoidFinder`) and/or ``VAST/Vsquared`` (for running `VoidRender` on output 
-   from `Vsquared`) and run::
+ * Navigate to ``VAST/VoidFender`` (for running **VoidRender** on output from 
+   **VoidFinder**) and/or ``VAST/Vsquared`` (for running **VoidRender** on 
+   output from :raw-html:`<strong>V<sup>2</sup></strong>`) and run::
     
     python setup.py install
     
    See :ref:`VF-install` or :ref:`V2-install` for installation options.  NOTE: 
-   The version of `VoidRender` that corresponds to the void-finding algorithm of 
-   choice will be built and installed when that algorithm (`VoidFinder`, for 
-   example) is built and installed.
+   The version of **VoidRender** that corresponds to the void-finding algorithm 
+   of choice will be built and installed when that algorithm (**VoidFinder**, 
+   for example) is built and installed.
    
  * Navigate to the ``scripts`` directory within your chosen void-finding 
    algorithm and modify ``visualize_voids.py`` if appropriate.  Parameters to 
@@ -37,15 +41,16 @@ A summary checklist for installing and running `VAST.VoidRender`.
 Example scripts
 ===============
 
-Included in the `VAST/VoidFinder` repository (`VAST/VoidFinder/scripts/`) is 
-``visualize_voids.py``, a working example script for how to run `VoidRender` to 
-visualize the voids found with `VoidFinder` in a given catalog.
+Included in the **VoidFinder** repository (``VAST/VoidFinder/scripts/``) is 
+``visualize_voids.py``, a working example script for how to run **VoidRender** 
+to visualize the voids found with **VoidFinder** in a given catalog.
 
-Included in the `VAST/Vsquared` repository (`VAST/Vsquared/scripts/`) is 
-``visualize_voids.py``, a working example script for how to run `VoidRender` to 
-visualize the voids found with `Vsquared` in a given catalog.  Note: `Vsquared` 
-must be run with the ``-v`` option to produce output needed for the 
-visualization.
+Included in the :raw-html:`<strong>V<sup>2</sup></strong>` repository 
+(``VAST/Vsquared/scripts/``) is ``visualize_voids.py``, a working example script 
+for how to run **VoidRender** to visualize the voids found with 
+:raw-html:`<strong>V<sup>2</sup></strong>` in a given catalog.  Note: 
+:raw-html:`<strong>V<sup>2</sup></strong>` must be run with the ``-v`` option to 
+produce output needed for the visualization.
 
 
 
@@ -54,7 +59,7 @@ visualization.
 Visualize voids
 ===============
 
-The easiest way to use `VoidRender` is to create a script that
+The easiest way to use **VoidRender** is to create a script that
 
 1. Reads in the output from a **VAST** void-finding algorithm and the 
    corresponding galaxy catalog within which the voids were located.
@@ -68,7 +73,7 @@ options and functions called.
 
 **NOTE:** Due to differences in the void structures found by each of the 
 different void-finding algorithms implemented in **VAST**, there is an 
-implementation of `VoidRender` that corresponds to each algorithm.
[…]
