+++ Documentation/source/VoidFinder_examples.rst
@@ -7,7 +7,7 @@ Using VoidFinder
 Quick start
 ===========
 
-A summary checklist for installing and running `VAST.VoidFinder`.
+A summary checklist for installing and running **VoidFinder**.
 
  * Clone the `GitHub repository <https://github.com/DESI-UR/VAST>`_
  * Navigate to ``VAST/VoidFinder`` and run::
@@ -44,13 +44,13 @@ See :ref:`VF-output` for a detailed description of each of these files.
 Example scripts
 ===============
 
-Included in the `VAST.VoidFinder` repository (``VAST/VoidFinder/scripts/``) are 
+Included in the **VoidFinder** repository (``VAST/VoidFinder/scripts/``) are 
 a finite selection of example scripts:
 
- * ``SDSS_VoidFinder_dr7.py`` will run `VoidFinder` on the SDSS DR7 main galaxy 
-   sample.  A volume-limited version of this galaxy catalog is provided with the 
-   package (``VAST/VoidFinder/vollim_dr7_cbp_102709.dat``).
- * ``classifyEnvironment.py`` takes the output of `VoidFinder` (identified 
+ * ``SDSS_VoidFinder_dr7.py`` will run **VoidFinder** on the SDSS DR7 main 
+   galaxy sample.  A volume-limited version of this galaxy catalog is provided 
+   with the package (``VAST/VoidFinder/vollim_dr7_cbp_102709.dat``).
+ * ``classifyEnvironment.py`` takes the output of **VoidFinder** (identified 
    voids) and determines which objects of a given catalog are within the voids 
    ("void" objects), which are outside the voids ("wall" objects), and which are 
    too close to the survey boundary and/or are outside the survey bounds to be 
@@ -62,10 +62,10 @@ a finite selection of example scripts:
 Finding voids
 =============
 
-The easiest way to use `VAST.VoidFinder` is to create a script that
+The easiest way to use **VoidFinder** is to create a script that
 
 1. Filters and processes the input galaxy catalog so that it is in the correct 
-   format for `VoidFinder`
+   format for **VoidFinder**
 2. Generates a survey mask based on the input galaxy catalog, defining the 
    boundaries within which voids can be found
 3. Finds voids within the galaxy catalog
@@ -82,15 +82,15 @@ Pre-processing the data
 ^^^^^^^^^^^^^^^^^^^^^^^
 
 Generally, the first function that should be called in a script running 
-`VoidFinder` is ``file_preprocess``::
+**VoidFinder** is ``file_preprocess``::
 
     from vast.voidfinder.preprocessing import file_preprocess
     
 This function reads the input data file into memory (as an 
 ``astropy.table.Table`` object), checks the various column names to ensure that 
 they match what is expected in other functions, and creates the file names for 
-the files produced by `VoidFinder`.  While not strictly necessary to run, it 
-encompasses many of the introductory steps necessary to run `VoidFinder`.
+the files produced by **VoidFinder**.  While not strictly necessary to run, it 
+encompasses many of the introductory steps necessary to run **VoidFinder**.
 
 The outputs from ``file_preprocess`` are
 
@@ -162,7 +162,7 @@ The outputs of ``generate_mask`` are
 3. Finding voids
 ----------------
 
-The heart of `VoidFinder`, voids are identified in the wall galaxy sample 
+The heart of **VoidFinder**, voids are identified in the wall galaxy sample 
 outputted from ``filter_galaxies`` in the ``find_voids`` fuunction::
 
     from vast.voidfinder import find_voids
@@ -216,9 +216,8 @@ To start ``find_voids`` from one of these files, set
 Input file
 ----------
 
-As `VAST.VoidFinder` is designed to identify dynamically distinct cosmic voids 
[…]
