+++ Documentation/source/index.rst
@@ -13,6 +13,7 @@ Welcome to VAST's documentation!
 
 The **Void Analysis Software Toolkit**, or VAST, provides Python implementations 
 of two popular classes of void-finding algorithms in galaxy catalogs:
+
 1. Void identification by growing spherical voids.
 2. Void identification using watershed algorithms.
 
@@ -24,9 +25,8 @@ large-scale structure improve constraints on the expansion of the universe as
 compared to using tracers alone.  However, what constitutes a void is vague and 
 formulating a concrete definition to use in a void-finding algorithm is not 
 trivial.  As a result, several different algorithms exist to identify these 
-cosmic underdensities.  The Void Analysis Software Toolkit, or VAST, provides 
-Python 3 implementations of two such algorithms: **VoidFinder** and 
-:raw-html:`<strong>V<sup>2</sup></strong>`.
+cosmic underdensities.  VAST provides Python 3 implementations of two such 
+algorithms: **VoidFinder** and :raw-html:`<strong>V<sup>2</sup></strong>`.
 
 .. toctree::
    :maxdepth: 2
@@ -38,7 +38,7 @@ Python 3 implementations of two such algorithms: **VoidFinder** and
 
 .. toctree::
    :maxdepth: 2
-   :caption: :raw-html:`<strong>V<sup>2</sup></strong>`
+   :caption: V&#178;
    
    Vsquared_intro
    Vsquared_examples

[…]
