### 2021-08-17 61174bd
Updated note syntax and added VoidFinder path to .readthedocs.yaml
