+++ .readthedocs.yaml
@@ -13,3 +13,5 @@ sphinx:
 python:
     install:
     - requirements: Documentation/source/requirements.txt
+    - method: setuptools
+      path: VoidFinder
