### 2021-08-17 4504570
auto-finding packages in VF setup
