### 2021-08-17 fb05acf
Fixing syntax errors in VF setup
