### 2021-08-18 5929a5c
Removed sys.path.insert for VF in conf.py
