### 2021-08-18 6c42ea8
Modified docs to be same for V2 as VF
