+++ VoidFinder/requirements.txt
@@ -1,5 +1,5 @@
 astropy>=4.1
-cython>=0.29.21
+Cython>=0.29.21
 h5py>=3.1.0
 matplotlib>=3.3.3
 numpy>=1.19.4
