+++ Vsquared/requirements.txt
@@ -3,7 +3,7 @@ scikit-learn>=0.23.2
 astropy>=4.1
 pytest>=6.1.2
 vispy>=0.6.5
-cython>=0.29.21
+Cython>=0.29.21
 PyOpenGL>=3.1.5
 PyOpenGL-accelerate>=3.1.5
 healpy>=1.15.0
