### 2021-08-18 a5b95ee
Readding cython, numpy to requirements file in docs
