+++ Documentation/source/requirements.txt
@@ -1,2 +1,4 @@
 numpydoc
 sphinx-argparse
+Cython
+numpy
