### 2021-08-18 a696c39
Added additional .voidfinder to main functions
