### 2021-08-18 d4500bb
Debugging docs: added print statements to conf file
