### 2021-08-18 911461f
Debugging docs: importing and printing vispy
