+++ Documentation/source/conf.py
@@ -20,8 +20,9 @@ for entry in sys.path:
     if 'site-packages' in entry:
         print(entry)
 
-from vast.voidfinder import filter_galaxies
-print('filter_galaxies:', filter_galaxies)
+#from vast.voidfinder import filter_galaxies
+import vispy
+print('vispy:', vispy)
 #print('vf.__file__:', vf.__file__)
 #parent_directory = os.path.dirname(vast.__file__)
 #print(os.listdir(parent_directory + '/voidfinder/'))
