### 2021-08-18 c950fe3
Debugging docs: printing site-packages entries containing vast_voidfinder
