### 2021-08-18 fe0ef59
Debugging docs: switched from setuptools to pip in .yaml file
