+++ .readthedocs.yaml
@@ -13,7 +13,7 @@ sphinx:
 python:
     install:
     - requirements: Documentation/source/requirements.txt
-    - method: setuptools
+    - method: pip
       path: VoidFinder
-    - method: setuptools
+    - method: pip
       path: Vsquared
