### 2021-08-18 c071549
Cleaning up debugging lines in conf file
