+++ Documentation/source/conf.py
@@ -4,29 +4,6 @@
 # list see the documentation:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
 
-# -- Path setup --------------------------------------------------------------
-
-# If extensions (or modules to document with autodoc) are in another directory,
-# add these directories to sys.path here.  If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
-
-import os
-import sys
-#sys.path.insert(0, os.path.abspath('../../VoidFinder/'))
-#sys.path.insert(1, os.path.abspath('../../Vsquared/'))
-#sys.path.insert(0, os.path.abspath('../../VoidFinder/vast/voidfinder/'))
-#print('sys.path:', sys.path)
-for entry in sys.path:
-    if 'site-packages' in entry and 'vast_voidfinder' in entry:
-        print(os.listdir(entry))
-
-#from vast.voidfinder import filter_galaxies
-#import vast.voidfinder as vf
-#print('dir(vf):', dir(vf))
-#print('vars(vf):', vars(vf))
-#parent_directory = os.path.dirname(vast.__file__)
-#print(os.listdir(parent_directory + '/voidfinder/'))
-
 # -- Project information -----------------------------------------------------
 
 project = 'VAST'
