+++ Documentation/source/VoidRender.rst
@@ -7,7 +7,8 @@ Introduction
 ############
 
 .. image:: ../voidfinder_viz.png
-**VoidRender** visualization of the output from SDSS DR7 [@Abazajian:2009].\label{fig:vfviz}
+**VoidRender** visualization of the output from SDSS DR7 
+[Abazajian et al. (2009)]_.\label{fig:vfviz}
 
 In order to aid in assessing the quality of the VoidFinder algorithm, the 
 ``vast.voidfinder.viz`` package includes a **VoidRender** class 
@@ -76,4 +77,4 @@ and installed when either **VoidFinder** or
 
 
 
-
+.. [Abazajian et al. (2009)] Abazajian, K. et al. 2009, ApJS, 182, 543
