### 2021-08-18 05ba5ca
Fixing figure formatting, hyperlink references
