### 2021-08-18 c6b6318
Adjusting VoidRender doc strings, citations in RTD
