+++ Documentation/source/VoidFinder_intro.rst
@@ -3,10 +3,13 @@ Introduction
 ############
 
 **VAST.VoidFinder** is a software package containing a Python 3 implementation 
-of the VoidFinder algorithm of [El-Ad97]_ that is based on the algorithm's 
-Fortran implementation by [Hoyle02]_.  Motivated by the expectation that voids 
-are spherical to first order, this algorithm defines voids as the unions of sets 
-of spheres grown in the underdense regions of the large-scale structure.
+of the VoidFinder algorithm 
+(`El-Ad & Piran (1997) <https://arxiv.org/abs/astro-ph/9702135>`_) that is based 
+on the algorithm's Fortran implementation by 
+`Hoyle & Vogeley (2002) <https://arxiv.org/abs/astro-ph/0109357>`_.  Motivated 
+by the expectation that voids are spherical to first order, this algorithm 
+defines voids as the unions of sets of spheres grown in the underdense regions 
+of the large-scale structure.
 
 The **VoidFinder** directory contains the package, which includes an efficient 
 Multi-Process Cythonized version of VoidFinder.  To import the main 
@@ -119,9 +122,10 @@ are using, add ``--user`` to your choice of build from above.  For example::
 Citation
 ========
 
-Please cite [Hoyle02]_ and [El-Ad97]_ when using this algorithm.
+Please cite `Hoyle & Vogeley (2002) <https://arxiv.org/abs/astro-ph/0109357>`_ 
+and `El-Ad & Piran (1997) <https://arxiv.org/abs/astro-ph/9702135>`_ when using 
+this algorithm.
 
 
 
-.. [El-Ad97] El-Ad, H. & Piran, T. 1997, ApJ, 491, 421
-.. [Hoyle02] Hoyle, F. & Vogeley, M. 2002, ApJ, 566, 641
+
