+++ Documentation/source/Vsquared_intro.rst
@@ -9,8 +9,9 @@ Introduction
 
 :raw-html:`<strong>V<sup>2</sup></strong>` is a software package for finding 
 voids based on the ZOBOV (ZOnes Bordering On Voidness) algorithm by 
-[Neyrinck07]_, which uses the gradient of the volume of adjacent voronoi cells 
-to flow multiple cells into larger void regions.
+`Neyrinck (2008) <https://arxiv.org/abs/0712.3049>`_, which uses the gradient of 
+the volume of adjacent voronoi cells to flow multiple cells into larger void 
+regions.
 
 The ``VAST/Vsquared`` directory contains the package, as well as an example 
 script for automatically running all steps of the algorithm. To import the main 
@@ -108,8 +109,8 @@ your choice of build from above.  For example::
 Citation
 ========
 
-Please cite [Neyrinck07]_ when using this algorithm.
+Please cite `Neyrinck (2008) <https://arxiv.org/abs/0712.3049>`_ when using this 
+algorithm.
 
 
 
-.. [Neyrinck (2007)] Neyrinck, M. 2007, MNRAS, 386, 2101
