### 2021-08-18 e1fd685
Removed () from EdAd citation, added figure numbers to captions
