### 2021-08-18 43445e0
Fixing greek letter rendering, V2 figure caption
