### 2021-08-18 98eefa1
Fixing V2 figure caption
