### 2021-08-18 74af32b
Fixing V2 figure caption
