+++ Documentation/source/VoidRender.rst
@@ -51,10 +51,10 @@ environment includes the ``ffmpeg`` library.
 .. _fig-v2viz:
 
 .. figure:: ../vsquared_viz.png
-   :alt: V\ :sup:`2` visualization of the output from SDSS DR7 with VoidRender
-
-   **Figure 2:** V2 visualization of the 
-   output from `SDSS DR7 <https://arxiv.org/abs/0812.0649>`_.
+   :alt: V2 visualization of the output from SDSS DR7 with VoidRender
+   
+   **Figure 2:** **V2** visualization of the output from 
+   `SDSS DR7 <https://arxiv.org/abs/0812.0649>`_.
 
 :raw-html:`<strong>V<sup>2</sup></strong>` also includes an ``OpenGL`` and 
 ``vispy`` based visualization for its output.  The surfaces of voids found by 
