### 2021-08-19 7f6fca7
Removing reference to V2 figure
