### 2021-08-19 526ab21
Minor changes to sample VF scripts
