+++ VoidFinder/scripts/visualize_voids.py
@@ -78,7 +78,7 @@ viz = VoidRender(holes_xyz=holes_xyz,
                  holes_group_IDs=holes_flags,
                  galaxy_xyz=field_galaxy_data,
                  wall_galaxy_xyz=wall_galaxy_data,
-                 wall_distance=3.24,
+                 wall_distance=5.955,
                  galaxy_display_radius=10,
                  remove_void_intersects=2,
                  #void_hole_color=np.array([0.0, 0.0, 1.0, 1.0], dtype=np.float32),
