+++ VoidFinder/scripts/classifyEnvironment.py
@@ -38,7 +38,7 @@ dist_metric = 'comoving'
 
 #-------------------------------------------------------------------------------
 # SURVEY MASK FILE
-mask_filename = '../SDSS_dr7_mask.pickle'
+mask_filename = '../vollim_dr7_cbp_102709_mask.pickle'
 #-------------------------------------------------------------------------------
 
 #-------------------------------------------------------------------------------
