### 2021-08-19 015c652
Added fits file option in for VF input in documentation
