### 2021-08-19 777c427
Incorporate sphinx documentation into main branch. (#40)
* Initializing source files for ReadTheDocs documentation

* Updated VF ReadTheDocs documentation and made necessary adjustments to docstrings

* Removed unusued package imports from example scripts

* Add to Vsquared documentation

* Fix bugs in documentation

* VoidRender documentation additions

* Changed VF mask generation to depend on maximum redshift of volume-limited sample

* Add Vsquared visualization comments

* Finished first draft of online docs

* Add static and templates folders.

* Update config to static/templates folders
[…]
