+++ Vsquared/vast/vsquared/viz/__init__.py
@@ -1,6 +1,6 @@
 
 """
-Visualization module for VoidFinder results.
+Visualization module for Vsquared results.
 
 Requires OpenGL >= 1.2
 vispy >= 0.6.3
@@ -9,7 +9,7 @@ optionally requires ffmpeg available on PATH for video recording
 Usage:
 ------
 
-from vast.voidfinder.viz import VoidRender, load_hole_data, load_galaxy_data
+from vast.vsquared.viz import VoidRender, load_hole_data, load_galaxy_data
 
 holes_xyz, holes_radii, holes_flags = load_hole_data("vollim_dr7_cbp_102709_holes.txt")
 
