+++ Documentation/source/Vsquared_functions.rst
@@ -1,0 +1,91 @@
+####################
+V\ :sup:`2` API
+####################
+
+
+
+.. _V2-init:
+
+``Zobov.__init__``
+==================
+
+.. autofunction:: vast.vsquared.zobov.Zobov.__init__
+
+
+
+.. _V2-sort:
+
+``Zobov.sortVoids``
+===================
+
+.. autofunction:: vast.vsquared.zobov.Zobov.sortVoids
+
+
+
+``Zobov.saveVoids``
+===================
+
+.. autofunction:: vast.vsquared.zobov.Zobov.saveVoids
+
+
+
+``Zobov.saveZones``
+===================
+
+.. autofunction:: vast.vsquared.zobov.Zobov.saveZones
+
+
+
+``Zobov.preViz``
+================
+
+.. autofunction:: vast.vsquared.zobov.Zobov.preViz
+
+
+
+``util.toCoord``
+================
+
+.. autofunction:: vast.vsquared.util.toCoord
+
+
+
+``util.toSky``
+==============
+
+.. autofunction:: vast.vsquared.util.toSky
+
+
+
+``util.inSphere``
+=================
+
+.. autofunction:: vast.vsquared.util.inSphere
+
+
+
+``util.wCen``
+=============
+
+.. autofunction:: vast.vsquared.util.wCen
+
+
+
+``util.getSMA``
+===============
+
+.. autofunction:: vast.vsquared.util.getSMA
+
+
+
+``util.P``
+==========
+
+.. autofunction:: vast.vsquared.util.P
+
+
+
+``util.flatten``
+================
+
+.. autofunction:: vast.vsquared.util.flatten

[…]
