### 2021-08-20 9ff5296
Propagated minimum maximal radius option out to find_voids
