+++ VoidFinder/vast/voidfinder/tests/vf_test.py
@@ -1,0 +1,228 @@
+'''
+Test various functions in VoidFinder by running through a pre-set fake galaxy 
+list with artificial voids removed.
+'''
+
+
+################################################################################
+# Import modules
+#
+# All modules from vast.voidfinder imported here are going to be tested below.
+#-------------------------------------------------------------------------------
+import numpy as np
+
+from astropy.table import Table, setdiff, vstack
+
+from sklearn import neighbors
+
+from vast.voidfinder.constants import c
+
+from vast.voidfinder import find_voids, filter_galaxies
+
+from vast.voidfinder.multizmask import generate_mask
+from vast.voidfinder.preprocessing import file_preprocess
+################################################################################
+
+
+
+
+################################################################################
+# Generate our set of "fake" galaxies and save to disk
+#-------------------------------------------------------------------------------
+ra_range = np.arange(10, 30, 0.5)
+dec_range = np.arange(-10, 10, 0.5)
+redshift_range = np.arange(0, 0.011, 0.0005)
+
+RA, DEC, REDSHIFT = np.meshgrid(ra_range, dec_range, redshift_range)
+
+galaxies_table = Table()
+galaxies_table['ra'] = np.ravel(RA)
+galaxies_table['dec'] = np.ravel(DEC)
+galaxies_table['redshift'] = np.ravel(REDSHIFT)
+
+# Shuffle the table (so that the KDtree does not die)
+rng = np.random.default_rng()
+galaxies_shuffled = Table(rng.permutation(galaxies_table))
+
+galaxies_shuffled['Rgal'] = c*galaxies_shuffled['redshift']/100.
+
+N_galaxies = len(galaxies_shuffled)
+
+# All galaxies will be brighter than the magnitude limit, so that none of them 
+# are removed
+galaxies_shuffled['rabsmag'] = 5*np.random.rand(N_galaxies) - 25.1
+
+galaxies_filename = 'test_galaxies.txt'
+galaxies_shuffled.write(galaxies_filename, 
+                        format='ascii.commented_header', 
+                        overwrite=True)
+
+gal = np.zeros((N_galaxies,3))
+gal[:,0] = galaxies_shuffled['Rgal']*np.cos(galaxies_shuffled['ra']*np.pi/180.)*np.cos(galaxies_shuffled['dec']*np.pi/180.)
+gal[:,1] = galaxies_shuffled['Rgal']*np.sin(galaxies_shuffled['ra']*np.pi/180.)*np.cos(galaxies_shuffled['dec']*np.pi/180.)
+gal[:,2] = galaxies_shuffled['Rgal']*np.sin(galaxies_shuffled['dec']*np.pi/180.)
+################################################################################
+
+
+
+
+################################################################################
+# Test vast.voidfinder.preprocessing.file_preprocess
+# 
+# This function takes in a file name (pointing to the galaxy data file) and 
+# returns the data table.  It also computes the redshift range in comoving 
+# coordinates, and generates the output filenames.
+#-------------------------------------------------------------------------------
+f_galaxy_table, f_dist_limits, f_out1_filename, f_out2_filename = file_preprocess(galaxies_filename, 
+                                                                                  '', 
+                                                                                  '', 
+                                                                                  dist_metric='redshift')
+
+# Check the galaxy table
+assert(len(setdiff(f_galaxy_table, galaxies_shuffled)) == 0)
+
+# Check the distance limits
+dist_limits = np.zeros(2)
+dist_limits[1] = c*redshift_range[-1]/100.
+assert(np.isclose(f_dist_limits, dist_limits).all())
+
+# Check the first output file name
[…]
