### 2021-08-27 7f1188b
Updates to find_voids test
