+++ Vsquared/data/test_config.ini
@@ -1,0 +1,23 @@
+[Paths]
+#Currently only accepts FITS input
+Input Catalog    = ../../data/test_data.fits
+Survey Name      = TEST
+Output Directory = ../../data/
+
+[Cosmology]
+#Currently only uses Flat LambdaCDM
+H_0     = 100
+Omega_m = 0.3
+
+[Settings]
+#Redshift limits
+redshift_min  = 0.03
+redshift_max  = 0.1
+#Magnitude limit, can be None
+rabsmag_min   = None
+#Minimum void radius (Mpc/h)
+radius_min    = 10.0
+#NSIDE for HEALPix pixelization
+nside         = 16
+#Step size for comoving-distance-to-redshift lookup table
+redshift_step = 5e-5
