### 2021-11-19 c3c234f
Updates to CircleCI config file
