### 2021-11-19 b6608a6
Adding test truth files
