### 2021-11-19 aab31c5
Prepended directory structure to test truth tables
