### 2021-11-22 25f9449
Tests (#49)
* Propagated minimum maximal radius option out to find_voids

* Adding VF tests

* Updates to find_voids test

* Updates to CircleCI config file

* Update config.yml

* Adding test truth files

* Prepended directory structure to test truth tables
