### 2022-01-05 #4033 @whedon
[REVIEW]: VAST: The Void Analysis Software Toolkit
**Submitting author:** @kadglass (<a href="http://orcid.org/0000-0002-9540-546X">Kelly Douglass</a>)
**Repository:** https://github.com/DESI-UR/VAST
**Branch with paper.md** (empty if default branch):
**Version:** 1.2.0
