### 2022-01-20 #4033 @changhoonhahn
@dfm I have a question regarding the conflict of interest. I am in the same collaboration (the Dark Energy Spectroscopic Instrument collaboration) as two of the authors: Kelly A. Douglass and Segev BenZvi. Is this a COI based on:
> recent (past four years) collaborators in funded research
I have not published any works with these authors yet, but will likely be co-authors in the near future. Does this count as a COI?
