### 2022-01-21 b6a3d57
First implementation for delta files
