+++ VoidFinder/vast/voidfinder/tests/test_init.py
@@ -1,4 +1,1 @@
-from vast import voidfinder as voidfinder
 
-def test_version_exists():
-    assert hasattr(voidfinder, '__version__')
