+++ VoidFinder/scripts/SDSS_VoidFinder_dr7.py
@@ -56,8 +56,8 @@ survey_name = 'vollim_dr7_cbp_102709_'
 # File header
 # Change these directory paths to where your data is stored, and where you want 
 # the output to be saved.
-in_directory = '../'
-out_directory = '../'
+in_directory = '/global/homes/s/schampat/Voids/VAST/VoidFinder/'
+out_directory = '/global/homes/s/schampat/Voids/Voidfinder_Output/SDSS Output/'
 
 
 # Input file name
