### 2022-01-21 #4033 @dfm
Good question, @changhoonhahn!

With this type of apparent conflict we are generally happy to proceed after disclosure, as long as the authors and reviewers feel confident that the review can be impartial. If not, that's no worries - I'm happy to recruit a different reviewer to fill in for you.

@changhoonhahn & @kadglass: can you let me know here or via email what your stance on this potential COI is? Based on that we can decide how to proceed. Thanks, both!
