### 2022-01-27 DESI-UR/VAST#50 @lavaux
Merge VoidFinder and Vsquared python package
The two components of VAST are installable as two packages. This effectively prevents the use of `pip install` directly on the repository such as:
```
pip install git+https://github.com/DESI-UR/VAST
```
and thus does not follow common practice. It also slightly confuses a user to have two pip installable packages that fills up the namespace `vast` though through two different subpackages.

Could the files `setup.py` be adjusted so that there is only one needed at the top of the source tree to satisfy this constraint?

Reviewing https://github.com/openjournals/joss-reviews/issues/4033
