### 2022-01-27 DESI-UR/VAST#51 @QuiteAFoxtrot
@kadglass @dveyrat @sybenzvi have any of you guys added anything that uses pyopengltk?  The 3D VoidRender stuff uses OpenGL, PyOpenGL, and Vispy, but I have never encountered pyopengltk unless maybe pip pulled the wool over my eyes and something is using it that we don't know about
