### 2022-02-01 DESI-UR/VAST#52 @changhoonhahn
comments on paper
openjournals/joss-reviews#4033

# major comments
- I think the "statement of need" for the paper needs to be sharpened. The speed of void finders is cited as the main problem the software is designed to solve at the moment. However, the paper does not state whether the Python3 implementations of VoidFinder and Vsquared are faster than the previous versions.
- Also, I think the target audience could be more clearly stated. Is the target audience cosmologists who are interested in constraining cosmological parameters using voids? If so, some citations to void forecasts, for instance, would be helpful.

# minor comments
- VoidFinder Section: I think some of the details are a bit vague. For instance, it's not clear to me how "the centers of empty cells" are defined or the "remaining holes".
- V2 Section: It would be helpful if a brief summary of one of the void pruning algorithm (VIDE or REVOLVER) is provided.
- Figure 1 captions: A more detailed description would be very helpful. For example, what is the difference between the black and red points? How are the black points connected?
