### 2022-02-02 72b1721
Updated paper based on comments in Issue #52
