### 2022-02-02 DESI-UR/VAST#52 @kadglass
@dveyrat I have updated the `paper.md` file on the docs branch based on some of these comments.  Could you please add a short description of one of the void pruning methods to the V2 section?
