### 2022-02-02 DESI-UR/VAST#50 @kadglass
We purposefully set up VAST so that each void-finding method could be installed separately, giving the user the flexibility to only install one if preferred.  If this is too confusing, or if this ability to install the package using this particular method of `pip install` is necessary, then we can reconsider the VAST package structure.
