### 2022-02-02 DESI-UR/VAST#52 @sybenzvi
@kadglass, your rewrite of the VoidFinder description is good. But I don't recall you profiling the FORTRAN version... I would have remembered 36 hours (!) to run on DR7.

About the ZOBOV packages, the initial motivation to write V2 was that @dveyrat found ZOBOV implementations in the wild to be abandonware or very difficult to run and install. (REVOLVER is an exception.) I wouldn't put that into the statement of need, but the statement could emphasize that V2 consolidates a large number of pruning methods in a single package.
