### 2022-02-02 DESI-UR/VAST#50 @sybenzvi
We did consider setting up VAST as a single package but ultimately settled on separate build systems that could then be set up as separate packages in pip:
- `pip install vast-voidfinder`
- `pip install vast-vsquared`
- ...

But I think it's not well documented, and we may have set ourselves up for more difficult maintenance in the future. Merging it into one package would require restructuring the repository, though I suspect it won't be a huge amount of work; probably a weekend afternoon. I'd be in favor of this.
