### 2022-02-02 DESI-UR/VAST#52 @kadglass
Michaela ran the Fortran code when we first started working on this project - that is how long it took her to run it!  I can see if I can compile and run the Fortran code again to confirm that runtime.

Good idea on V2 - I will add something about that to the statement.
