### 2022-02-02 b417d7f
Updates related to Issue #52
