### 2022-02-03 DESI-UR/VAST#50 @lavaux
Indeed for clarity and long term maintenance I would be in favor of the merging this into one package.
