### 2022-02-04 836ecd7
Updated paper based on comments in Issue #52
