+++ python/vast/voidfinder/_voidfinder_cython.pyx
@@ -1,0 +1,1388 @@
+
+
+
+
+
+from __future__ import print_function
+
+cimport cython
+
+import numpy as np
+
+cimport numpy as np
+
+np.import_array()  # required in order to use C-API
+
+
+from typedefs cimport DTYPE_CP128_t, \
+                      DTYPE_CP64_t, \
+                      DTYPE_F64_t, \
+                      DTYPE_F32_t, \
+                      DTYPE_B_t, \
+                      ITYPE_t, \
+                      DTYPE_INT32_t, \
+                      DTYPE_INT64_t, \
+                      DTYPE_INT8_t
+
+from numpy.math cimport NAN, INFINITY
+
+from libc.math cimport fabs, sqrt, asin, atan#, exp, pow, cos, sin, asin
+
+from _voidfinder_cython_find_next cimport find_next_galaxy, \
+                                          not_in_mask, \
+                                          _query_first, \
+                                          DistIdxPair, \
+                                          Cell_ID_Memory, \
+                                          GalaxyMapCustomDict, \
+                                          HoleGridCustomDict, \
+                                          FindNextReturnVal, \
+                                          NeighborMemory
+
+import time
+
+
+
+
+
+
+@cython.boundscheck(False)
+@cython.wraparound(False)
+@cython.cdivision(True)
+cpdef DTYPE_INT64_t fill_ijk_zig_zag(DTYPE_INT64_t[:,:] i_j_k_array,
+                                     DTYPE_INT64_t start_idx,
+                                     DTYPE_INT64_t batch_size,
+                                     DTYPE_INT64_t i_lim,
+                                     DTYPE_INT64_t j_lim,
+                                     DTYPE_INT64_t k_lim,
+                                     HoleGridCustomDict cell_ID_dict
+                                     ):
+    """
+    Description
+    ===========
+    
+    Given a starting index and a batch size, generate and fill in the (i,j,k) coordinates
+    into the provided i_j_k_array.
+    
+    This version of fill_ijk goes in zigzag/snake order in an attempt to preserve spatial locality of
+    the array data while processing.  That is to say, on a 3x3 grid, this functions goes:
+    (0,0,0), (0,0,1), (0,0,2), (0,1,2), (0,1,1), (0,1,0), (0,2,0), ...
+    whaeras a normal ordering would be:
+    (0,0,0), (0,0,1), (0,0,2), (0,1,0), (0,1,1), (0,1,2), (0,2,0)
+    so that the next cell grid returned is always adjacent to both its predecessor and
+    its antecedent
+    
+    Also filter out cell IDs which we don't want to grow a hole in (since there are galaxies
+    there) using the cell_ID_dict.
+    
+    
+    
+    Parameters
+    ==========
+    
+    i_j_k_array : memoryview onto a numpy array of shape (N, 3) where N is >= batch_size
+        the array to fill in result values into
+        
+    start_idx : int
+        the sequential index into the grid to start at
+        
+    batch_size : int
+        the maximum number of (i,j,k) triplets to write
+    
+    i_lim, j_lim, k_lim : int
+        the number of cells in each i,j,k direction of the grid we're searching
+        
+    cell_ID_dict : custom dictionary/hash table
+        provides a contains() method that can check if a grid cell is in the
+        galaxy map
+        
+    Output
+    ======
+    
+    Writes i-j-k triplets into the i_j_k_array memory
+    
+    
+    Notes
+    =====
+    We are given a block of batch_size indices to attempt to put
+    into the ijk array.  In order to stay synced with the main
+    process-shared counter, we can ONLY return indices from this block
+    of batch_size.  This means we have 'batch_size' chances to generate
+    a cell ID, NOT that we have to generate until we get 'batch_size'
+    results.
+    
[…]
