+++ python/vast/voidfinder/volume_cut.py
@@ -1,0 +1,1446 @@
+#imports 
+
+import numpy as np
+import os
+import sys
+import mmap
+import struct
+import socket
+import select
+import atexit
+import signal
+import tempfile
+import multiprocessing
+from psutil import cpu_count
+from astropy.table import Table
+
+
+from multiprocessing import Queue, Process, RLock, Value, Array
+from ctypes import c_int64, c_double, c_float
+
+
+from .voidfinder_functions import in_mask, not_in_mask
+from .hole_combine import spherical_cap_volume
+from ._voidfinder_cython_find_next import not_in_mask as nim_cython
+from ._vol_cut_cython import _check_holes_mask_overlap, _check_holes_mask_overlap_2
+from ._voidfinder import process_message_buffer
+
+import time
+
+
+# function to find which spheres stick out of the mask
+def max_range_check(spheres_table, direction, sign, survey_mask, mask_resolution, r_limits):
+    '''
+    Given the list of potential hole locations and their radii in spheres_table,
+    and an axes x,y,z and direction +/-, add the radii of each hole to the hole
+    location and check if that location is within the mask.
+    
+    Returns a boolean array of length N where True indicates the location is valid.
+    '''
+
+    #print("Max Range Check", direction, sign, "hole_table ID: ", id(spheres_table))
+    #print(spheres_table['x'][0])
+
+
+
+    if sign == '+':
+       spheres_table[direction] += spheres_table['radius']
+    else:
+       spheres_table[direction] -= spheres_table['radius']
+       
+       
+       
+    #print(spheres_table['x'][0])
+    #print(spheres_table)
+
+    boolean = in_mask(spheres_table, survey_mask, mask_resolution, r_limits)
+
+    return boolean
+
+
+
+
+
+def check_coordinates(coord, direction, sign, survey_mask, mask_resolution, r_limits):
+
+    dr = 0
+    check_coord = coord
+    #mask_check = True
+    mask_check2 = False
+    #mask_check3 = False
+    
+    #print(id(check_coord), id(coord))
+    
+    np_check_coord = np.empty((1,3), dtype=np.float64)
+    np_check_coord[0,0] = coord['x']
+    np_check_coord[0,1] = coord['y']
+    np_check_coord[0,2] = coord['z']
+    
+    if direction == 'x':
+        np_dir = 0
+    elif direction == 'y':
+        np_dir = 1
+    elif direction == 'z':
+        np_dir = 2
+    
+    #out_log = open("VF_DEBUG_volume_cut.txt", 'a')
+
+    #while dr < coord['radius'] and mask_check:
+    while dr < coord['radius'] and not mask_check2:
+
+        dr += 1
+
+        if sign == '+':
+        #    check_coord[direction] = coord[direction] + dr
+            np_check_coord[0,np_dir] = np_check_coord[0,np_dir] + dr
+        else:
+        #    check_coord[direction] = coord[direction] - dr
+            np_check_coord[0,np_dir] = np_check_coord[0,np_dir] - dr
+
+        #mask_check = in_mask(check_coord, survey_mask, mask_resolution, r_limits)
+        
+        mask_check2 = nim_cython(np_check_coord, survey_mask, mask_resolution, r_limits[0], r_limits[1])
+        
+        #mask_check3 = not_in_mask(np_check_coord, survey_mask, mask_resolution, r_limits[0], r_limits[1])
+        
+        #if mask_check == mask_check3: # or \
+        #   mask_check != mask_check3 or \
+        #if mask_check2 != mask_check3:
+            #out_log.write(str(check_coord)+"\n")
[…]
