### 2022-02-05 e5d2f93
Add tests, visualization, and legacy code.
