+++ python/vast/voidfinder/legacy/deprecated_hole_mask_intersection.py
@@ -1,0 +1,285 @@
+
+
+#
+# This strategy was briefly used in VoidFinder to check whether a given hole overlapped
+# too much outside of the survey mask in order to discard it.  We refactored this process
+# into a monte-carlo method which replaced this strategy.
+#
+
+
+
+
+
+############################################################################
+# Calculate the radial value at which we need to check a finished hole
+# for overlap with the mask
+#
+# We're assuming axis-aligned intersection only, so this calculation uses
+# the area of a "spherical cap" https://en.wikipedia.org/wiki/Spherical_cap
+# We set the area of the spherical cap equal to some fraction p times the
+# area of the whole sphere, and if that volume of the sphere is in the mask
+# then we discard the current hole.  Instead of calculating the actual 
+# volume at any point, instead we calculate the percentage of the radius
+# at which that volume is achieved, which is the same for every sphere, then
+# we can check the 6 directions from the center of a hole as a proxy for
+# an actual 10% volume overlap.  This is a good-enough approximation, since
+# the mask is already composed of cubic cells anyway.
+#
+#
+# let l = r - Y, where r is the radius of a sphere, and Y is the distance
+# along the radius at the volume we care about
+# V_cap = pi*l*l*(3r-l)/3 
+# V_sphere = (4/3)*pi*r^3
+# let 'p' be the fraction of the volume we care about (say, 10% or 0.1)
+#
+# pi*(r-Y)*(r-Y)*(3r-r+Y)/3 = p*(4/3)*pi*r^3
+#
+# algebra
+#
+# (Y/r)^3 - 3(Y/r) + (2-4p) = 0
+#
+# Solve for the value (Y/r) given parameter p using the numpy.roots, which
+# solves the roots of polynomials in the form ax^n + bx^(n-1) ... + C = 0
+# https://docs.scipy.org/doc/numpy/reference/generated/numpy.roots.html
+# The root we care about will be in the interval (0,1)
+#
+# a = 1
+# b = 0
+# c = -3
+# d = 2 - 4*p
+#
+# DEPRECATED THIS IN FAVOR OF A NEW STRATEGY
+# 
+############################################################################
+
+coeffs = [1.0, 0.0, -3.0, 2.0 - 4.0*max_hole_mask_overlap]
+
+roots = np.roots(coeffs)
+
+radial_mask_check = None
+
+for root in roots:
+    
+    if root > 0.0 and root < 1.0:
+        
+        radial_mask_check = root
+        
+if radial_mask_check is None:
+    
+    raise ValueError("Could not calculate appropriate radial check value for input max_hole_mask_overlap "+str(max_hole_mask_overlap))
+
+
+print("For mask volume check of: ", max_hole_mask_overlap, "Using radial hole value of: ", radial_mask_check)
+
+
+
+
+
+
+
+
+
+#---------------------------------------------------------------------------
+#
+# from voidfinder.find_voids, after the _hole_finder has been run
+'''
+# Pure python version
+#---------------------------------------------------------------------------
+coordinates = np.empty((1,3), dtype=np.float64)
+temp_coordinates = np.empty((1,3), dtype=np.float64)
+mask = mask.astype(np.uint8)
+
+keep_holes = np.ones(x_y_z_r_array.shape[0], dtype=np.bool)
+
+for idx in range(x_y_z_r_array.shape[0]):
+
+    hole_radius = x_y_z_r_array[idx,3]
+    
+    coordinates[0,0] = x_y_z_r_array[idx,0]
+    coordinates[0,1] = x_y_z_r_array[idx,1]
+    coordinates[0,2] = x_y_z_r_array[idx,2]
+
+    discard = check_mask_overlap(coordinates,
+                       temp_coordinates,
+                       radial_mask_check,
+                       hole_radius,
[…]
