+++ python/vast/vsquared/util.py
@@ -1,0 +1,183 @@
+import numpy as np
+import collections
+import astropy.units as u
+from astropy.cosmology import FlatLambdaCDM, z_at_value
+from scipy import interpolate
+
+c    = 3e5
+D2R  = np.pi/180.
+
+
+def toCoord(z,ra,dec,H0,Om_m):
+    """Convert redshift, RA, and Dec to comoving coordinates.
+
+    Parameters
+    ----------
+    z : list or ndarray
+        Object redshift.
+    ra : list or ndarray
+        Object right ascension, in decimal degrees.
+    dec : list or ndarray
+        Object declination, in decimal degrees.
+    H0 : float
+        Hubble's constant in km/s/Mpc.
+    Om_m : float
+        Value of matter density.
+
+    Returns
+    -------
+    cs : list
+        Comoving xyz-coordinates, assuming input cosmology.
+    """
+    Kos = FlatLambdaCDM(H0,Om_m)
+    r = Kos.comoving_distance(z)
+    r = np.array([d.value for d in r])
+    #r = c*z/H0
+    c1 = r*np.cos(ra*D2R)*np.cos(dec*D2R)
+    c2 = r*np.sin(ra*D2R)*np.cos(dec*D2R)
+    c3 = r*np.sin(dec*D2R)
+    return c1,c2,c3
+
+
+def toSky(cs,H0,Om_m,zstep):
+    """Convert redshift, RA, and Dec to comoving coordinates.
+
+    Parameters
+    ----------
+    cs : ndarray
+        Comoving xyz-coordinates table [x,y,z], assuming input cosmology.
+    H0 : float
+        Hubble's constant in km/s/Mpc.
+    Om_m : float
+        Value of matter density.
+    zstep : float
+        Redshift step size for converting distance to redshift.
+
+    Returns
+    -------
+    z : float
+        Object redshift.
+    ra : float
+        Object right ascension, in decimal degrees.
+    dec : float
+        Object declination, in decimal degrees.
+    """
+    Kos = FlatLambdaCDM(H0,Om_m)
+    c1 = cs.T[0]
+    c2 = cs.T[1]
+    c3 = cs.T[2]
+    r   = np.sqrt(c1**2.+c2**2.+c3**2.)
+    dec = np.arcsin(c3/r)/D2R
+    ra  = (np.arccos(c1/np.sqrt(c1**2.+c2**2.))*np.sign(c2)/D2R)%360
+    zmn = z_at_value(Kos.comoving_distance,np.amin(r)*u.Mpc)
+    zmx = z_at_value(Kos.comoving_distance,np.amax(r)*u.Mpc)
+    zmn = zmn-(zstep+zmn%zstep)
+    zmx = zmx+(2*zstep-zmx%zstep)
+    ct  = np.array([np.linspace(zmn,zmx,int(np.ceil(zmn/zstep))),Kos.comoving_distance(np.linspace(zmn,zmx,int(np.ceil(zmn/zstep)))).value]).T
+    r2z = interpolate.pchip(*ct[:,::-1].T)
+    z = r2z(r)
+    #z = H0*r/c
+    return z,ra,dec
+
+
+def inSphere(cs, r, coords):
+    """
+    Checks if a set of comoving coordinates are within a sphere.
+
+    Parameters
+    ==========
+
+    cs : list or ndarray
+        Center of sphere.
+
+    r : float
+        Sphere volume.
+
+    coords : list or ndarray
+        Comoving xyz-coordinates.
+
+    Returns
+    =======
+
+    inSphere : bool
+        True if abs(coords - cs) < r.
+    """
+    return np.sum((cs.reshape(3,1) - coords.T)**2., axis=0)<r**2.
+
+
+def wCen(vols,coords):
[…]
