+++ python/vast/vsquared/tests/test_init.py
@@ -1,0 +1,4 @@
+from vast import vsquared as V2
+
+def test_version_exists():
+    assert hasattr(V2, '__version__')
