+++ python/vast/vsquared/viz/void_render.py
@@ -1,0 +1,2338 @@
+"""Void rendering program based on OpenGL and vispy.
+"""
+
+import os
+import subprocess
+
+import numpy as np
+
+from vispy import app, gloo
+import vispy.io as io
+from vispy.util.transforms import perspective, translate, rotate
+
+#from vispy.color import Color
+
+from scipy.spatial import cKDTree
+
+from sklearn import neighbors
+
+import time
+
+#import gc
+
+#Vertex shader for the Galaxy Vispy Program
+vert = """
+#version 120
+// Uniforms
+// ------------------------------------
+//uniform mat4 u_model;
+uniform mat4 u_view;
+uniform mat4 u_projection;
+uniform float u_linewidth;
+uniform float u_antialias;
+uniform float u_size;
+// Attributes
+// ------------------------------------
+//attribute vec3  a_position;
+attribute vec4  a_position;
+attribute vec4  a_fg_color;
+attribute vec4  a_bg_color;
+attribute float a_size;
+// Varyings
+// ------------------------------------
+varying vec4 v_fg_color;
+varying vec4 v_bg_color;
+varying float v_size;
+varying float v_linewidth;
+varying float v_antialias;
+void main (void) {
+    v_size = a_size * u_size;
+    v_linewidth = u_linewidth;
+    v_antialias = u_antialias;
+    v_fg_color  = a_fg_color;
+    v_bg_color  = a_bg_color;
+    
+    //gl_Position = u_projection * u_view * u_model * vec4(a_position,1.0);
+    //gl_Position = u_projection * u_view * u_model * a_position;
+    gl_Position = u_projection * u_view * a_position;
+    
+    gl_PointSize = v_size + 2.*(v_linewidth + 1.5*v_antialias);
+}
+"""
+#Fragment shader for the Galaxy Vispy Program
+frag = """
+#version 120
+// Constants
+// ------------------------------------
+// Varyings
+// ------------------------------------
+varying vec4 v_fg_color;
+varying vec4 v_bg_color;
+varying float v_size;
+varying float v_linewidth;
+varying float v_antialias;
+// Functions
+// ------------------------------------
+// ----------------
+float disc(vec2 P, float size)
+{
+    float r = length((P.xy - vec2(0.5,0.5))*size);
+    r -= v_size/2.;
+    return r;
+}
+// ----------------
+float arrow_right(vec2 P, float size)
+{
+    float r1 = abs(P.x -.50)*size + abs(P.y -.5)*size - v_size/2.;
+    float r2 = abs(P.x -.25)*size + abs(P.y -.5)*size - v_size/2.;
+    float r = max(r1,-r2);
+    return r;
+}
+// ----------------
+float ring(vec2 P, float size)
+{
+    float r1 = length((P.xy - vec2(0.5,0.5))*size) - v_size/2.;
+    float r2 = length((P.xy - vec2(0.5,0.5))*size) - v_size/4.;
+    float r = max(r1,-r2);
+    return r;
+}
+// ----------------
+float clober(vec2 P, float size)
+{
+    const float PI = 3.14159265358979323846264;
+    const float t1 = -PI/2.;
+    const vec2  c1 = 0.2*vec2(cos(t1),sin(t1));
+    const float t2 = t1+2.*PI/3.;
+    const vec2  c2 = 0.2*vec2(cos(t2),sin(t2));
+    const float t3 = t2+2.*PI/3.;
+    const vec2  c3 = 0.2*vec2(cos(t3),sin(t3));
+    float r1 = length((P.xy- vec2(0.5,0.5) - c1)*size);
+    r1 -= v_size/3;
[…]

[…]
