### 2022-02-05 dfe7f60
First attempt at top-level setup.py.
