### 2022-02-05 e3a493c
Clean up setup.
