### 2022-02-05 b15bd72
Add V^2 utility tests.
