+++ python/vast/test/__init__.py
@@ -1,0 +1,1 @@
+from .vast_test_suite import runtests
