+++ python/vast/voidfinder/__init__.py
@@ -1,0 +1,7 @@
+# This line allows you to do "from vast.voidfinder import X" instead of
+# "from vast.voidfinder.voidfinder import X"
+from .voidfinder import filter_galaxies, \
+                        ra_dec_to_xyz, \
+                        calculate_grid, \
+                        wall_field_separation, \
+                        find_voids
