### 2022-02-05 4ec8c83
Set up working VoidFinder unit test.
