### 2022-02-05 fa7797d
Add file deletion to test teardown.
