### 2022-02-05 588a150
Add run scripts to voidfinder subfolder.
