+++ python/vast/voidfinder/scripts/SDSS_VoidFinder_dr7.py
@@ -1,0 +1,209 @@
+'''VoidFinder - Hoyle & Vogeley (2002)'''
+
+
+
+################################################################################
+#
+# If you have control over your python environment, voidfinder can be installed
+# as a normal python package via 'python setup.py install', in which case the 
+# below import of 'sys' and 'sys.path.insert(0, '/abspath/to/VoidFinder/python'
+# is unnecessary.  If you aren't able to install the voidfinder package,
+# you can use the sys.path.insert to add it to the list of available packages
+# in your python environment.
+#
+# Alternately, "python setup.py develop" will 'install' some symlinks which
+# point back to the current directory and you can run off the same voidfinder
+# repository that you're working on as if it was installed
+#
+################################################################################
+
+
+#import sys
+#sys.path.insert(1, 'local/path/VAST/VoidFinder/vast/voidfinder/')
+
+################################################################################
+#
+#   IMPORT MODULES
+#
+################################################################################
+
+from vast.voidfinder import find_voids, filter_galaxies
+
+from vast.voidfinder.multizmask import generate_mask
+from vast.voidfinder.preprocessing import file_preprocess
+
+import pickle
+
+
+
+
+
+################################################################################
+#
+#   USER INPUTS
+#
+################################################################################
+
+
+# Number of CPUs available for analysis.
+# A value of None will use one less than all available CPUs.
+num_cpus = 1
+
+#-------------------------------------------------------------------------------
+survey_name = 'vollim_dr7_cbp_102709_'
+
+# File header
+# Change these directory paths to where your data is stored, and where you want 
+# the output to be saved.
+in_directory = '../'
+out_directory = '../'
+
+
+# Input file name
+galaxies_filename = 'vollim_dr7_cbp_102709.dat'  # File format: RA, dec, redshift, comoving distance, absolute magnitude
+#-------------------------------------------------------------------------------
+
+#-------------------------------------------------------------------------------
+# Survey parameters
+# Note: These can be set to None, in which case VoidFinder will use the limits 
+# of the galaxy catalog.
+min_z = 0
+max_z = 0.107
+
+# Cosmology (uncomment and change values to change cosmology)
+# Need to also uncomment relevent inputs in function calls below
+Omega_M = 0.26
+#h = 1
+
+# Uncomment if you do NOT want to use comoving distances
+# Need to also uncomment relevent inputs in function calls below
+dist_metric = 'comoving'
+#-------------------------------------------------------------------------------
+
+#-------------------------------------------------------------------------------
+# Uncomment if you do NOT want to remove galaxies with Mr > -20
+# Need to also uncomment relevent input in function calls below
+#mag_cut = False
+
+
+# Uncomment if you do NOT want to remove isolated galaxies
+# Need to also uncomment relevent input in function calls below
+#rm_isolated = False
+#-------------------------------------------------------------------------------
+
+
+
+
+################################################################################
+#
+#   PREPROCESS DATA
+#
+################################################################################
+
+galaxy_data_table, dist_limits, out1_filename, out2_filename = file_preprocess(galaxies_filename, 
+                                                                               in_directory, 
+                                                                               out_directory, 
+                                                                               #mag_cut=mag_cut,
+                                                                               #rm_isolated=rm_isolated,
+                                                                               dist_metric=dist_metric,
+                                                                               min_z=min_z, 
+                                                                               max_z=max_z,
+                                                                               Omega_M=Omega_M,
+                                                                               #h=h,
+                                                                               verbose=0)
+
+print("Dist limits: ", dist_limits)
+
[…]
