### 2022-02-05 24d49e0
Add setuptools manifest for pypi.
