+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include requirements.txt
+recursive-include python/vast/voidfinder *.pxd *.pyx *.c
