### 2022-02-05 40f795e
Remove obsolete VoidFinder code.
