+++ VoidFinder/vast/voidfinder/_voidfinder_cython_find_next.pyx
@@ -1,2513 +1,0 @@
-
-
-
-
-from __future__ import print_function
-
-cimport cython
-
-from cpython.mem cimport PyMem_Malloc, PyMem_Realloc, PyMem_Free
-
-import numpy as np
-
-cimport numpy as np
-
-np.import_array()  # required in order to use C-API
-
-from typedefs cimport DTYPE_CP128_t, \
-                      DTYPE_CP64_t, \
-                      DTYPE_F64_t, \
-                      DTYPE_F32_t, \
-                      DTYPE_B_t, \
-                      ITYPE_t, \
-                      DTYPE_INT32_t, \
-                      DTYPE_INT64_t, \
-                      DTYPE_UINT16_t, \
-                      CELL_ID_t
-
-from numpy.math cimport NAN, INFINITY
-
-from libc.math cimport fabs, sqrt, asin, atan, ceil#, exp, pow, cos, sin, asin
-
-#from libc.stdlib cimport malloc, free
-
-import time
-
-
-
-
-
-
-
-@cython.boundscheck(False)
-@cython.wraparound(False)
-@cython.cdivision(True)
-@cython.profile(True)
-cdef FindNextReturnVal find_next_galaxy(DTYPE_F64_t[:,:] hole_center_memview, 
-                                        DTYPE_F64_t[:,:] temp_hole_center_memview,
-                                        DTYPE_F64_t search_radius, 
-                                        DTYPE_F64_t dr, 
-                                        DTYPE_F64_t direction_mod,
-                                        DTYPE_F64_t[:] unit_vector_memview, 
-                                        galaxy_tree, 
-                                        DTYPE_INT64_t[:] nearest_gal_index_list, 
-                                        ITYPE_t num_neighbors,
-                                        DTYPE_F64_t[:,:] w_coord, 
-                                        DTYPE_B_t[:,:] mask, 
-                                        DTYPE_INT32_t mask_resolution,
-                                        DTYPE_F64_t min_dist, 
-                                        DTYPE_F64_t max_dist, 
-                                        DTYPE_F64_t[:] Bcenter_memview,
-                                        Cell_ID_Memory cell_ID_mem,
-                                        NeighborMemory neighbor_mem,
-                                        ):          
-
-    '''
-    Description:
-    ============
-    Function to locate the next nearest galaxy during hole center propagation 
-    along direction defined by unit_vector_memview.
-
-    The algorithm needs to find 4 bounding galaxies per cell.  The 
-    first galaxy is found as the minimum of regular euclidean distance to the hole center.  The
-    second galaxy is determined using a ratio of distance with respect to projected distance
-    along the hole center search path.  To find galaxy 2, the hole center search path and 
-    distance ratios are calculated with respect to neighbor 1.  
-
-    Implementation Detail:
-    For galaxies 3 and 4, the 
-    distance ratios are calculated with respect to the previous galaxies, so this code uses
-    an if-block to detect whether we're calculating for galaxy 2/B, 3/C, or 4/D, and runs slightly
-    different distance ratio calculations.  The if block looks for the number of elements
-    in the input nearest_gal_index_list, when its 1 it assumes we're finding for galaxy 2
-    and when its not 1 it assumes we have the first 2 galaxies and we're looking for galaxy
-    3 or 4.
-
-
-
-    Parameters:
-    ===========
-
-    hole_center_memview : memview of shape (1,3)
-        x,y,z coordinate of current center of hole in units of Mpc/h
-        
-    NEEDS DEFINITION temp_hole_center_memview :
-
-    search_radius : float
-        Radius of hole in units of Mpc/h
-
-    dr : float
-        Incrememt value for hole propagation
-        
-    direction_mod : -1.0 or 1.0
[…]
