### 2022-02-05 4c57855
Delete obsolete Vsquared folder.
