+++ Vsquared/vast/vsquared/classes.py
@@ -1,312 +1,0 @@
-"""Utility classes for the ZOBOV algorithm using a voronoi tesselation of an
-input catalog.
-"""
-
-import numpy as np
-import healpy as hp
-from astropy.io import fits
-from scipy.spatial import ConvexHull, Voronoi, Delaunay, KDTree
-
-from vast.vsquared.util import toCoord, flatten
-
-class Catalog:
-    """Catalog data for void calculation.
-    """
-
-    def __init__(self,catfile,nside,zmin,zmax,maglim=None,H0=100,Om_m=0.3,maskfile=None):
-        """Initialize catalog.
-
-        Parameters
-        ----------
-        catfile: str
-            Object catalog file (FITS format).
-        nside : int
-            HEALPix map nside parameter (2,4,8,16,...,2^k).
-        zmin : float
-            Minimum redshift boundary.
-        zmax : float
-            Maximum redshift boundary.
-        maglim : float or None
-            Catalog object magnitude limit.
-        H0 : float
-            Hubble parameter, in units of km/s/Mpc.
-        Om_m : float
-            Matter density.
-        maskfile : str or None
-            Mask file giving HEALPixels with catalog objects.
-        """
-        print("Extracting data...")
-        hdulist = fits.open(catfile)
-        z    = hdulist[1].data['z']
-        ra   = hdulist[1].data['ra']
-        dec  = hdulist[1].data['dec']
-        zcut = np.logical_and(z>zmin,z<zmax)
-        if not zcut.any():
-            print("Choose valid redshift limits")
-            return
-        scut = zcut
-        c1,c2,c3   = toCoord(z,ra,dec,H0,Om_m)
-        self.coord = np.array([c1,c2,c3]).T
-        nnls = np.arange(len(z))
-        nnls[zcut<1] = -1
-        if maglim is not None:
-            mag  = hdulist[1].data['rabsmag']
-            print("Applying magnitude cut...")
-            mcut = np.logical_and(mag<maglim,zcut)
-            if not mcut.any():
-                print("Choose valid magnitude limit")
-                return
-            scut = mcut
-            tree = KDTree(self.coord[mcut])
-            nnls[zcut][mcut[zcut]<1] = tree.query(self.coord[zcut][mcut[zcut]<1])[1]
-            self.mcut = mcut
-        self.nnls = nnls
-        if maskfile is None:
-            print("Generating mask...")
-            mask = np.zeros(hp.nside2npix(nside),dtype=bool)
-            pids = hp.ang2pix(nside,ra[scut],dec[scut],lonlat=True)
-            mask[pids] = True
-        self.mask = mask
-        pids = hp.ang2pix(nside,ra,dec,lonlat=True)
-        self.imsk = mask[pids]*zcut
-
-class Tesselation:
-    """Implementation of Voronoi tesselation of the catalog.
-    """
-
-    def __init__(self,cat,viz=False):
-        """Initialize tesselation.
-
-        Parameters
-        ----------
-        cat : Catalog
-            Catalog of objects used to compute the Voronoi tesselation.
-        viz : bool
-            Compute visualization.
-        """
-        coords = cat.coord[cat.nnls==np.arange(len(cat.nnls))]
-        print("Tesselating...")
-        Vor = Voronoi(coords)
-        ver = Vor.vertices
-        reg = np.array(Vor.regions)[Vor.point_region]
-        del Vor
-        ve2 = ver.T
-        vth = np.arctan2(np.sqrt(ve2[0]**2.+ve2[1]**2.),ve2[2])
-        vph = np.arctan2(ve2[1],ve2[0])
[…]
