### 2022-02-05 DESI-UR/VAST#50 @sybenzvi
@kadglass, I did the reorganization and made a PR (#53). Note that you'll see complaints about the unit tests failing, but this is because I deleted the old CircleCI tests in favor of using GitHub Actions, which will make them easier to maintain going forward. I checked the unit tests separately and they definitely work.
