### 2022-02-06 51c6160
added single threaded mode back in
