### 2022-02-06 080b940
Updated find_voids function call in VF unit tests
