### 2022-02-06 3665421
Xyz mask enabling enhancement #41 Cartesian input (#46)
* Initializing source files for ReadTheDocs documentation

* Updated VF ReadTheDocs documentation and made necessary adjustments to docstrings

* Removed unusued package imports from example scripts

* Add to Vsquared documentation

* Fix bugs in documentation

* Added enhancement #41 Cartesian input for VoidFinder ability to utilize xyz-based mask

* Added switch to turn on/off non-empty hole checking; commented out 'install_requires=open('requirements.txt') in setup.py until we can establish >= semantics for package requirement versions
[…]
