+++ VoidFinder/vast/voidfinder/voidfinder.py
@@ -36,6 +36,8 @@ from .constants import c
 
 from ._voidfinder_cython import check_mask_overlap
 
+from ._voidfinder_cython_find_next import MaskChecker
+
 
 ################################################################################
 DtoR = np.pi/180.
@@ -177,8 +179,8 @@ def filter_galaxies(galaxy_table,
     ############################################################################
     # Grid shape
     #---------------------------------------------------------------------------
-    hole_grid_shape, coords_min = calculate_grid(coords_xyz,
-                                                 hole_grid_edge_length)
+    hole_grid_shape, coords_min, coords_max = calculate_grid(coords_xyz,
+                                                             hole_grid_edge_length)
     ############################################################################
 
 
@@ -378,7 +380,7 @@ def calculate_grid(galaxy_coords_xyz,
     
     hole_grid_shape = tuple(np.ceil(ngrid).astype(int))
     
-    return hole_grid_shape, coords_min
+    return hole_grid_shape, coords_min, coords_max
     
     
 
@@ -492,12 +494,16 @@ def wall_field_separation(galaxy_coords_xyz,
 
 
 def find_voids(galaxy_coords_xyz,
-               dist_limits,
-               mask, 
-               mask_resolution,
                coords_min,
                hole_grid_shape,
                survey_name,
+               mask_type='ra_dec_z',
+               mask=None, 
+               mask_resolution=None,
+               min_dist=None,
+               max_dist=None,
+               xyz_limits=None,
+               check_only_empty_holes=True,
                max_hole_mask_overlap=0.1,
                hole_grid_edge_length=5.0,
                min_maximal_radius=10.0,
@@ -610,10 +616,26 @@ def find_voids(galaxy_coords_xyz,
     galaxy_coords_xyz : numpy.ndarray of shape (num_galaxies, 3)
         coordinates of the galaxies in the survey, units of Mpc/h 
         (xyz space)
+        
+    mask_type : string, 'ra_dec_z' or 'xyz'
+        Determines the mode of mask checking to use and what mask parameters
+        to use.  ra_dec_z means the mask, mask_resolution, and dist_limits parameters
+        will be provided representing an angular `mask` in Right Ascension and Declination
+        and a corresponding mask_resolution integer, as well as dist_limits representing
+        the min and max redshift values represented as radial distances in the xyz space
+        
+        'xyz' means that the xyz_limits parameter will be provided which directly encodes
+        a bounding box for the survey in xyz space
+        
     
     dist_limits : numpy array of shape (2,)
         minimum and maximum distance limit of the survey in units of Mpc/h 
         (xyz space)
+        
+    xyz_limits : numpy array of shape (2,3)
+        format [x_min, y_min, z_min]
+               [x_max, y_max, z_max]
+        to be used for checking against the mask when mask_type == 'xyz'
         
     mask : numpy.ndarray of shape (N,M) type bool
         represents the survey footprint in scaled ra/dec space.  Value of True 
@@ -688,9 +710,13 @@ def find_voids(galaxy_coords_xyz,
         choose a good balance between saving too often and not often enough if 
         using this parameter.  Note that it is an approximate value because it 
         depends on the number of worker processes and the provided batch_size 
-        value, if you batch size is 10,000 and your save_after is 1,000,000 you 
+        value, if your batch size is 10,000 and your save_after is 1,000,000 you 
         might actually get a checkpoint at say 1,030,000.  If None, disables 
         saving the checkpoint file.
+
+    check_only_empty_holes : bool
+        whether or not to start growing a hole in a cell which has galaxies in
[…]
