+++ Documentation/source/Vsquared_examples.rst
@@ -1,3 +1,4 @@
+
 .. role:: raw-html(raw)
     :format: html
 
@@ -31,6 +32,7 @@ A summary checklist for installing and running
    * etc.
 
  * Run ``vsquared.py`` from the ``VAST/Vsquared/scripts`` directory on your 
+
    machine or using a cluster::
    
     python vsquared.py -c DR7_config.ini
@@ -485,5 +487,5 @@ does not belong to any void, and any objects contained within that zone are not
 in a void.
  
  
- 
- 
+
+
