### 2022-02-07 ef92ca7
Add DR7 analysis configuration file.
