+++ python/vast/vsquared/data/DR7_config.ini
@@ -1,0 +1,23 @@
+[Paths]
+#Currently only accepts FITS input
+Input Catalog    = ../data/vollim_dr7_cbp_102709.fits
+Survey Name      = DR7
+Output Directory = ../data/
+
+[Cosmology]
+#Currently only uses Flat LambdaCDM
+H_0     = 100
+Omega_m = 0.3
+
+[Settings]
+#Redshift limits
+redshift_min  = 0.0
+redshift_max  = 0.1
+#Magnitude limit, can be None
+rabsmag_min   = None
+#Minimum void radius (Mpc/h)
+radius_min    = 10.0
+#NSIDE for HEALPix pixelization
+nside         = 32
+#Step size for comoving-distance-to-redshift lookup table
+redshift_step = 5e-5
