### 2022-02-07 02dddb3
Add test of V^2 zone and void formation.
