+++ python/vast/vsquared/data/test_config.ini
@@ -1,8 +1,8 @@
 [Paths]
 #Currently only accepts FITS input
-Input Catalog    = ../../data/test_data.fits
+Input Catalog    = python/vast/vsquared/data/test_data.fits
 Survey Name      = TEST
-Output Directory = ../../data/
+Output Directory = ./
 
 [Cosmology]
 #Currently only uses Flat LambdaCDM
