### 2022-02-07 c1b41ca
Finalize V^2 unit tests from dveyrat branch.
