+++ python/vast/test/test_V2.py
@@ -5,6 +5,8 @@ import unittest
 
 from vast.vsquared import util, classes, zobov
 
+import os
+import copy
 import numpy as np
 
 class TestV2(unittest.TestCase):
@@ -12,11 +14,14 @@ class TestV2(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         # Set up global varaibles for tests.
+        TestV2.inifile = 'python/vast/vsquared/data/test_config.ini'
         TestV2.catfile = 'python/vast/vsquared/data/test_data.fits'
+
         TestV2.cat = None
         TestV2.tess = None
         TestV2.zones = None
         TestV2.voids = None
+        TestV2.zobov = None
 
     def test_cat_coord(self):
         """Check catalog coordinate access
@@ -82,3 +87,41 @@ class TestV2(unittest.TestCase):
         # Test ovols
         self.assertTrue(np.isclose(np.mean([np.mean(ov) for ov in self.voids.ovols]), 5519.078018084154))
 
+    def test_zobov_4_zobov(self):
+        """Test full ZOBOV algorithm
+        """
+        TestV2.zobov = zobov.Zobov(TestV2.inifile, save_intermediate=False)
+
+        # Sort voids.
+        self.assertFalse(hasattr(TestV2.zobov, 'vrads')) # before sortVoids
+
+        zobov1 = copy.deepcopy(TestV2.zobov)
+        zobov2 = copy.deepcopy(TestV2.zobov)
+        zobov4 = copy.deepcopy(TestV2.zobov)
+        TestV2.zobov.sortVoids()
+        zobov1.sortVoids(1)
+        zobov2.sortVoids(2)
+        zobov4.sortVoids(4)
+        self.assertEqual(len(TestV2.zobov.vrads), 62)
+        self.assertEqual(len(zobov1.vrads), 87)
+        self.assertEqual(len(zobov2.vrads), 14)
+        self.assertEqual(len(zobov4.vrads), 49)
+
+        # Save voids.
+        self.assertTrue(hasattr(TestV2.zobov, 'vrads')) # after sortVoids
+
+        TestV2.zobov.saveVoids()
+        self.assertTrue(os.path.exists('TEST_zobovoids.dat'))
+        self.assertTrue(os.path.exists('TEST_zonevoids.dat'))
+
+        # Save zones.
+        TestV2.zobov.saveZones()
+        self.assertTrue(os.path.exists('TEST_galzones.dat'))
+
+    def tearDown(self):
+        """Delete files produced for the unit tests.
+        """
+        files = [ 'TEST_zobovoids.dat', 'TEST_zonevoids.dat', 'TEST_galzones.dat' ]
+        for f in files:
+            if os.path.exists(f):
+                os.remove(f)
