### 2022-02-08 8e80391
Make branch compatible with PR #42.
