+++ python/vast/voidfinder/_voidfinder.py
@@ -35,7 +35,8 @@ from ._voidfinder_cython_find_next import GalaxyMap, \
                                           Cell_ID_Memory, \
                                           GalaxyMapCustomDict, \
                                           HoleGridCustomDict, \
-                                          NeighborMemory
+                                          NeighborMemory, \
+                                          MaskChecker
 
 
 
@@ -67,12 +68,15 @@ def _hole_finder(hole_grid_shape,
                  hole_center_iter_dist,
                  galaxy_map_grid_edge_length,
                  coord_min, 
-                 mask,
-                 mask_resolution,
-                 min_dist,
-                 max_dist,
                  galaxy_coords,
                  survey_name,
+                 mask_mode=0,
+                 mask=None,
+                 mask_resolution=None,
+                 min_dist=None,
+                 max_dist=None,
+                 xyz_limits=None,
+                 check_only_empty_holes=True,
                  #hole_radial_mask_check_dist,
                  save_after=None,
                  use_start_checkpoint=False,
@@ -177,8 +181,8 @@ def _hole_finder(hole_grid_shape,
         number of cpus to use while running the main algorithm.  None will 
         result in using number of physical cores on the machine.  Some speedup 
         benefit may be obtained from using additional logical cores via Intel 
-        Hyperthreading but with diminishing returns.  This can safely be set 
-        above the number of physical cores without issue if desired.
+        Hyperthreading but with diminishing returns based on some basic
+        spot testing
     
     
     
@@ -194,101 +198,28 @@ def _hole_finder(hole_grid_shape,
         combining stage so will not represent the final output of VoidFinder
     '''
 
-
-    '''
-    #ijk reordertest
-    
-    batch_size = 126
-    
-    start_idx = 0
-
-    test_cell_ID_dict = {(4,4,0) : 1,
-                         (0,2,4) : 1}
-    
-    i_j_k_array = np.empty((batch_size, 3), dtype=np.int64)
-
-    test_gen = CellIDGenerator(5,5,5,test_cell_ID_dict)
-    
-    num_write = test_gen.gen_cell_ID_batch(start_idx, batch_size, i_j_k_array)
-
-    print("num_write: ", num_write)
-    
-    for idx, row in enumerate(i_j_k_array):
-        
-        print(idx, row)
-    
-    
-    print("Start idx 10")
-    
-    num_write = test_gen.gen_cell_ID_batch(10, 10, i_j_k_array)
-    
-    print(num_write)
-    print(i_j_k_array[0:10])
-    
-    print("Start idx 125")
-    
-    num_write = test_gen.gen_cell_ID_batch(126, 10, i_j_k_array)
-    
-    print(num_write)
-    
-    exit()
-    '''
-
-
-
-
-    ############################################################################
-    # Run single or multi-processed
-    ############################################################################
-    
-    if isinstance(num_cpus, int) and num_cpus == 1:
-        
-        #cProfile.runctx("run_single_process_cython(ngrid, dl, dr, coord_min, mask, mask_resolution, min_dist, max_dist, w_coord, batch_size=batch_size, verbose=verbose, print_after=print_after, num_cpus=num_cpus)", globals(), locals(), 'prof_single.prof')
-        #x_y_z_r_array = None
-        #n_holes = None
-        
-        
-        x_y_z_r_array, n_holes = _hole_finder_single_process(hole_grid_shape, 
-                                                             hole_grid_edge_length, 
-                                                             hole_center_iter_dist,
-                                                             galaxy_map_grid_edge_length,
-                                                             coord_min, 
-                                                             mask,
-                                                             mask_resolution,
-                                                             min_dist,
-                                                             max_dist,
-                                                             galaxy_coords,
-                                                             survey_name,
[…]
