+++ Documentation/source/VoidRender_functions.rst
@@ -1,64 +1,0 @@
-####################
-VoidRender API
-####################
-
-
-
-**************************
-VoidRender for VoidFinder
-**************************
-
-
-
-``load_galaxy_data``
-====================
-
-.. autofunction:: vast.voidfinder.viz.load_galaxy_data
-
-
-
-``load_void_data``
-===================
-
-.. autofunction:: vast.voidfinder.viz.load_void_data
-
-
-
-.. _VR-VF-docstring:
-
-``VoidRender``
-=================
-
-.. autofunction:: vast.voidfinder.viz.VoidRender.__init__
-
-
-
-
-
-
-**************************
-VoidRender for V\ :sup:`2`
-**************************
-
-
-
-``load_galaxy_data``
-====================
-
-.. autofunction:: vast.vsquared.viz.load_galaxy_data
-
-
-
-``load_void_data``
-===================
-
-.. autofunction:: vast.vsquared.viz.load_void_data
-
-
-
-.. _VR-V2-docstring:
-
-``VoidRender``
-=================
-
-.. autofunction:: vast.vsquared.viz.VoidRender.__init__
