+++ Documentation/source/VoidFinder_examples.rst
@@ -1,0 +1,605 @@
+################
+Using VoidFinder
+################
+
+
+
+Quick start
+===========
+
+A summary checklist for installing and running **VoidFinder**.
+
+ * Clone the `GitHub repository <https://github.com/DESI-UR/VAST>`_
+ * Navigate to ``VAST/VoidFinder`` and run::
+    
+    python setup.py install
+    
+   See :ref:`VF-install` for installation options.
+   
+ * Navigate to ``VAST/VoidFinder/scripts`` and modify
+   ``SDSS_VoidFinder_dr7.py`` if appropriate.  Variables to edit might include:
+   
+   * ``in_directory`` and ``out_directory``
+   * ``data_filename``
+   * redshift and/or magnitude limits
+   * Comment out ``dist_metric`` from all functions if you are using comoving distances.
+   * etc.
+
+ * Run your script (``SDSS_VoidFinder_dr7.py``, in this case) on your machine or 
+   using a cluster.
+
+The output files will be located in the directory specified by ``out_directory``.
+
+ * ``[survey_name]_[comoving/redshift]_holes.txt``
+ * ``[survey_name]_[comoving/redshift]_maximal.txt``
+ * ``[survey_name]_mask.pickle``
+ * ``[survey_name]_field_gal_file.txt``
+ * ``[survey_name]_wall_gal_file.txt``
+
+See :ref:`VF-output` for a detailed description of each of these files.
+
+
+
+
+Example scripts
+===============
+
+Included in the **VoidFinder** repository (``VAST/VoidFinder/scripts/``) are 
+a finite selection of example scripts:
+
+ * ``SDSS_VoidFinder_dr7.py`` will run **VoidFinder** on the SDSS DR7 main 
+   galaxy sample.  A volume-limited version of this galaxy catalog is provided 
+   with the package (``VAST/VoidFinder/vollim_dr7_cbp_102709.dat``).
+ * ``classifyEnvironment.py`` takes the output of **VoidFinder** (identified 
+   voids) and determines which objects of a given catalog are within the voids 
+   ("void" objects), which are outside the voids ("wall" objects), and which are 
+   too close to the survey boundary and/or are outside the survey bounds to be 
+   classified ("edge" objects).
+
+
+
+
+Finding voids
+=============
+
+The easiest way to use **VoidFinder** is to create a script that
+
+1. Filters and processes the input galaxy catalog so that it is in the correct 
+   format for **VoidFinder**
+2. Generates a survey mask based on the input galaxy catalog, defining the 
+   boundaries within which voids can be found
+3. Finds voids within the galaxy catalog
+
+An example of this script is the ``SDSS_VoidFinder_dr7.py`` file, located in 
+``VAST/VoidFinder/scripts/``.  What follows is a breakdown of this script, 
+explaining the various options and functions called.
+
+
+1. Preparing the data
+---------------------
+
+Pre-processing the data
+^^^^^^^^^^^^^^^^^^^^^^^
+
+Generally, the first function that should be called in a script running 
+**VoidFinder** is ``file_preprocess``::
+
+    from vast.voidfinder.preprocessing import file_preprocess
+    
+This function reads the input data file into memory (as an 
+``astropy.table.Table`` object), checks the various column names to ensure that 
+they match what is expected in other functions, and creates the file names for 
+the files produced by **VoidFinder**.  While not strictly necessary to run, it 
+encompasses many of the introductory steps necessary to run **VoidFinder**.
+
+The outputs from ``file_preprocess`` are
+
+ * The input galaxy catalog as an ``astropy.table.Table`` object
+ * The distance limits within which to locate voids
[…]
