+++ python/vast/voidfinder/_voidfinder_cython.pyx
@@ -36,9 +36,30 @@ from _voidfinder_cython_find_next cimport find_next_galaxy, \
                                           GalaxyMapCustomDict, \
                                           HoleGridCustomDict, \
                                           FindNextReturnVal, \
-                                          NeighborMemory
+                                          NeighborMemory, \
+                                          MaskChecker
+                                          
 
 import time
+
+
+#Create a typedef for the type of pointer which can point
+#to the mask check function
+ctypedef DTYPE_B_t (*MASK_CHECK_FUNCTION_t)(DTYPE_F64_t[:,:], DTYPE_B_t[:,:], DTYPE_INT32_t, DTYPE_F64_t, DTYPE_F64_t)
+
+
+
+
+
+
+cdef DTYPE_B_t test_function_woo(DTYPE_F64_t[:,:] a, 
+                                 DTYPE_B_t[:,:] b, 
+                                 DTYPE_INT32_t c,  
+                                 DTYPE_F64_t d, 
+                                 DTYPE_F64_t e):
+
+    return <DTYPE_B_t>0
+
 
 
 
@@ -292,10 +313,11 @@ cpdef void main_algorithm(DTYPE_INT64_t[:,:] i_j_k_array,
                           DTYPE_F64_t dl, 
                           DTYPE_F64_t dr,
                           DTYPE_F64_t[:,:] coord_min, 
-                          DTYPE_B_t[:,:] mask,
-                          DTYPE_INT32_t mask_resolution,
-                          DTYPE_F64_t min_dist,
-                          DTYPE_F64_t max_dist,
+                          MaskChecker mask_checker,
+                          #DTYPE_B_t[:,:] mask,
+                          #DTYPE_INT32_t mask_resolution,
+                          #DTYPE_F64_t min_dist,
+                          #DTYPE_F64_t max_dist,
                           #DTYPE_F64_t hole_radial_mask_check_dist,
                           DTYPE_F64_t[:,:] return_array,
                           Cell_ID_Memory cell_ID_mem,
@@ -388,6 +410,9 @@ cpdef void main_algorithm(DTYPE_INT64_t[:,:] i_j_k_array,
     NAN or (x,y,z,r) rows filled into the return_array parameter.
     
     '''
+    
+    
+    #cdef MASK_CHECK_FUNCTION_t mask_check_function = not_in_mask
     
     
     ############################################################################
@@ -555,7 +580,9 @@ cpdef void main_algorithm(DTYPE_INT64_t[:,:] i_j_k_array,
             hole_center_memview[0,idx] = (hole_center_memview[0,idx] + 0.5)*dl + coord_min[0,idx]
             
             
-        if not_in_mask(hole_center_memview, mask, mask_resolution, min_dist, max_dist):
+        #if not_in_mask(hole_center_memview, mask, mask_resolution, min_dist, max_dist):
+        #if mask_check_function(hole_center_memview, mask, mask_resolution, min_dist, max_dist):
+        if mask_checker.not_in_mask(hole_center_memview):
             
             '''
             out_str = ""
@@ -652,10 +679,11 @@ cpdef void main_algorithm(DTYPE_INT64_t[:,:] i_j_k_array,
                                              nearest_gal_index_list, 
                                              1,
                                              w_coord, 
-                                             mask, 
-                                             mask_resolution,
-                                             min_dist, 
-                                             max_dist,
+                                             mask_checker,
+                                             #mask, 
+                                             #mask_resolution,
+                                             #min_dist, 
+                                             #max_dist,
                                              Bcenter_memview,
                                              cell_ID_mem,
[…]
