+++ python/vast/voidfinder/_voidfinder_cython_find_next.pyx
@@ -53,10 +53,7 @@ cdef FindNextReturnVal find_next_galaxy(DTYPE_F64_t[:,:] hole_center_memview,
                                         DTYPE_INT64_t[:] nearest_gal_index_list, 
                                         ITYPE_t num_neighbors,
                                         DTYPE_F64_t[:,:] w_coord, 
-                                        DTYPE_B_t[:,:] mask, 
-                                        DTYPE_INT32_t mask_resolution,
-                                        DTYPE_F64_t min_dist, 
-                                        DTYPE_F64_t max_dist, 
+                                        MaskChecker mask_checker,
                                         DTYPE_F64_t[:] Bcenter_memview,
                                         Cell_ID_Memory cell_ID_mem,
                                         NeighborMemory neighbor_mem,
@@ -538,7 +535,8 @@ cdef FindNextReturnVal find_next_galaxy(DTYPE_F64_t[:,:] hole_center_memview,
                 return retval
                 
 
-        elif not_in_mask(temp_hole_center_memview, mask, mask_resolution, min_dist, max_dist):
+        #elif not_in_mask(temp_hole_center_memview, mask, mask_resolution, min_dist, max_dist):
+        elif mask_checker.not_in_mask(temp_hole_center_memview):
             
             retval.in_mask = False
             
@@ -560,11 +558,113 @@ cdef DTYPE_F64_t DtoR = np.pi/180.
 cdef DTYPE_F64_t dec_offset = -90
 
 
+
+
+
+cdef class MaskChecker:
+    
+    def __init__(self, 
+                 mode,
+                 survey_mask_ra_dec=None,
+                 n=None,
+                 rmin=None,
+                 rmax=None,
+                 xyz_limits=None):
+        
+        self.mode = mode
+        
+        if survey_mask_ra_dec is not None:
+            self.survey_mask_ra_dec = survey_mask_ra_dec
+            
+        if n is not None:
+            self.n = n
+            
+        if rmin is not None:
+            self.rmin = rmin
+            
+        if rmax is not None:
+            self.rmax = rmax
+            
+        if xyz_limits is not None:
+            self.xyz_limits = xyz_limits
+        
+        
+    cdef DTYPE_B_t not_in_mask(self, DTYPE_F64_t[:,:] coordinates):
+        
+        if self.mode == 0:
+            
+            return not_in_mask(coordinates,
+                               self.survey_mask_ra_dec,
+                               self.n,
+                               self.rmin,
+                               self.rmax)
+            
+        elif self.mode == 1:
+            
+            return not_in_mask_xyz(coordinates, self.xyz_limits)
+
+
+
+
+
 @cython.boundscheck(False)
 @cython.wraparound(False)
 @cython.cdivision(True)
 @cython.profile(True)
-cpdef DTYPE_B_t not_in_mask(DTYPE_F64_t[:,:] coordinates, 
+cdef DTYPE_B_t not_in_mask_xyz(DTYPE_F64_t[:,:] coordinates, 
+                               DTYPE_F64_t[:,:] xyz_limits):
+    """
+    Parameters
+    ==========
+    
+    coordinates : array shape (1,3)
+        format [x,y,z]
+        
+    xyz_limits : array shape (2,3)
+        format [x_min, y_min, z_min]
+               [x_max, y_max, z_max]
+    """
+    
+    cdef DTYPE_F64_t coord_x
+    cdef DTYPE_F64_t coord_y
+    cdef DTYPE_F64_t coord_z
+    
+    coord_x = coordinates[0,0]
[…]
