### 2022-02-08 086e053
Repo reorg feb22 (#53)
* Create a top-level folder for code.

* Copy voidfinder module.

* Remove executable permissions.

* Add tests, visualization, and legacy code.

* Add Voronoi Voids repository.

* Delete CircleCI tests.

* Add GitHub test actions.

* Add version finding code.

* First attempt at top-level setup.py.

* Add missing Cython imports.

* Add package requirements.

* Clean up setup.

* Clean up old folders.

* Add test suite based on unittest.

* Remove old separate package init files.

* Add V^2 utility tests.

* Add init to simplify some imports.

* Set up working VoidFinder
[…]
