+++ doc/paper.bib
@@ -1,0 +1,98 @@
+@article{Abazajian:2009,
+   title         = {The Seventh Data Release of the Sloan Digital Sky Survey},
+   volume        = {182},
+   ISSN          = {1538-4365},
+   eprint        = "0812.0649",
+   archivePrefix = "arXiv",
+   primaryClass  = "astro-ph",
+   DOI           = {10.1088/0067-0049/182/2/543},
+   number        = {2},
+   journal       = {The Astrophysical Journal Supplement Series},
+   publisher     = {IOP Publishing},
+   author        = {Abazajian, Kevork N. and Adelman-McCarthy, Jennifer K. and Agüeros, Marcel A. and Allam, Sahar S. and Prieto, Carlos Allende and An, Deokkeun and Anderson, Kurt S. J. and Anderson, Scott F. and Annis, James and Bahcall, Neta A. and et al.},
+   year          = {2009},
+   month         = {5},
+   pages         = {543–558}
+}
+
+@ARTICLE{Cython,
+    author  = {S. {Behnel} and R. {Bradshaw} and C. {Citro} and L. {Dalcin} and D. S. {Seljebotn} and K. {Smith}},
+    journal = {Computing in Science   Engineering}, 
+    title   = {Cython: The Best of Both Worlds}, 
+    year    = {2011},
+    volume  = {13},
+    number  = {2},
+    pages   = {31-39}
+}
+
+@misc{DESI,
+    Author = {Michael Levi and Chris Bebek and Timothy Beers and Robert Blum and Robert Cahn and Daniel Eisenstein and Brenna Flaugher and Klaus Honscheid and Richard Kron and Ofer Lahav and Patrick McDonald and Natalie Roe and David Schlegel and representing the DESI collaboration},
+    Title  = {The DESI Experiment, a whitepaper for Snowmass 2013},
+    Year   = {2013},
+    Eprint = {arXiv:1308.0847}
+}
+
+@article{El-Ad:1997,
+    author        = {{El-Ad}, Hagai and {Piran}, Tsvi},
+    title         = "{Voids in the Large-Scale Structure}",
+    journal       = "ApJ",
+    keywords      = {Cosmology: Observations, Galaxies: Clusters: General, Cosmology: Large-Scale Structure of Universe, Methods: Data Analysis, Astrophysics},
+    year          = 1997,
+    month         = dec,
+    volume        = {491},
+    number        = {2},
+    pages         = {421-435},
+    doi           = {10.1086/304973},
+    archivePrefix = {arXiv},
+    eprint        = {astro-ph/9702135},
+    primaryClass  = {astro-ph}
+}
+
+@article{Hoyle:2002,
+    title     = {Voids in the point source catalogue survey and the updated zwicky catalog},
+    author    = {Hoyle, Fiona and Vogeley, Michael S},
+    journal   = {ApJ},
+    volume    = {566},
+    number    = {2},
+    pages     = {641},
+    year      = {2002},
+    publisher = {IOP Publishing}
+}
+
+@article{Neyrinck:2007,
+    author        = "Neyrinck, Mark C.",
+    title         = "{ZOBOV: a parameter-free void-finding algorithm}",
+    journal       = "Mon. Not. Roy. Astron. Soc.",
+    volume        = "386",
+    year          = "2008",
+    pages         = "2101",
+    doi           = "10.1111/j.1365-2966.2008.13180.x",
+    eprint        = "0712.3049",
+    archivePrefix = "arXiv",
+    primaryClass  = "astro-ph"
+}
+
+@misc{REVOLVER:2018,
[…]
