+++ doc/paper.md
@@ -1,0 +1,171 @@
+---
+title: 'VAST: the Void Analysis Software Toolkit'
+tags:
+  - Python
+  - astronomy
+  - voids
+  - large-scale structure
+  - redshift survey
+authors:
+  - name: Kelly A. Douglass
+    orcid: 0000-0002-9540-546X
+    affiliation: 1
+  - name: Dylan Veyrat
+    orcid: 0000-0001-8101-2836
+    affiliation: 1
+  - name: Stephen W. O'Neill, Jr.
+    orcid: 0000-0003-1366-7325
+    affiliation: 2
+  - name: Segev BenZvi
+    orcid: 0000-0001-5537-4710
+    affiliation: 1
+  - name: Fatima Zaidouni
+    orcid: 0000-0003-0931-0868
+    affiliation: 1
+  - name: Michaela Guzzetti
+    orcid: 0000-0001-6866-8291
+    affiliation: "1, 3, 4"
+affiliations:
+  - name: University of Rochester
+    index: 1
+  - name: Independent Researcher
+    index: 2
+  - name: Smith College
+    index: 3
+  - name: University of Washington
+    index: 4
+date: December 2020
+bibliography: paper.bib
+---
+
+
+# Summary
+
+Voids are expansive regions in the universe containing significantly fewer 
+galaxies than surrounding galaxy clusters and filaments.  They are a fundamental 
+feature of the cosmic web and provide important information about galaxy physics 
+and cosmology.  For example, correlations between voids and luminous tracers of 
+large-scale structure improve constraints on the expansion of the universe as 
+compared to using tracers alone.  However, what constitutes a void is vague and 
+formulating a concrete definition to use in a void-finding algorithm is not 
+trivial.  As a result, several different algorithms exist to identify these 
+cosmic underdensities.  Our Void Analysis Software Toolkit, or VAST, provides 
+Python 3 implementations of two such algorithms: **VoidFinder** and 
+**V<sup>2</sup>**.
+
+
+# Statement of Need
+
+Analyzing the next generation of cosmological surveys will require the ability 
+to process large volumes of data (for example, at least 30 million galaxies and 
+quasars from the Dark Energy Spectroscopic Instrument [@DESI]).  To more 
+efficiently process such large datasets, this Python 3 implementation of 
+**VoidFinder** includes a Cythonized [@Cython] version of the algorithm which 
+also allows for multi-process void-finding.  The void-finding algorithm in the 
+`Vsquared` package uses the `scipy.spatial` [@SciPy] submodule for fast 
+computation of the Voronoi tessellation and convex hulls involved in the 
+algorithm.
+
+
+
+
+# VoidFinder
+
+`vast.voidfinder` is a software package containing a Python 3 implementation of 
+the Void Finder algorithm [@El-Ad:1997] that is based on the algorithm's Fortran 
+implementation by @Hoyle:2002.  Motivated by the expectation that voids are 
+spherical to first order, this algorithm defines voids as the unions of sets of 
+spheres grown in the underdense regions of the large-scale structure.  It begins 
+by removing all isolated tracers from the catalog of objects, defined as having 
+significantly ($1.5\sigma$) larger than average third-nearest neighbor 
+distances.  The remaining tracers are then placed on a grid, and spheres are 
+grown from the centers of the empty cells until they are bounded by four tracers 
+on their surfaces.  All spheres larger than a specified radius (typically 
+around 10 Mpc/h) are considered possible maximal spheres -- the largest sphere 
+that can fit in a given void region.  Filtering through these candidate maximal 
+spheres by order of decreasing radius, no maximal sphere can overlap by more 
+than 10% of its volume with any other previously identified (larger) maximal 
[…]
