+++ Vsquared/vast/vsquared/__init__.py
@@ -1,13 +1,0 @@
-# Licensed under a 3-clause BSD-style license - see LICENSE.
-# -*- coding: utf-8 -*-
-"""
-========
-Vsquared
-========
-Implementation of the galaxy void finder algorithm ZOnes Bordering On Voidness
-(ZOBOV) by Mark Neyrinck, MNRAS 386:2101-2109, 2007.
-"""
-from __future__ import absolute_import
-
-name = 'vsquared'
-__version__ = '0.3.0'
