### 2022-02-08 1fc60d6
Removing obsolete Documentation folder
