+++ Documentation/source/requirements.txt
@@ -1,4 +1,0 @@
-numpydoc
-sphinx-argparse
-Cython
-numpy
