+++ Documentation/source/VoidFinder_functions.rst
@@ -1,45 +1,0 @@
-####################
-VoidFinder API
-####################
-
-
-
-``file_preprocess``
-===================
-
-.. autofunction:: vast.voidfinder.preprocessing.file_preprocess
-
-
-
-``filter_galaxies``
-===================
-
-.. autofunction:: vast.voidfinder.filter_galaxies
-
-
-``ra_dec_to_xyz``
-=================
-
-.. autofunction:: vast.voidfinder.ra_dec_to_xyz
-
-
-``z_to_comoving_dist``
-======================
-
-.. autofunction:: vast.voidfinder.distance.z_to_comoving_dist
-
-
-
-.. _VF-mask:
-
-``generate_mask``
-=================
-
-.. autofunction:: vast.voidfinder.multizmask.generate_mask
-
-
-
-``find_voids``
-==============
-
-.. autofunction:: vast.voidfinder.find_voids
