### 2022-02-08 a164279
Remove obsolete top-level VoidFinder directory
