+++ VoidFinder/vast/voidfinder/voidfinder.py
@@ -1,916 +1,0 @@
-#VoidFinder Function to do just about everything
-
-
-import numpy as np
-
-from astropy.table import Table
-
-import time
-
-from .hole_combine import combine_holes, combine_holes_2
-
-from .voidfinder_functions import mesh_galaxies, \
-                                  in_mask, \
-                                  not_in_mask, \
-                                  in_survey, \
-                                  save_maximals, \
-                                  mesh_galaxies_dict
-                                  #build_mask, \
-
-from .table_functions import add_row, \
-                             subtract_row, \
-                             to_vector, \
-                             to_array, \
-                             table_dtype_cast, \
-                             table_divide
-
-from .volume_cut import volume_cut, check_hole_bounds
-
-from .avsepcalc import av_sep_calc
-
-from .mag_cutoff_function import mag_cut, field_gal_cut
-
-from ._voidfinder import _hole_finder
-
-from .constants import c
-
-from ._voidfinder_cython import check_mask_overlap
-
-from ._voidfinder_cython_find_next import MaskChecker
-
-
-################################################################################
-DtoR = np.pi/180.
-RtoD = 180./np.pi
-################################################################################
-
-
-
-def filter_galaxies(galaxy_table,
-                    survey_name, 
-                    out_directory,
-                    mag_cut=True, 
-                    dist_limits=None,
-                    rm_isolated=True,
-                    write_table=True,
-                    sep_neighbor=3,
-                    dist_metric='comoving', 
-                    h=1.0,
-                    hole_grid_edge_length=5.0,
-                    magnitude_limit=-20.09,
-                    verbose=0):
-    """
-    A hodge podge of miscellaneous tasks which need to be done to format the data into
-    something the main find_voids() function can use.
-    
-    1) Optional magnitude cut
-    2) Convert from ra-dec-redshift space into xyz space
-    3) Calculate the hole search grid shape
-    4) Optional remove isolated galaxies by partitioning them into wall (non-isolated)
-       and field (isolated) groups
-    5) Optionally write out the wall and field galaxies to disk
-    
-    
-    Parameters
-    ==========
-    
-    galaxy_table : astropy.table of shape (N,?)
-        variable number of required columns.  If doing magnitude cut, must include
-        'rabsmag' column. If distance metric is 'comoving', must include 'Rgal'
-        column, otherwise must include 'redshift'.  Also must always include 'ra' 
-        and 'dec'
-        
-    survey_name : str
-        Name of the galxy catalog, string value to prepend or append to output names
-
-    out_directory : string
-        Directory path for output files
-        
-    mag_cut : bool
-        whether or not to cut on magnitude, removing galaxies less than
-        magnitude_limit
-
-    dist_limits : list of length 2
-        [Minimum distance, maximum distance] of galaxy sample (in units of Mpc/h)
-        
-    magnitude_limit : float
-        value at which to perform magnitude cut
-        
-    rm_isolated : bool
-        whether or not to perform Nth neighbor distance calculation, and use it
-        to partition the input galaxies into wall and field galaxies
-        
-    write_table : bool
-        use astropy.table.Table.write to write out the wall and field galaxies
-        to file
-        
-    sep_neighbor : int, positive
-        if rm_isolated_flag is true, find the Nth galaxy neighbors based on this value
-        
-    dist_metric : str
-        Distance metric to use in calculations.  Options are 'comoving' 
-        (default; distance dependent on cosmology) and 'redshift' (distance 
-        independent of cosmology).
-
-    h : float
[…]
