+++ VoidFinder/vast/voidfinder/preprocessing.py
@@ -1,258 +1,0 @@
-
-import numpy as np
-
-from astropy.table import Table
-from astropy.io import ascii, fits
-
-from vast.voidfinder.distance import z_to_comoving_dist
-
-import time
-
-import h5py
-
-from vast.voidfinder.constants import c #speed of light
-
-import os
-
-
-
-def load_data_to_Table(input_filepath):
-    """
-    Load a .dat, .txt, .fits, or .h5 file representing a table of numerical 
-    values.  For HDF5 (.h5) files, assumes the "columns" are stored as 
-    individual datasets on the root h5py.File object.
-    
-    If the input filepath ends with a .h5 extension, this will attempt to read 
-    it in as the HDF5 format, otherwise for all other extensions it will attempt 
-    the normal Table.read()
-    
-    Parameters
-    ==========
-    
-    input_filepath : str
-        path to the input .dat, .txt, .fits, or .h5 file
-        
-    Returns
-    =======
-    
-    astropy.table.Table
-    """
-        
-    if input_filepath.endswith(".h5"):
-        
-        infile = h5py.File(input_filepath, 'r')
-        
-        data_table = Table()
-        
-        for col in list(infile.keys()):
-            
-            data_table[col] = infile[col][()]
-            
-        infile.close()
-
-    elif input_filepath.endswith('.fits') or input_filepath.endswith('.fit'):
-
-        hdu = fits.open(input_filepath)
-
-        data_table = Table(hdu[1].data)
-
-        hdu.close()
-        
-    else:
-        
-        if os.path.getsize(input_filepath) < 5e9:
-            
-            data_table = Table.read(input_filepath, format='ascii.commented_header')
-
-        else:
-        
-            # Import header line
-            data_table_fobj = open(input_filepath, 'r')
-
-            header_line = data_table_fobj.readline()
-
-            data_table_fobj.close()
-
-            # Parse header line
-            col_names = header_line.split(' ')
-
-            # Read in the data in 100 Mb chunks
-            data_table = ascii.read(input_filepath, 
-                                    format='no_header', 
-                                    names=col_names[1:], 
-                                    guess=False, 
-                                    fast_reader={'chunk_size': 100*1000000})
-        
-    return data_table
-    
-    
-    
-
-
-def file_preprocess(galaxies_filename, 
-                    in_directory, 
-                    out_directory, 
-                    mag_cut=True,
-                    rm_isolated=True,
-                    dist_metric='comoving', 
-                    min_z=None,
-                    max_z=None,
-                    Omega_M=0.3,
-                    h=1.0,
-                    verbose=0):
-    '''
-    Set up output file names, calculate distances, etc.
-    
-    
-    PARAMETERS
-    ==========
-    
-    galaxies_filename : string
-        File name of galaxy catalog.  Should be readable by 
-        astropy.table.Table.read as a ascii.commented_header file.  Required 
-        columns include 'ra', 'dec', 'z', and absolute magnitude (either 
-        'rabsmag' or 'magnitude'.
-        
-    in_directory : string
-        Directory path for input files
-    
-    out_directory : string
-        Directory path for output files
-        
-    mag_cut : boolean
-        Determines whether or not to implement a magnitude cut on the galaxy 
-        survey.  Default is True (remove all galaxies fainter than Mr = -20).
-        
-    rm_isolated : boolean
-        Determines whether or not to remove isolated galaxies (defined as those 
-        with the distance to their third nearest neighbor greater than the sum 
-        of the average third-nearest-neighbor distance and 1.5 times the 
-        standard deviation of the third-nearest-neighbor distances).
-    
-    dist_metric : string
-        Description of which distance metric to use.  Options should include 
-        'comoving' (default) and 'redshift'.
-        
-    min_z, max_z : float
[…]
