### 2022-02-08 22f2172
Moved example scripts to top level
