+++ example_scripts/DR7_config.ini
@@ -1,8 +1,8 @@
 [Paths]
 #Currently only accepts FITS input
-Input Catalog    = ../data/vollim_dr7_cbp_102709.fits
+Input Catalog    = vollim_dr7_cbp_102709.fits
 Survey Name      = DR7
-Output Directory = ../data/
+Output Directory = ./
 
 [Cosmology]
 #Currently only uses Flat LambdaCDM
