### 2022-02-08 9e80127
Minor changes to docstring of find_voids
