+++ README.md
@@ -29,6 +29,5 @@ for 3D OpenGL-based visualization of VoidFinder's voids in SDSS DR7!
 
 **V<sup>2</sup>** is a voronoi-tesselation-based algorithm for finding the void 
 regions, based on the ZOBOV algorithm.  ZOBOV uses the gradient of the volume of 
-adjacent voronoi cells to flow multiple cells together into large void regions.  
-This code is still under development.
+adjacent voronoi cells to flow multiple cells together into large void regions.
 
