### 2022-02-08 6398a66
Install VAST for unit test.
