### 2022-02-09 9b20afa
Different call to add sys.path to cwd.
