### 2022-02-09 756e21b
Use pytest-runner for cythonized modules.
