+++ .github/workflows/tests.yml
@@ -27,7 +27,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m pip install flake8 pytest
+        python -m pip install flake8 pytest pytest-runner
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Install VAST
       run: |
@@ -40,4 +40,4 @@ jobs:
 #        flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
       run: |
-        python -m pytest
+        python setup.py pytest
