### 2022-02-09 7eb8740
Temporarily disable comparison of holes.
