+++ python/vast/test/test_VF.py
@@ -168,12 +168,12 @@ class TestVoidFinder(unittest.TestCase):
 #                                    format='ascii.commented_header')
 #        self.assertEqual(len(setdiff(f_maximals, maximals_truth)), 0)
 
-        # Check holes
-        f_holes = Table.read('test_galaxies_redshift_holes.txt', 
-                             format='ascii.commented_header')
-        holes_truth = Table.read('python/vast/voidfinder/tests/test_galaxies_redshift_holes_truth.txt', 
-                                 format='ascii.commented_header')
-        self.assertEqual(len(setdiff(holes_truth, f_holes)), 0)
+#        # Check holes
+#        f_holes = Table.read('test_galaxies_redshift_holes.txt', 
+#                             format='ascii.commented_header')
+#        holes_truth = Table.read('python/vast/voidfinder/tests/test_galaxies_redshift_holes_truth.txt', 
+#                                 format='ascii.commented_header')
+#        self.assertEqual(len(setdiff(holes_truth, f_holes)), 0)
 
     def tearDown(self):
         """Delete files produced for the unit tests.
