+++ doc/source/conf.py
@@ -8,6 +8,8 @@ import sys
 import os
 import re
 
+import sphinx_rtd_theme
+
 sys.path.insert(0, os.path.abspath('../../python/'))
 import vast
 
@@ -36,6 +38,7 @@ extensions = [
     'sphinx.ext.ifconfig',
     'sphinx.ext.viewcode',
     'sphinx.ext.githubpages',
+    'sphinx_rtd_theme',
     'numpydoc',
     'sphinxarg.ext'
     #'pyquickhelper.sphinxext.sphinx_autosignature'
@@ -58,7 +61,7 @@ exclude_patterns = []
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'classic'
+html_theme = 'sphinx_rtd_theme'
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
