+++ .readthedocs.yaml
@@ -2,15 +2,18 @@
 # Read the Docs configuration file
 # See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
 
-# Required
 version: 2
+    
+# Python version and requirements
+python:
+    version: 3.8
+    install:
+    - method: pip
+      path: .
+      extra_requirements:
+        - docs
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
     configuration: doc/source/conf.py
-    
-# Python version and requirements
-python:
-    install:
-    - requirements: doc/source/requirements.txt
-    - method: pip
+    fail_on_warning: true
