+++ .readthedocs.yaml
@@ -16,4 +16,4 @@ python:
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
     configuration: doc/source/conf.py
-    fail_on_warning: true
+    fail_on_warning: false
