### 2022-02-13 58086fb
Modified code with zero-array error in filter_galaxies
