### 2022-03-03 15007ce
voidfinder.py code, removed all the unecessary prints
