### 2022-03-12 DESI-UR/VAST#51 @QuiteAFoxtrot
Was able to reproduce the issue
