+++ README.md
@@ -11,6 +11,8 @@ catalogs:
 1. Void identification using watershed algorithms.
 
 
+In case you missed the link above, the docs can be found here: https://vast.readthedocs.io/en/latest/
+
 ## VoidFinder
 
 **VoidFinder** is an algorithm which utilizes a sphere-growing method on a grid 
