+++ python/vast/voidfinder/_hole_combine_cython.pyx
@@ -1,3 +1,4 @@
+#cython: language_level=3
 
 from __future__ import print_function
 
@@ -10,7 +11,7 @@ cimport numpy as np
 np.import_array()  # required in order to use C-API
 
 
-from typedefs cimport DTYPE_CP128_t, \
+from .typedefs cimport DTYPE_CP128_t, \
                       DTYPE_CP64_t, \
                       DTYPE_F64_t, \
                       DTYPE_F32_t, \
