+++ python/vast/voidfinder/_voidfinder.py
@@ -62,168 +62,6 @@ import matplotlib.pyplot as plt
 
 
 
-
-def _hole_finder(hole_grid_shape, 
-                 hole_grid_edge_length, 
-                 hole_center_iter_dist,
-                 galaxy_map_grid_edge_length,
-                 coord_min, 
-                 galaxy_coords,
-                 survey_name,
-                 mask_mode=0,
-                 mask=None,
-                 mask_resolution=None,
-                 min_dist=None,
-                 max_dist=None,
-                 xyz_limits=None,
-                 check_only_empty_holes=True,
-                 #hole_radial_mask_check_dist,
-                 save_after=None,
-                 use_start_checkpoint=False,
-                 batch_size=1000,
-                 verbose=0,
-                 print_after=5.0,
-                 num_cpus=1):
-    
-    '''
-    Description
-    ===========
-
-    See help(voidfinder.find_voids)
-
-    This function is basically a glorified switch between single-threaded mode 
-    and multi-processed mode of running VoidFinder.
-
-    
-    Parameters
-    ==========
-    
-    hole_grid_shape : array or tuple of length 3
-        the number of grid cells in each of the 3 x,y,z dimensions
-    
-    hole_grid_edge_length : scalar float
-        length of each cell in Mpc/h
-        
-    hole_center_iter_dist : scalar float
-        distance to shift hole centers during iterative void hole growing in 
-        Mpc/h
-        
-        
-    galaxy_map_grid_edge_length : float or None
-        edge length in Mpc/h for the secondary grid for finding nearest neighbor 
-        galaxies.  If None, will default to 3*void_grid_edge_length (which 
-        results in a cell volume of 3^3 = 27 times larger cube volume).  This 
-        parameter yields a tradeoff between number of galaxies in a cell, and 
-        number of cells to search when growing a sphere.  Too large and many 
-        redundant galaxies may be searched, too small and too many cells will 
-        need to be searched.
-        (xyz space)
-        
-    coord_min : numpy.ndarray of shape (1,3) 
-        minimum coordinates of the survey in x,y,z in Mpc/h
-
-        Note that this coordinate is used for transforming values into the i,j,k 
-        search grid space and also into the p,q,r galaxy map grid space
-        
-    mask : numpy.ndarray of shape (N,M) type bool
-        represents the survey footprint in scaled ra/dec space.  Value of True 
-        indicates that a location is within the survey
-
-    mask_resolution : integer
-        Scale factor of coordinates needed to index mask
-    
-    min_dist : float
-        minimum redshift in units of Mpc/h
-        
-    max_dist : float
-        maximum redshift in units of Mpc/h
-        
-    galaxy_coords : numpy.ndarray of shape (num_galaxies, 3)
-        coordinates of the galaxies in the survey, units of Mpc/h
-        (xyz space)
-        
-    survey_name : str
-        identifier for the survey running, may be prepended or appended to 
-        output filenames including the checkpoint filename
-        
-    DEPRECATED hole_radial_mask_check_dist : float in (0.0,1.0)
-        radial distance to check whether or not a hole overlaps with outside the 
-        mask too much
-        
-    save_after : int or None
-        save a VoidFinderCheckpoint.h5 file after *approximately* every 
-        save_after cells have been processed.  This will over-write this 
-        checkpoint file every save_after cells, NOT append to it.  Also, saving 
-        the checkpoint file forces the worker processes to pause and synchronize 
-        with the master process to ensure the correct values get written, so 
-        choose a good balance between saving too often and not often enough if 
-        using this parameter.  Note that it is an approximate value because it 
-        depends on the number of worker processes and the provided batch_size 
-        value, if you batch size is 10,000 and your save_after is 1,000,000 you 
[…]

[…]
