### 2022-03-13 d00eb82
Consolidating parameters.  Updated cython language_level to 3.  Added resize() methods to custom dict class to help enable shared memory resizing for periodic boundary conditions implementation
