+++ python/vast/voidfinder/_voidfinder.py
@@ -5,24 +5,21 @@
 import os
 import stat
 import sys
-import fcntl
 import mmap
 import struct
 import socket
 import select
 import atexit
-import signal
 import tempfile
 import multiprocessing
 import h5py
+import pickle
+import time
 from psutil import cpu_count
-
 
 import cProfile
 
 import numpy as np
-
-import time
 
 from .voidfinder_functions import not_in_mask
 
@@ -30,33 +27,21 @@ from ._voidfinder_cython import main_algorithm, \
                                 fill_ijk, \
                                 fill_ijk_zig_zag
 
-
 from ._voidfinder_cython_find_next import GalaxyMap, \
                                           Cell_ID_Memory, \
                                           GalaxyMapCustomDict, \
                                           HoleGridCustomDict, \
                                           NeighborMemory, \
-                                          MaskChecker
-
-
+                                          MaskChecker, \
+                                          find_next_prime
 
 from multiprocessing import Queue, Process, RLock, Value, Array
 
 from ctypes import c_int64, c_double, c_float
 
-from queue import Empty
-
-from copy import deepcopy
-
-import pickle
-
-from astropy.table import Table
-
-from .table_functions import to_array
-
-import matplotlib
+#import matplotlib
 #matplotlib.use('TkAgg')
-import matplotlib.pyplot as plt
+#import matplotlib.pyplot as plt
 
 
 
@@ -218,11 +203,9 @@ def SaveCheckpointFile(checkpoint_filepath,
 
 
 
-def _hole_finder(hole_grid_shape, 
-                 hole_grid_edge_length, 
+def _hole_finder(hole_grid_edge_length, 
                  hole_center_iter_dist,
                  galaxy_map_grid_edge_length,
-                 coord_min, 
                  galaxy_coords,
                  survey_name,
                  mask_mode=0,
@@ -402,6 +385,35 @@ def _hole_finder(hole_grid_shape,
         
         mask = mask.astype(np.uint8)
     
+        coords_max = np.max(galaxy_coords, axis=0)
+    
+        coords_min = np.min(galaxy_coords, axis=0)
+        
+        box = coords_max - coords_min
+    
+        ngrid = box/hole_grid_edge_length
+        
+        #print("Ngrid: ", ngrid)
+        
+        hole_grid_shape = tuple(np.ceil(ngrid).astype(int))
+        
+    else:
+        
+        box = xyz_limits[1,:] - xyz_limits[0,:]
+        
+        ngrid = box/hole_grid_edge_length
+        
+        hole_grid_shape = tuple(np.ceil(ngrid).astype(int))
+        
+        coords_min = xyz_limits[0,:]
+    
+    
+    
+    coords_min = coords_min.reshape(1,3)
+    
+    if galaxy_map_grid_edge_length is None:
+        
+        galaxy_map_grid_edge_length = 3.0*hole_grid_edge_length
     
     
     
@@ -434,8 +446,6 @@ def _hole_finder(hole_grid_shape,
         num_cpus = cpu_count(logical=False)
         
     if verbose > 0:
-        
-        print("Running multi-process mode,", str(num_cpus), "cpus", flush=True)
         
         print("Grid: ", hole_grid_shape, flush=True)
     ############################################################################
@@ -494,47 +504,31 @@ def _hole_finder(hole_grid_shape,
     # First build a helper for the i,j,k generator, using the hole grid edge 
     # length.  We basically need a flag that says "there is a galaxy in this ijk 
     # cell" so VoidFinder can skip that i,j,k value when growing holes
+    #
[…]
