### 2022-03-14 10e4db6
Updated public interface simplifying/removing grid shape and coord_min parameters.  Added logic to check mask and grid search edge length for alignment in periodic mode
