+++ python/vast/voidfinder/_voidfinder.py
@@ -202,11 +202,33 @@ def SaveCheckpointFile(checkpoint_filepath,
 
 
 
+def get_common_divisors(values):
+    
+    
+    min_value = int(numpy.ceil(min(values)/2))
+    
+    divisors = []
+    
+    for idx in range(1, min_value):
+        
+        if all([value % idx == 0 for value in values]):
+        
+            divisors.append(idx)
+            
+    return divisors
+            
+        
 
-def _hole_finder(hole_grid_edge_length, 
+
+
+
+
+
+
+def _hole_finder(galaxy_coords,
+                 hole_grid_edge_length, 
                  hole_center_iter_dist,
                  galaxy_map_grid_edge_length,
-                 galaxy_coords,
                  survey_name,
                  mask_mode=0,
                  mask=None,
@@ -246,13 +268,13 @@ def _hole_finder(hole_grid_edge_length,
     5). Start the workers
     6). Make sure workers connect to the comm socket
     7). Checkpoint the progress if those parameters are enabled
-    8). Collect progress results from the workers
+    8). Collect & print progress results from the workers
     
     This function is designed to be run on Linux on an SMP (Symmetric 
     Multi-Processing) architecture.  It takes advantage of 2 Linux-specific 
     properties: the /dev/shm filesystem and the fork() method of spawning 
     processes. /dev/shm is used as the preferred location for creating memory 
-    maps to share information between the worker processes since on Linux is is 
+    maps to share information between the worker processes since on Linux it is 
     a RAMdisk, and the implementation of fork() on Linux is used to share file 
     descriptor values between the master and worker processes, whereas on 
     mac/OSX fork() is wonky and Windows does not offer fork() at all.  This has 
@@ -266,8 +288,10 @@ def _hole_finder(hole_grid_edge_length,
     Parameters
     ==========
     
-    hole_grid_shape : array or tuple of length 3
-        the number of grid cells in each of the 3 x,y,z dimensions
+    
+    galaxy_coords : numpy.ndarray of shape (num_galaxies, 3)
+        coordinates of the galaxies in the survey, units of Mpc/h
+        (xyz space)
     
     hole_grid_edge_length : scalar float
         length of each cell in Mpc/h
@@ -275,7 +299,6 @@ def _hole_finder(hole_grid_edge_length,
     hole_center_iter_dist : scalar float
         distance to shift hole centers during iterative void hole growing in 
         Mpc/h
-        
         
     galaxy_map_grid_edge_length : float or None
         edge length in Mpc/h for the secondary grid for finding nearest neighbor 
@@ -287,11 +310,13 @@ def _hole_finder(hole_grid_edge_length,
         need to be searched.
         (xyz space)
         
-    coord_min : numpy.ndarray of shape (1,3) 
-        minimum coordinates of the survey in x,y,z in Mpc/h
-
-        Note that this coordinate is used for transforming values into the i,j,k 
-        search grid space and also into the p,q,r galaxy map grid space
+    survey_name : str
+        identifier for the survey running, may be prepended or appended to 
+        output filenames including the checkpoint filename
+        
+    mask_mode : int, one of [0,1,2]
+        Determines which mode VoidFinder is running in with regards to the Mask
+        checking.  0 == 'ra-dec-redshift', 1 == 'xyz' and 2 == 'periodic'
         
     mask : numpy.ndarray of shape (N,M) type bool
         represents the survey footprint in scaled ra/dec space.  Value of True 
@@ -306,17 +331,8 @@ def _hole_finder(hole_grid_edge_length,
     max_dist : float
         maximum redshift in units of Mpc/h
         
-    galaxy_coords : numpy.ndarray of shape (num_galaxies, 3)
-        coordinates of the galaxies in the survey, units of Mpc/h
-        (xyz space)
         
-    survey_name : str
[…]
