### 2022-03-14 a53e490
Updated all references to w_coord to refer to GalaxyMap attribute as an interface so we can implement the periodic switcheroo underneath
