+++ python/vast/voidfinder/_voidfinder_cython.pyx
@@ -300,17 +300,11 @@ cpdef DTYPE_INT64_t fill_ijk(DTYPE_INT64_t[:,:] i_j_k_array,
 @cython.cdivision(True)
 @cython.profile(False)
 cpdef void main_algorithm(DTYPE_INT64_t[:,:] i_j_k_array,
-                          GalaxyMap galaxy_tree,
-                          DTYPE_F64_t[:,:] w_coord,
+                          GalaxyMap galaxy_map,
                           DTYPE_F64_t dl, 
                           DTYPE_F64_t dr,
                           DTYPE_F64_t[:,:] coord_min, 
                           MaskChecker mask_checker,
-                          #DTYPE_B_t[:,:] mask,
-                          #DTYPE_INT32_t mask_resolution,
-                          #DTYPE_F64_t min_dist,
-                          #DTYPE_F64_t max_dist,
-                          #DTYPE_F64_t hole_radial_mask_check_dist,
                           DTYPE_F64_t[:,:] return_array,
                           Cell_ID_Memory cell_ID_mem,
                           NeighborMemory neighbor_mem,
@@ -346,12 +340,11 @@ cpdef void main_algorithm(DTYPE_INT64_t[:,:] i_j_k_array,
         a batch of N cell IDs on which to run the VoidFinder hole-growing 
         algorithm
       
-    galaxy_tree : python object
-        a glorified wrapper around a couple of the data structures used by 
-        VoidFinder
+    galaxy_map : _voidfinder_cython_find_next.GalaxyMap
+        An interface to searching, accessing, and retrieving the galaxies
+        in the survey efficiently.  handles logic like regular vs. periodic
+        access, and memory mapping the arrays for the workers
         
-    w_coord : memoryview to (K,3) numpy array
-        xyz space galaxy coordinates
         
     dl : float
         edge length in Mpc/h of the hole grid cell size
@@ -606,14 +599,14 @@ cpdef void main_algorithm(DTYPE_INT64_t[:,:] i_j_k_array,
         #
         ########################################################################
         
-        query_vals = _query_first(galaxy_tree.reference_point_ijk,
-                                  galaxy_tree.coord_min,
-                                  galaxy_tree.dl,
-                                  galaxy_tree.shell_boundaries_xyz,
-                                  galaxy_tree.cell_center_xyz,
-                                  galaxy_tree.galaxy_map,
-                                  galaxy_tree.galaxy_map_array,
-                                  galaxy_tree.w_coord,
+        query_vals = _query_first(galaxy_map.reference_point_ijk,
+                                  galaxy_map.coord_min,
+                                  galaxy_map.dl,
+                                  galaxy_map.shell_boundaries_xyz,
+                                  galaxy_map.cell_center_xyz,
+                                  galaxy_map.galaxy_map,
+                                  galaxy_map.galaxy_map_array,
+                                  galaxy_map.wall_galaxy_coords,
                                   cell_ID_mem,
                                   hole_center_memview)
         
@@ -635,7 +628,7 @@ cpdef void main_algorithm(DTYPE_INT64_t[:,:] i_j_k_array,
         
         for idx in range(3):
             
-            unit_vector_memview[idx] = (w_coord[k1g,idx] - hole_center_memview[0,idx])/vector_modulus
+            unit_vector_memview[idx] = (galaxy_map.wall_galaxy_coords[k1g,idx] - hole_center_memview[0,idx])/vector_modulus
             
         hole_radius = vector_modulus
         
@@ -667,10 +660,10 @@ cpdef void main_algorithm(DTYPE_INT64_t[:,:] i_j_k_array,
                                              dr, 
                                              -1.0,
                                              unit_vector_memview, 
-                                             galaxy_tree, 
+                                             galaxy_map, 
[…]
