+++ python/vast/voidfinder/_voidfinder_cython_find_next.pyx
@@ -55,10 +55,9 @@ cdef FindNextReturnVal find_next_galaxy(DTYPE_F64_t[:,:] hole_center_memview,
                                         DTYPE_F64_t dr, 
                                         DTYPE_F64_t direction_mod,
                                         DTYPE_F64_t[:] unit_vector_memview, 
-                                        galaxy_tree, 
+                                        GalaxyMap galaxy_tree, 
                                         DTYPE_INT64_t[:] nearest_gal_index_list, 
                                         ITYPE_t num_neighbors,
-                                        DTYPE_F64_t[:,:] w_coord, 
                                         MaskChecker mask_checker,
                                         DTYPE_F64_t[:] Bcenter_memview,
                                         Cell_ID_Memory cell_ID_mem,
@@ -68,6 +67,8 @@ cdef FindNextReturnVal find_next_galaxy(DTYPE_F64_t[:,:] hole_center_memview,
     '''
     Description:
     ============
+    
+    
     Function to locate the next nearest galaxy during hole center propagation 
     along direction defined by unit_vector_memview.
 
@@ -291,7 +292,7 @@ cdef FindNextReturnVal find_next_galaxy(DTYPE_F64_t[:,:] hole_center_memview,
             
             for idx in range(3):
                 
-                temp_f64_val = w_coord[nearest_gal_index_list[0],idx] - temp_hole_center_memview[0,idx]
+                temp_f64_val = galaxy_tree.wall_galaxy_coords[nearest_gal_index_list[0],idx] - temp_hole_center_memview[0,idx]
                 
                 temp_f64_accum += temp_f64_val*temp_f64_val
                 
@@ -312,7 +313,7 @@ cdef FindNextReturnVal find_next_galaxy(DTYPE_F64_t[:,:] hole_center_memview,
         #print("Loc-C1", flush=True)
         
         _query_shell_radius(galaxy_tree.reference_point_ijk,
-                            galaxy_tree.w_coord,
+                            galaxy_tree.wall_galaxy_coords,
                             galaxy_tree.coord_min,
                             galaxy_tree.dl, 
                             galaxy_tree.galaxy_map,
@@ -410,14 +411,14 @@ cdef FindNextReturnVal find_next_galaxy(DTYPE_F64_t[:,:] hole_center_memview,
                     
                     if num_neighbors == 1:
                         
-                        neighbor_mem.candidate_minus_A[3*idx+jdx] = w_coord[nearest_gal_index_list[0], jdx] - w_coord[temp_idx, jdx]
+                        neighbor_mem.candidate_minus_A[3*idx+jdx] = galaxy_tree.wall_galaxy_coords[nearest_gal_index_list[0], jdx] - galaxy_tree.wall_galaxy_coords[temp_idx, jdx]
                         
                         
                     else:
 
-                        neighbor_mem.candidate_minus_A[3*idx+jdx] = w_coord[temp_idx, jdx] - w_coord[nearest_gal_index_list[0], jdx]
+                        neighbor_mem.candidate_minus_A[3*idx+jdx] = galaxy_tree.wall_galaxy_coords[temp_idx, jdx] - galaxy_tree.wall_galaxy_coords[nearest_gal_index_list[0], jdx]
                         
-                    neighbor_mem.candidate_minus_center[3*idx+jdx] = w_coord[temp_idx, jdx] - hole_center_memview[0, jdx]
+                    neighbor_mem.candidate_minus_center[3*idx+jdx] = galaxy_tree.wall_galaxy_coords[temp_idx, jdx] - hole_center_memview[0, jdx]
 
             ############################################################################
             # Calculate bottom of ratio to be minimized
[…]

[…]
