### 2022-03-14 e717bb5
Sanitized all uses of former w_coord and galaxy_map_array to use GalaxyMap interface instead for the ol periodic switcheroo
