+++ python/vast/voidfinder/_voidfinder_cython_find_next.pyx
@@ -313,11 +313,9 @@ cdef FindNextReturnVal find_next_galaxy(DTYPE_F64_t[:,:] hole_center_memview,
         #print("Loc-C1", flush=True)
         
         _query_shell_radius(galaxy_tree.reference_point_ijk,
-                            galaxy_tree.wall_galaxy_coords,
                             galaxy_tree.coord_min,
                             galaxy_tree.dl, 
-                            galaxy_tree.galaxy_map,
-                            galaxy_tree.galaxy_map_array,
+                            galaxy_tree,
                             cell_ID_mem,
                             neighbor_mem,
                             temp_hole_center_memview, 
@@ -1708,6 +1706,49 @@ cdef class GalaxyMap:
         
         self.gma_fd = gma_fd
         
+        
+        
+    @cython.boundscheck(False)
+    @cython.wraparound(False)
+    @cython.cdivision(True)
+    cdef DTYPE_B_t contains(self,
+                            CELL_ID_t i, 
+                            CELL_ID_t j, 
+                            CELL_ID_t k):
+                            
+        return self.galaxy_map.contains(i, j, k)
+        
+        
+        
+        
+        
+    @cython.boundscheck(False)
+    @cython.wraparound(False)
+    @cython.cdivision(True)
+    cdef OffsetNumPair getitem(self,
+                               CELL_ID_t i, 
+                               CELL_ID_t j, 
+                               CELL_ID_t k):
+                                
+        return self.galaxy_map.getitem(i, j, k)
+        
+        
+        
+    @cython.boundscheck(False)
+    @cython.wraparound(False)
+    @cython.cdivision(True)
+    cdef void setitem(self, 
+                      CELL_ID_t i,
+                      CELL_ID_t j,
+                      CELL_ID_t k, 
+                      DTYPE_INT64_t offset,
+                      DTYPE_INT64_t num_elements):
+                       
+        self.galaxy_map.setitem(i, j, k, offset, num_elements)
+        
+        
+        
+        
     def close(self):
         
         self.wall_galaxy_buffer.close()
@@ -2124,9 +2165,7 @@ cdef DistIdxPair _query_first(CELL_ID_t[:,:] reference_point_ijk,
                               DTYPE_F64_t dl,
                               DTYPE_F64_t[:,:] shell_boundaries_xyz,
                               DTYPE_F64_t[:,:] cell_center_xyz,
-                              GalaxyMapCustomDict galaxy_map,
-                              DTYPE_INT64_t[:] galaxy_map_array,
-                              DTYPE_F64_t[:,:] w_coord,
+                              GalaxyMap galaxy_map,
                               Cell_ID_Memory cell_ID_mem,
                               DTYPE_F64_t[:,:] reference_point_xyz
                               ):
@@ -2251,9 +2290,9 @@ cdef DistIdxPair _query_first(CELL_ID_t[:,:] reference_point_ijk,
             
             for idx in range(num_elements):
                 
-                potential_neighbor_idx = <ITYPE_t>galaxy_map_array[offset+idx]
+                potential_neighbor_idx = <ITYPE_t>(galaxy_map.galaxy_map_array[offset+idx])
                 
-                potential_neighbor_xyz = w_coord[potential_neighbor_idx]
+                potential_neighbor_xyz = galaxy_map.wall_galaxy_coords[potential_neighbor_idx]
                 
                 temp1 = potential_neighbor_xyz[0] - reference_point_xyz[0,0]
                 temp2 = potential_neighbor_xyz[1] - reference_point_xyz[0,1]
@@ -2297,11 +2336,9 @@ cdef DistIdxPair _query_first(CELL_ID_t[:,:] reference_point_ijk,
[…]
