+++ python/vast/voidfinder/_hole_combine_cython.pyx
@@ -29,6 +29,7 @@ from libc.math cimport fabs, sqrt, asin, atan, ceil#, exp, pow, cos, sin, asin
 #from ._voidfinder import find_next_prime
 
 from ._voidfinder_cython_find_next cimport GalaxyMapCustomDict, \
+                                           GalaxyMap, \
                                            Cell_ID_Memory, \
                                            _gen_cube, \
                                            OffsetNumPair, \
@@ -535,7 +536,7 @@ def find_maximals_3(DTYPE_F64_t[:,:] x_y_z_r_array,
         
     
     
-    new_galaxy_map = GalaxyMapCustomDict((n_grid_x, n_grid_y, n_grid_z), 
+    cust_galaxy_map = GalaxyMapCustomDict((n_grid_x, n_grid_y, n_grid_z), 
                                          RESOURCE_DIR)
     
     
@@ -543,15 +544,29 @@ def find_maximals_3(DTYPE_F64_t[:,:] x_y_z_r_array,
     
     for curr_pqr in galaxy_map:
         
-        new_galaxy_map.setitem(curr_pqr[0],
-                               curr_pqr[1],
-                               curr_pqr[2],
-                               offset, 
-                               0)
+        cust_galaxy_map.setitem(curr_pqr[0],
+                           curr_pqr[1],
+                           curr_pqr[2],
+                           offset, 
+                           0)
         
         num_elements = galaxy_map[curr_pqr]
         
         offset += num_elements
+        
+        
+    # Using a dummy workaround for now, changed the code to use the
+    # GalaxyMap interface instead of the GalaxyMapCustomDict directly,
+    # so instantiate the interface with some dummy variables here
+    # and the real galaxy map custom dict we were using before
+    new_galaxy_map = GalaxyMap(RESOURCE_DIR,
+                               0, #mask mode 0 or 1 should be good
+                               np.empty((2,3), dtype=np.float64),
+                               np.empty((1,3), dtype=np.float64),
+                               1.0,
+                               cust_galaxy_map,
+                               np.empty(5, dtype=np.int64)
+                               )
         
     
     
@@ -950,7 +965,7 @@ cpdef np.ndarray join_holes_to_maximals(DTYPE_F64_t[:,:] x_y_z_r_array,
     
     cdef DTYPE_F64_t twice_largest_radius = maximals_info["twice_largest_radius"]
     
-    cdef GalaxyMapCustomDict maximals_map = maximals_info["maximals_map"]
+    cdef GalaxyMap maximals_map = maximals_info["maximals_map"]
     
     cdef DTYPE_INT64_t[:] maximals_map_array = maximals_info["maximals_cell_array"]
     '''
