+++ python/vast/_version.py
@@ -12,6 +12,9 @@ changes and so forth.
 Note that this scheme is still pretty loosey-goosey and ultimately it still falls
 on the package maintainers to debate and find consensus on each issue individually
 but this can provide a useful frame of reference for the versioning.
+
+Also when we hit something like 1.9.x we dont have to go to 2.0.0, we can go to 1.10.x
+just in case anyone wondered about that like I used to.
 """
 
 __version__ = '1.1.1'
