+++ python/vast/voidfinder/_voidfinder_cython_find_next.pyx
@@ -34,7 +34,8 @@ from libc.math cimport fabs, sqrt, asin, atan, ceil#, exp, pow, cos, sin, asin
 import time
 import os
 import tempfile
-from multiprocessing import RLock
+from multiprocessing import RLock, Value
+from ctypes import c_int64
 import mmap
 
 
@@ -1177,6 +1178,7 @@ cdef class HoleGridCustomDict:
         
         os.close(self.lookup_fd)
 
+
 cdef class GalaxyMapCustomDict:
     """
     Description
@@ -1258,6 +1260,7 @@ cdef class GalaxyMapCustomDict:
                         ("r", np.int16, ()),
                         ("offset", np.int64, ()),
                         ("num_elements", np.int64, ())]
+        # 1 + 2 + 2 + 2 + 8 + 8 = 23 bytes per element
         
         self.numpy_dtype = np.dtype(lookup_dtype, align=False)
         
@@ -1272,14 +1275,40 @@ cdef class GalaxyMapCustomDict:
         
         self.lookup_memory[:] = curr_element
         
-        self.num_elements = 0
+        #self.num_elements = 0
+        self.num_elements = Value(c_int64, 0, lock=False)
+        
+        self.process_local_num_elements = 0
         
         self.num_collisions = 0
             
             
+    @cython.boundscheck(False)
+    @cython.wraparound(False)
+    @cython.cdivision(True)
+    cdef void refresh(self):
+        """
+        If someone has updated the underlying shared memory, need to refresh the size of our 
+        mappings.  Don't need to call the expensive resize() method, just refresh.
+        """
+        
+    
+        self.lookup_buffer.resize(self.lookup_buffer.size()) #resize to full file length
+        
+        self.lookup_memory = np.frombuffer(self.lookup_buffer, dtype=self.numpy_dtype)
+        
+        self.mem_length = self.lookup_buffer.size() // 23
+        
+        self.process_local_num_elements = self.num_elements.value
+        
+        # Whoever updated the memory needs to have also updated the self.num_elements value
+        # so we do not update it here
+        
+        
+        
     
     def __len__(self):
-        return self.num_elements
+        return self.num_elements.value
             
     @cython.boundscheck(False)
     @cython.wraparound(False)
@@ -1366,7 +1395,7 @@ cdef class GalaxyMapCustomDict:
     cpdef OffsetNumPair getitem(self,
                                 CELL_ID_t i, 
                                 CELL_ID_t j, 
-                                CELL_ID_t k):
+                                CELL_ID_t k) except *:
         """
         TODO: update names to p-q-r
         """
@@ -1458,7 +1487,9 @@ cdef class GalaxyMapCustomDict:
                 
                 self.lookup_memory[curr_hash_addr] = out_element
                 
-                self.num_elements += 1
+                self.num_elements.value += 1
+                
+                self.process_local_num_elements += 1
                 
                 break
             
@@ -1481,7 +1512,7 @@ cdef class GalaxyMapCustomDict:
                 
             first_try = False
     
-        if self.num_elements >= (<DTYPE_INT64_t>(0.90*self.mem_length)):
+        if self.num_elements.value >= (<DTYPE_INT64_t>(0.90*self.mem_length)):
             
             self.resize(2*self.mem_length)
             
@@ -1532,9 +1563,10 @@ cdef class GalaxyMapCustomDict:
         # pointed to by self.lookup_fd.  Then point our self.lookup_memory memoryview
         # object to the extended version of where it was already pointing
         ################################################################################
-        self.lookup_buffer.close()
+        #self.lookup_buffer.close()
         
[…]
