### 2022-03-15 510fabe
Updated unit test for VoidFinder
