### 2022-03-15 641d36d
Added return values to find_voids
