+++ python/vast/voidfinder/voidfinder.py
@@ -898,4 +898,9 @@ def find_voids(galaxy_coords_xyz,
     #---------------------------------------------------------------------------
     ############################################################################
     
+    
+    return maximal_spheres_table, myvoids_table
+
+
+
 
