### 2022-03-16 bbee73e
Updated section in docs about OS compatibility and fixed VoidRender mouse movement issue
