### 2022-03-17 701bed7
Added option for wall_distance=None to disable line connecting in VoidRender, full periodic mode support, implemented enhancement #69 separated galaxy_map into two separate shared memory hash tables for faster access to static source survey portion
