### 2022-03-17 f861f4d
Updated example scripts
