+++ example_scripts/SDSS_VoidFinder_dr7_xyz.py
@@ -45,8 +45,7 @@ find_voids(wall_coords_xyz,
            mask_type='xyz',
            mask=None, 
            mask_resolution=None,
-           min_dist=None,
-           max_dist=None,
+           dist_limits=None,
            xyz_limits=xyz_limits,
            #save_after=50000,
            #use_start_checkpoint=True,
