+++ python/vast/test/test_VF.py
@@ -164,8 +164,7 @@ class TestVoidFinder(unittest.TestCase):
                    'test_', 
                    mask=TestVoidFinder.mask, 
                    mask_resolution=1,
-                   min_dist=TestVoidFinder.dist_limits[0],
-                   max_dist=TestVoidFinder.dist_limits[1],
+                   dist_limits=TestVoidFinder.dist_limits,
                    hole_grid_edge_length=1.0,
                    hole_center_iter_dist=0.2, 
                    min_maximal_radius=1.0, 
