### 2022-03-20 DESI-UR/VAST#71 @lavaux
Make example_scripts default path consistent
I have been running the example script, however it seems that it is either:
* lacking in the documentation to modify classifyEnvironment.py to point to the location of the output files generated by SDSS_VoidFinder_dr7.py
* or use the current path as default for the files and not ".."
I would suggest to just remove the ".." from the paths in classifyEnvironment.py
