### 2022-03-20 DESI-UR/VAST#72 @lavaux
Visualization not working on Linux/Wayland
Running `example_scripts/visualize_voids_VoidFinder.py` on linux/wayland triggers an error in PyOpenGL and PyOpenGLTK.

I have not reported that issue to PyOpenGLTK yet. However I would suggest to include a word of warning in the documentation/readme about compatibility issues.

Here is the stacktrace:
```
Traceback (most recent call last):
  File "/home/guilhem/Software/VAST/example_scripts/visualize_voids_VoidFinder.py", line 76, in <module>
    viz = VoidRender(holes_xyz=holes_xyz,
  File "/home/guilhem/venv/vast/lib64/python3.9/site-packages/vast-1.1.1-py3.9-linux-x86_64.egg/vast/voidfinder/viz/void_render.py", line 551, in __init__
    app.Canvas.__init__(self,
  File "/home/guilhem/venv/vast/lib64/python3.9/site-packages/vispy/app/canvas.py", line 173, in __init__
    self._app = use_app(call_reuse=False)
  File "/home/guilhem/venv/vast/lib64/python3.9/site-packages/vispy/app/_default_app.py", line 47, in use_app
    default_app = Application(backend_name)
  File "/home/guilhem/venv/vast/lib64/python3.9/site-packages/vispy/app/application.py", line 47, in __init__
    self._use(backend_name)
  File "/home/guilhem/venv/vast/lib64/python3.9/site-packages/vispy/app/application.py", line 233, in _use
    __import__(mod_name, globals(), level=1)
  File "/home/guilhem/venv/vast/lib64/python3.9/site-packages/vispy/app/backends/_tk.py", line 29, in <module>
    import pyopengltk
  File "/home/guilhem/venv/vast/lib64/python3.9/site-packages/pyopengltk/__init__.py", line 28, in <module>
    from pyopengltk.linux import OpenGLFrame
  File "/home/guilhem/venv/vast/lib64/python3.9/site-packages/pyopengltk/linux.py", line 8, in <module>
    from OpenGL import GL, GLX
  File "/home/guilhem/venv/vast/lib64/python3.9/site-packages/OpenGL/GLX/__init__.py", line 3, in <module>
    from OpenGL.GLX.VERSION.GLX_1_0 import *
  File "/home/guilhem/venv/vast/lib64/python3.9/site-packages/OpenGL/GLX/VERSION/GLX_1_0.py", line 14, in <module>
    from OpenGL.raw.GLX.VERSION.GLX_1_0 import *
  File "/home/guilhem/venv/vast/lib64/python3.9/site-packages/OpenGL/raw/GLX/VERSION/GLX_1_0.py", line 44, in <module>
    def glXChooseVisual(dpy,screen,attribList):pass
  File "/home/guilhem/venv/vast/lib64/python3.9/site-packages/OpenGL/raw/GLX/VERSION/GLX_1_0.py", line 13, in _f
    return _p.createFunction( function,_p.PLATFORM.GLX,'GLX_VERSION_GLX_1_0',error_checker=_errors._error_checker)
AttributeError: 'EGLPlatform' object has no attribute 'GLX'
```
