### 2022-03-20 DESI-UR/VAST#73 @lavaux
Documentation of community guidelines
I do not think that I have found the community guidelines in the documentation to VAST. Could that be added?
