### 2022-03-20 #4033 @lavaux
@kadglass I submitted a few additional minor issues for the documentation as issue tickets.

https://github.com/DESI-UR/VAST/issues/73
https://github.com/DESI-UR/VAST/issues/72
https://github.com/DESI-UR/VAST/issues/71
