### 2022-03-21 DESI-UR/VAST#71 @QuiteAFoxtrot
Thanks @lavaux I also encountered this and fixed it in https://github.com/DESI-UR/VAST/pull/70
