### 2022-03-21 DESI-UR/VAST#72 @QuiteAFoxtrot
Have you encountered this issue when not using the Wayland backend?  I could be a bit out of date but isn't Wayland still considered a somewhat experimental choice of graphical server?  I don't see any harm in pointing this out in the docs though.
