### 2022-03-28 DESI-UR/VAST#73 @sybenzvi
@kadglass, here are the JOSS review criteria discussing community guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines
