### 2022-03-28 DESI-UR/VAST#73 @sybenzvi
Added community guidelines indicating how to contribute to VAST and where to seek support. See current build of online documentation: https://vast.readthedocs.io/en/latest/contributing.html.
