### 2022-04-06 4f36d43
Added periodic example
