### 2022-04-06 bf60c04
Added periodic example file
