### 2022-04-08 3d03785
Maximum redshift reset to 0.1026, resulting in 887 voids
