### 2022-04-08 a22cc00
Updated for periodic boundary conditions
