+++ doc/source/VoidFinder_intro.rst
@@ -12,8 +12,9 @@ defines voids as the unions of sets of spheres grown in the underdense regions
 of the large-scale structure.
 
 The **VoidFinder** directory contains the package, which includes an efficient 
-Multi-Process Cythonized version of VoidFinder.  To import the main 
-void-finding function of VoidFinder::
+Multi-Process Cythonized version of VoidFinder.  Options are available to 
+identify voids in both observational galaxy surveys and periodic cosmological 
+simulations.  To import the main void-finding function of VoidFinder::
     
     from vast.voidfinder import find_voids
 
@@ -42,20 +43,22 @@ Installation
 ============
 
 Operating System Support & Requirements
------------------------------
+---------------------------------------
 
-Linux - **VoidFinder** is currently only fully supported on Linux, but may run on OSX and other Unix variants,
-and potentially even Windows in single-process mode.
+*Linux* --- **VoidFinder** is currently only fully supported on Linux, but may run 
+on OSX and other Unix variants.
 
-OSX - The authors have successfully run **VoidFinder** with full multi-processing power on Mac/OSX,
-but until Apple guarantees a POSIX- or Single Unix Specification-compliant ``fork()`` system call, 
-full multi-processing **VoidFinder** cannot be guaranteed.  However, the single-process version of 
-**VoidFinder** should always work on OSX to the best of our knowledge, and (as of March 2022), **VoidFinder**
-seems to be working correctly on OSX.
+*OSX* --- The authors have successfully run **VoidFinder** with full 
+multi-processing power on Mac/OSX, but until Apple guarantees a POSIX- or Single 
+Unix Specification-compliant ``fork()`` system call, full multi-processing 
+**VoidFinder** cannot be guaranteed.  However, the single-process version of 
+**VoidFinder** should always work on OSX to the best of our knowledge, and (as 
+of March 2022), **VoidFinder** seems to be working correctly on OSX.
 
-Windows - The authors have encountered difficulty compiling the Cython code on the Windows platform,
-and more importantly the ``fork()`` system call is not supported by Windows, so given its small popularity 
-as a scientific computing platform we have no plans (as of March 2022) to support Windows.  
+*Windows* --- The authors have encountered difficulty compiling the Cython code 
+on the Windows platform, and more importantly the ``fork()`` system call is not 
+supported by Windows.  Given its small popularity as a scientific computing 
+platform, we have no plans (as of March 2022) to support Windows.  
 
 
 
@@ -79,6 +82,7 @@ environment that you are using.  To check on this, you can type::
 into a normal unix shell and it will give you a path like ``/usr/bin/python`` or 
 ``/opt/anaconda3/bin/python``, which lets you know which python binary your 
 ``python`` command actually points to.
+
 
 Developing VoidFinder
 ^^^^^^^^^^^^^^^^^^^^^
@@ -109,6 +113,7 @@ Occasionally, it can be helpful to know the following command::
 which can be run from within the directory where the .pyx files live 
 (currently ``VAST/python/vast/voidfinder/``) to "manually" build the cython 
 (.pyx) files.
+
 
 Installing VoidFinder without admin privileges
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
