### 2022-04-08 21ae692
Periodic (#70)
* Updated cython to language_level=3 directive

* Updated cython to language_level=3 directive

* Consolidating parameters.  Updated cython language_level to 3.  Added resize() methods to custom dict class to help enable shared memory resizing for periodic boundary conditions implementation

* Updated public interface simplifying/removing grid shape and coord_min parameters.  Added logic to check mask and grid search edge length for alignment in periodic mode

* Correcting minor errors in interfaces after updates

* Updated all references to w_coord to refer to GalaxyMap attribu
[…]
