+++ python/vast/voidfinder/_voidfinder_cython.pyx
@@ -1,5 +1,4 @@
-
-
+#cython: language_level=3
 
 
 
@@ -14,7 +13,7 @@ cimport numpy as np
 np.import_array()  # required in order to use C-API
 
 
-from typedefs cimport DTYPE_CP128_t, \
+from .typedefs cimport DTYPE_CP128_t, \
                       DTYPE_CP64_t, \
                       DTYPE_F64_t, \
                       DTYPE_F32_t, \
@@ -28,16 +27,17 @@ from numpy.math cimport NAN, INFINITY
 
 from libc.math cimport fabs, sqrt, asin, atan#, exp, pow, cos, sin, asin
 
-from _voidfinder_cython_find_next cimport find_next_galaxy, \
-                                          not_in_mask, \
-                                          _query_first, \
-                                          DistIdxPair, \
-                                          Cell_ID_Memory, \
-                                          GalaxyMapCustomDict, \
-                                          HoleGridCustomDict, \
-                                          FindNextReturnVal, \
-                                          NeighborMemory, \
-                                          MaskChecker
+from ._voidfinder_cython_find_next cimport find_next_galaxy, \
+                                           not_in_mask, \
+                                           _query_first, \
+                                           DistIdxPair, \
+                                           Cell_ID_Memory, \
+                                           GalaxyMapCustomDict, \
+                                           HoleGridCustomDict, \
+                                           FindNextReturnVal, \
+                                           NeighborMemory, \
+                                           MaskChecker, \
+                                           GalaxyMap
                                           
 
 import time
@@ -45,13 +45,12 @@ import time
 
 #Create a typedef for the type of pointer which can point
 #to the mask check function
-ctypedef DTYPE_B_t (*MASK_CHECK_FUNCTION_t)(DTYPE_F64_t[:,:], DTYPE_B_t[:,:], DTYPE_INT32_t, DTYPE_F64_t, DTYPE_F64_t)
+#ctypedef DTYPE_B_t (*MASK_CHECK_FUNCTION_t)(DTYPE_F64_t[:,:], DTYPE_B_t[:,:], DTYPE_INT32_t, DTYPE_F64_t, DTYPE_F64_t)
 
 
 
 
-
-
+'''
 cdef DTYPE_B_t test_function_woo(DTYPE_F64_t[:,:] a, 
                                  DTYPE_B_t[:,:] b, 
                                  DTYPE_INT32_t c,  
@@ -59,11 +58,7 @@ cdef DTYPE_B_t test_function_woo(DTYPE_F64_t[:,:] a,
                                  DTYPE_F64_t e):
 
     return <DTYPE_B_t>0
-
-
-
-
-
+'''
 
 
 @cython.boundscheck(False)
@@ -217,7 +212,6 @@ cpdef DTYPE_INT64_t fill_ijk_zig_zag(DTYPE_INT64_t[:,:] i_j_k_array,
 
 
 
-
 @cython.boundscheck(False)
 @cython.wraparound(False)
 @cython.cdivision(True)
@@ -301,24 +295,16 @@ cpdef DTYPE_INT64_t fill_ijk(DTYPE_INT64_t[:,:] i_j_k_array,
 
 
 
-
-
 @cython.boundscheck(False)
 @cython.wraparound(False)
 @cython.cdivision(True)
-@cython.profile(True)
+@cython.profile(False)
 cpdef void main_algorithm(DTYPE_INT64_t[:,:] i_j_k_array,
-                          galaxy_tree,
-                          DTYPE_F64_t[:,:] w_coord,
+                          GalaxyMap galaxy_map,
                           DTYPE_F64_t dl, 
                           DTYPE_F64_t dr,
                           DTYPE_F64_t[:,:] coord_min, 
[…]
