+++ python/vast/voidfinder/volume/void_volume.pyx
@@ -1,3 +1,4 @@
+#cython: language_level=3
 
 
 cimport cython
