### 2022-04-09 DESI-UR/VAST#72 @lavaux
Sorry I did not reply to this. But indeed no, this problem does not occur on X backend and I did not try on Windows or OSX. Wayland is not experimental anymore for ~2 years minimum. Actually this is the recommended backend for past version of Ubuntu and it will be the default for Ubuntu 22.04 including for NVIDIA.

Some references:
* https://discourse.ubuntu.com/t/trying-wayland-by-default-again/20575
* https://www.omgubuntu.co.uk/2021/01/ubuntu-21-04-will-use-wayland-by-default
* https://wiki.ubuntu.com/Wayland
* https://ubuntu.com/blog/ubuntu-17-10-releases-with-gnome-kubernetes-1-8-minimal-base-images:
> The new default desktop features the latest version of GNOME with extensions developed in collaboration with the GNOME Shell team to provide a familiar experience to long-standing Ubuntu users. 17.10 will run Wayland as the default display server on compatible hardware, with the option of Xorg where required.
