### 2022-04-09 DESI-UR/VAST#72 @QuiteAFoxtrot
Ok thanks @lavaux I think we'll need to take a bit of a deeper look at this
