### 2022-04-11 4b1fee3
Fixed table and docstring for docs
