+++ python/vast/voidfinder/voidfinder.py
@@ -600,33 +600,35 @@ def find_voids(galaxy_coords_xyz,
     ==========
     
     galaxy_coords_xyz : numpy.ndarray of shape (num_galaxies, 3)
-        coordinates of the galaxies in the survey, units of Mpc/h 
+        Coordinates of the galaxies in the survey, units of Mpc/h 
         (xyz space)
         
     survey_name : str
-        identifier for the survey running, may be prepended or appended to 
+        Identifier for the survey running, may be prepended or appended to 
         output filenames including the checkpoint filename
         
     mask_type : string, one of ['ra_dec_z', 'xyz', 'periodic']
-        Determines the mode of mask checking to use and which mask parameters to use.  
+        Determines the mode of mask checking to use and which mask parameters to 
+        use.  
         
-        'ra_dec_z' means the mask, mask_resolution, and dist_limits parameters
-            must be provided.  The 'mask' represents an angular space in Right Ascension and 
-            Declination, the corresponding mask_resolution integer represents the scale needed
-            to index into the Right Ascension and Declination of the mask, and the dist_limits
-            represent the min and max redshift values (as radial distances in xyz space).
+        - 'ra_dec_z' : the mask, mask_resolution, and dist_limits parameters
+          must be provided.  The 'mask' represents an angular space in Right 
+          Ascension and Declination, the corresponding mask_resolution integer 
+          represents the scale needed to index into the Right Ascension and 
+          Declination of the mask, and the dist_limits represent the min and max 
+          redshift values (as radial distances in xyz space).
         
-        'xyz' means that the xyz_limits parameter must be provided which directly encodes
-            a bounding box for the survey in xyz space
+        - 'xyz' : the xyz_limits parameter must be provided which directly 
+          encodes a bounding box for the survey in xyz space
             
-        'periodic' means that the xyz_limits parameter must be provided, which directly
-            encodes a bounding box representing the periodic boundary of the survey, and
-            the survey will be treated as if its bounding box were tiled to infinity in
-            all directions.  Spheres will still only be grown starting from within the
-            original bounding box.
+        - 'periodic' : the xyz_limits parameter must be provided, which directly
+          encodes a bounding box representing the periodic boundary of the 
+          survey, and the survey will be treated as if its bounding box were 
+          tiled to infinity in all directions.  Spheres will still only be grown 
+          starting from within the original bounding box.
         
     mask : numpy.ndarray of shape (N,M) type bool
-        represents the survey footprint in scaled ra/dec space.  Value of True 
+        Represents the survey footprint in scaled ra/dec space.  Value of True 
         indicates that a location is within the survey
         (ra/dec space)
 
@@ -637,28 +639,28 @@ def find_voids(galaxy_coords_xyz,
         [min_dist, max_dist] in units of Mpc/h (xyz space)
         
     xyz_limits : numpy array of shape (2,3)
-        format [x_min, y_min, z_min]
-               [x_max, y_max, z_max]
-        to be used for checking against the mask when mask_type == 'xyz' or for
+        Format [x_min, y_min, z_min], [x_max, y_max, z_max]
+        
+        To be used for checking against the mask when mask_type == 'xyz' or for
         periodic conditions when mask_type == 'periodic'
         
     hole_grid_edge_length : float
-        size in Mpc/h of the edge of 1 cube in the search grid, or distance 
+        Size in Mpc/h of the edge of 1 cube in the search grid, or distance 
         between 2 grid cells
         (xyz space)
 
     min_maximal_radius : float
-        the minimum radius in units of distance for a hole to be considered
+        The minimum radius in units of distance for a hole to be considered
         for maximal status.  Default value is 10 Mpc/h.
         
     max_hole_mask_overlap : float in range (0, 0.5)
[…]
