### 2022-04-18 3073c3a
Added periodic option for Vsquared
