+++ example_scripts/vsquared.py
@@ -16,6 +16,8 @@ p.add_argument('-v', '--visualize', action='store_true', default=False,
                help='Enable void visualization.')
 p.add_argument('-w', '--save_intermediate', action='store_true', default=False,
                help='Save intermediate files in void calculation.')
+p.add_argument('-p', '--periodic', action='store_true', default=False,
+               help='Use periodic boundary contitions.')
 
 req = p.add_argument_group('required named arguments')
 req.add_argument('-c', '--config', dest='config_file', required=True, default="DR7_config.ini",
@@ -25,7 +27,8 @@ args = p.parse_args()
 
 newZobov = Zobov(args.config_file, args.method, 3,
                  save_intermediate=args.save_intermediate,
-                 visualize=args.visualize)
+                 visualize=args.visualize
+                 periodic=args.periodic)
 
 newZobov.sortVoids(method=args.method)
 
