+++ python/vast/vsquared/classes.py
@@ -7,13 +7,13 @@ import healpy as hp
 from astropy.io import fits
 from scipy.spatial import ConvexHull, Voronoi, Delaunay, KDTree
 
-from vast.vsquared.util import toCoord, flatten
+from vast.vsquared.util import toCoord, getBuff, flatten
 
 class Catalog:
     """Catalog data for void calculation.
     """
 
-    def __init__(self,catfile,nside,zmin,zmax,maglim=None,H0=100,Om_m=0.3,maskfile=None):
+    def __init__(self,catfile,nside,zmin,zmax,maglim=None,H0=100,Om_m=0.3,periodic=False,cmin=None,cmax=None,maskfile=None):
         """Initialize catalog.
 
         Parameters
@@ -34,19 +34,30 @@ class Catalog:
             Matter density.
         maskfile : str or None
             Mask file giving HEALPixels with catalog objects.
+        periodic : bool
+            Use periodic boundary conditions.
+        cmin : ndarray or None
+            Array of coordinate minima.
+        cmax : ndarray or None
+            Array of coordinate maxima.
         """
         print("Extracting data...")
         hdulist = fits.open(catfile)
-        z    = hdulist[1].data['z']
-        ra   = hdulist[1].data['ra']
-        dec  = hdulist[1].data['dec']
-        zcut = np.logical_and(z>zmin,z<zmax)
-        if not zcut.any():
-            print("Choose valid redshift limits")
-            return
-        scut = zcut
-        c1,c2,c3   = toCoord(z,ra,dec,H0,Om_m)
-        self.coord = np.array([c1,c2,c3]).T
+        if periodic:
+            self.coord = np.array(hdulist[1].data['x'],hdulist[1].data['y'],hdulist[1].data['z']).T
+            self.cmin = cmin
+            self.cmax = cmax
+        else:
+            z    = hdulist[1].data['z']
+            ra   = hdulist[1].data['ra']
+            dec  = hdulist[1].data['dec']
+            zcut = np.logical_and(z>zmin,z<zmax)
+            if not zcut.any():
+                print("Choose valid redshift limits")
+                return
+            scut = zcut
+            c1,c2,c3   = toCoord(z,ra,dec,H0,Om_m)
+            self.coord = np.array([c1,c2,c3]).T
         nnls = np.arange(len(z))
         nnls[zcut<1] = -1
         if maglim is not None:
@@ -62,20 +73,21 @@ class Catalog:
             nnls[ncut] = lut[tree.query(self.coord[ncut])[1]]
             self.mcut = mcut
         self.nnls = nnls
-        if maskfile is None:
-            print("Generating mask...")
-            mask = np.zeros(hp.nside2npix(nside),dtype=bool)
-            pids = hp.ang2pix(nside,ra[scut],dec[scut],lonlat=True)
-            mask[pids] = True
-        self.mask = mask
-        pids = hp.ang2pix(nside,ra,dec,lonlat=True)
-        self.imsk = mask[pids]*zcut
+        if not periodic:
+            if maskfile is None:
+                print("Generating mask...")
+                mask = np.zeros(hp.nside2npix(nside),dtype=bool)
+                pids = hp.ang2pix(nside,ra[scut],dec[scut],lonlat=True)
+                mask[pids] = True
+            self.mask = mask
+            pids = hp.ang2pix(nside,ra,dec,lonlat=True)
+            self.imsk = mask[pids]*zcut
 
 class Tesselation:
     """Implementation of Voronoi tesselation of the catalog.
     """
 
-    def __init__(self,cat,viz=False):
+    def __init__(self,cat,viz=False,periodic=False,buff=5.):
         """Initialize tesselation.
 
         Parameters
[…]
