### 2022-04-18 d70d976
Fixed periodic option for Vsquared
