### 2022-04-21 e6b9b70
Updated Vsquared config file
