### 2022-04-21 59d5baf
Fixed Vsquared periodic output
