### 2022-05-02 724e0b5
Updated Vsquared docs
