### 2022-05-08 DESI-UR/VAST#72 @QuiteAFoxtrot
Had a chance to do some testing - indeed this visualization does not work on Wayland, but I am marking this issue Closed because I cannot get anything else to work on Wayland either, even a simple script like:
```import numpy
import matplotlib.pyplot as plt
data = numpy.arange(10)
plt.plot(data, data)
plt.show()```
fails with libGL errors.  I believe this is a Wayland problem, not a VAST issue, but we will update our docs to reflect this known problem.
