### 2022-05-08 DESI-UR/VAST#72 @QuiteAFoxtrot
Sorry - realized it would be more accurate to re-open this and mark it as Docs instead of bugs then we'll close when we link in a PR for this docs update.
