### 2022-05-09 e999c65
Doc updates; consolidated install instructions under main VAST header, added notes for Linux/Wayland compatibility issue
