+++ doc/source/VoidRender.rst
@@ -71,24 +71,4 @@ above.
 
 
 
-Installation
-============
-
-**VoidRender** requires OpenGL :math:`\geq` 1.2 and ``vispy`` (``pip install 
-vispy``).
-
-Also, on Ubuntu, you may need to install the following::
-
-    sudo apt-get install mesa-common-dev libgl1-mesa-dev libgl1-mesa-dev
-
-
-There is a version of **VoidRender** that corresponds to each void-finding 
-algorithm implemented in **VAST**.  As a result, **VoidRender** will be built 
-and installed when either **VoidFinder** or 
-:raw-html:`<strong>V<sup>2</sup></strong>` is built and installed.  Please see 
-:ref:`VF-install` and/or :ref:`V2-install` for installation details.
-
-
-
-
 
