+++ doc/source/index.rst
@@ -14,6 +14,7 @@ of two popular classes of void-finding algorithms in galaxy catalogs:
 
 1. Void identification by growing spherical voids.
 2. Void identification using watershed algorithms.
+  
 
 Voids are expansive regions in the universe containing significantly fewer 
 galaxies than surrounding galaxy clusters and filaments.  They are a fundamental 
@@ -25,6 +26,142 @@ formulating a concrete definition to use in a void-finding algorithm is not
 trivial.  As a result, several different algorithms exist to identify these 
 cosmic underdensities.  VAST provides Python 3 implementations of two such 
 algorithms: **VoidFinder** and :raw-html:`<strong>V<sup>2</sup></strong>`.
+
+**VAST** also contains an implementation called **VoidRender** which utilizes OpenGL to
+create interactive 3D visualizations of the output of the void-finding algorithms
+overlaid on top of the reference galaxy survey.  Each of the void-finding algorithms
+in **VAST** provides its own visualization class.
+
+
+
+
+
+.. _VAST-install:
+
+Installation
+============
+
+Operating System Support & Requirements
+---------------------------------------
+
+Linux
+^^^^^
+
+**VoidFinder** is currently only fully supported on Linux, but may run on 
+OSX and other Unix variants.
+
+:raw-html:`<strong>V<sup>2</sup></strong>` is fully supported on Linux.
+
+**VoidRender** is not expected to work well with the Wayland
+graphical backend.  VoidRender works well on the older and more stable Xorg 
+backend, on Intel, Nvidia, and AMD graphical drivers.  Note that for extreme
+performance settings (very high numbers of voids or a ``SPHERE_TRIANGULARIZATION_DEPTH``
+greater than 3), a discrete gpu from AMD or Nvidia is highly recommended for a smooth
+visualization.
+
+It may be helpful to install the following for **VoidRender** on Ubuntu::
+
+    sudo apt-get install mesa-common-dev libgl1-mesa-dev libgl1-mesa-dev
+
+
+OSX
+^^^
+
+The authors have successfully run **VoidFinder** with full 
+multi-processing power on Mac/OSX, but until Apple guarantees a POSIX- or Single 
+Unix Specification-compliant ``fork()`` system call, full multi-processing 
+**VoidFinder** cannot be guaranteed.  However, the single-process version of 
+**VoidFinder** should always work on OSX to the best of our knowledge, and (as 
+of March 2022), **VoidFinder** seems to be working correctly on OSX.
+
+:raw-html:`<strong>V<sup>2</sup></strong>` runs on OSX.
+
+**VoidRender** is expected to work on OSX using the Intel graphical drivers.
+
+
+Windows
+^^^^^^^
+
+The authors have encountered difficulty compiling the **VAST** Cython code 
+on the Windows platform, and more importantly the ``fork()`` system call is not 
+supported by Windows.  Given its small popularity as a scientific computing 
+platform, we have no plans (as of May 2022) to support Windows.  
+
+
+
+
+Prerequisite Libraries
+----------------------
+
+**VoidFinder** uses Cython to speed up its operations, so when building from source
+you will need a C compiler like gcc.
+
+**VoidRender** requires OpenGL :math:`\geq` 1.2.
+
+
+
+
+Building VAST
+-------------------
+
+**VAST** does not yet have any pre-built wheels or distribution packages, 
+so clone the repository from https://github.com/DESI-UR/VAST.
+
+**VAST** will install like a normal python package via the shell command::
+
+    python setup.py install
+    
+It is important to remember that this will attempt to install 
+`vast` into the ``site-packages`` directory of whatever python 
+environment that you are using.  To check on this, you can type::
+
+    which python
[…]
