### 2022-05-13 c4fa7e0
Fixed minor doc bugs, restructured intro page
