+++ doc/source/Vsquared_functions.rst
@@ -65,7 +65,7 @@ V\ :sup:`2` API
 
 
 ``util.getBuff``
-=============
+================
 
 .. autofunction:: vast.vsquared.util.getBuff
 
