### 2022-05-13 cc01642
Minor docs updates for note on Linux/Wayland (#77)
* Doc updates; consolidated install instructions under main VAST header, added notes for Linux/Wayland compatibility issue

* Fixed minor doc bugs, restructured intro page

Co-authored-by: Stephen W. O'Neill Jr <stephen.w.o'neill@lmco.com>
Co-authored-by: Kelly <kad347@drexel.edu>
