### 2022-06-07 a560d5f
missing comma in example script
