### 2022-06-07 #4033 @dfm
@kadglass — I've opened a PR with some minor edits to the paper. After merging or responding to that, here are the final steps that I'll need from you:

1. Comment `@editorialbot generate pdf`, and take one last read through the manuscript to make sure that you're happy with it (it's harder to make changes later!), especially the author names and affiliations. I've taken a pass and it looks good to me!

Thanks all!
