+++ python/vast/vsquared/data/test_config.ini
@@ -21,7 +21,7 @@ radius_min    = 10.0
 nside         = 16
 #Step size for comoving-distance-to-redshift lookup table
 redshift_step = 5e-5
-#Box limits for periodic boundary conditions
+#Coordinate limits (Mpc/h) for periodic and xyz modes
 x_min         = -100.
 y_min         = -100.
 z_min         = -100.
