+++ example_scripts/DR7_config.ini
@@ -21,7 +21,7 @@ radius_min    = 10.0
 nside         = 32
 #Step size for comoving-distance-to-redshift lookup table
 redshift_step = 5e-5
-#Periodic limits (Mpc/h)
+#Coordinate limits (Mpc/h) for periodic and xyz modes
 x_min         = -50.
 y_min         = -50.
 z_min         = -50.
