### 2022-07-18 a031549
Added missing comma to sample V2 script, code formatting adjustments
