### 2022-07-20 7617352
Fixed fake void galaxy removal
