### 2022-07-20 50c0757
Added 'edge' column to truth table, other minor changes
