### 2022-07-21 6aedd4e
Added min_maximal_radius to generate_mask so that it is not hard-coded
