### 2022-07-22 5344946
Added pts_per_unit_volume kwarg to find_voids
