+++ python/vast/voidfinder/volume_cut.py
@@ -235,13 +235,10 @@ def check_hole_bounds(x_y_z_r_array,
                       num_cpus=1,
                       verbose=0):
     """
-    Description
-    ===========
-    
     Remove holes from the output of _hole_finder() whose volume falls outside of 
     the mask by X % or more.  
     
-    This is accomplished by a 2-phase approach, first, N points are distributed 
+    This is accomplished by a 2-phase approach.  First, N points are distributed 
     on the surface of each sphere, and those N points are checked against the 
     mask.  If any of those N points fall outside the mask, the percentage of the
     volume of the sphere which falls outside the mask is calculated by using a
@@ -250,6 +247,7 @@ def check_hole_bounds(x_y_z_r_array,
     The percentage of volume outside the mask is then approximated as the 
     percentage of those points which fall outside the mask.
     
+
     Parameters
     ==========
     
@@ -323,6 +321,113 @@ def check_hole_bounds(x_y_z_r_array,
 
 
 
+def build_unit_sphere_points(num_surf_pts):
+    '''
+    Distribute N points on a unit sphere.
+
+    Reference algorithm "Golden Spiral" method: 
+    https://stackoverflow.com/questions/9600801/evenly-distributing-n-points-on-a-sphere
+
+
+    PARAMETERS
+    ==========
+
+    num_surf_pts : int
+        Number of points to distribute on the surface of the sphere
+
+
+    RETURNS
+    =======
+
+    unit_sphere_pts : ndarray of shape (num_surf_pts, 3)
+        Cartesian coordinates of the surface points
+    '''
+
+    indices = np.arange(0, num_surf_pts, dtype=float) + 0.5
+    
+    phi = np.arccos(1 - 2*indices/num_surf_pts)
+    
+    theta = np.pi * (1 + 5**0.5) * indices
+    
+    x = np.cos(theta) * np.sin(phi)
+    y = np.sin(theta) * np.sin(phi)
+    z = np.cos(phi)
+    
+    unit_sphere_pts = np.empty((num_surf_pts, 3), dtype=np.float64)
+    unit_sphere_pts[:,0] = x
+    unit_sphere_pts[:,1] = y
+    unit_sphere_pts[:,2] = z
+
+    return unit_sphere_pts
+
+
+
+
+
+def generate_mesh(radius, pts_per_unit_volume):
+    '''
+    Generate a mesh of constant density such that a sphere will fit in this 
+    mesh.
+
+    Cut the extraneous points, and sort all of the points in order of smallest 
+    radius to largest radius, so that when we iterate later for the smaller 
+    holes we can stop early at the largest necessary radius - the cythonized 
+    code critically depends on this sort.
+
+
+    PARAMETERS
+    ==========
+
+    radius : float
+        Radius of sphere around which to define the mesh grid.
+
+    pts_per_unit_volume : int
+        Number of points per unit volume in the mesh grid.
+
+
+    RETURNS
+    =======
+
+    mesh_points : ndarray of shape (N,3)
+        Cartesian coordinates of all points in mesh
+
+    mesh_points_radii : ndarray of shape (N,)
+        Distance from center to each point in mesh_points
+    '''
+
+    gen_radius = radius*1.05 # add a bit of margin for the mesh
+    
+    step = 1.0/np.power(pts_per_unit_volume, 0.33)
+    
+    mesh_pts = np.arange(-1.0*gen_radius, gen_radius, step)
+    
+    n_pts = mesh_pts.shape[0]
+    
+    mesh_x, mesh_y, mesh_z = np.meshgrid(mesh_pts, mesh_pts, mesh_pts)
+    
[…]
