### 2022-07-22 8b429dc
Fixed indexing issue for identifying edge voids
