### 2022-07-25 eccd226
Changed deprecated np.bool to np.bool_
