### 2022-07-27 83423bc
Added field galaxies to find_voids so that the hole grid is defined based on both wall and field galaxies
