+++ python/vast/voidfinder/voidfinder.py
@@ -323,8 +323,12 @@ def ra_dec_to_xyz(galaxy_table,
 
 
 
-def calculate_grid(galaxy_coords_xyz,
-                   hole_grid_edge_length):
+def calculate_grid(galaxy_coords,
+                   mask_type='ra_dec_z',
+                   xyz_limits=None,
+                   hole_grid_edge_length=5.0, 
+                   galaxy_map_grid_edge_length=None, 
+                   verbose=0):
     """
     Given a galaxy survey in xyz/Cartesian coordinates and the length of a 
     cubical grid cell, calculate the cubical grid shape which will contain the 
@@ -344,11 +348,52 @@ def calculate_grid(galaxy_coords_xyz,
     Parameters
     ==========
     
-    galaxy_coords_xyz : numpy.ndarray of shape (N,3)
-        coordinates of survey galaxies in xyz space
+    galaxy_coords : astropy Table or numpy.ndarray of shape (N,3)
+        coordinates of survey galaxies in either sky coordinates or xyz space
+
+    mask_type : string, one of ['ra_dec_z', 'xyz', 'periodic']
+        Determines the mode of mask checking to use and which mask parameters to 
+        use.  
+        
+        'ra_dec_z' means the mask, mask_resolution, and dist_limits parameters
+            must be provided.  The 'mask' represents an angular space in Right 
+            Ascension and Declination, the corresponding mask_resolution integer 
+            represents the scale needed to index into the Right Ascension and 
+            Declination of the mask, and the dist_limits represent the min and 
+            max redshift values (as radial distances in xyz space).
+        
+        'xyz' means that the xyz_limits parameter must be provided which 
+            directly encodes a bounding box for the survey in xyz space
+            
+        'periodic' means that the xyz_limits parameter must be provided, which 
+            directly encodes a bounding box representing the periodic boundary 
+            of the survey, and the survey will be treated as if its bounding box 
+            were tiled to infinity in all directions.  Spheres will still only 
+            be grown starting from within the original bounding box.
+
+    xyz_limits : numpy array of shape (2,3)
+        format [x_min, y_min, z_min]
+               [x_max, y_max, z_max]
+        to be used for checking against the mask when mask_type == 'xyz' or for
+        periodic conditions when mask_type == 'periodic'
         
     hole_grid_edge_length : float
-        length in xyz space of the edge of 1 cubical cell in the grid
+        Length in xyz space of the edge of 1 cubical cell in the grid.  Default 
+        value is 5 Mpc/h.
+
+    galaxy_map_grid_edge_length : float or None
+        Edge length in Mpc/h for the secondary grid for finding nearest neighbor 
+        galaxies.  If None, will default to 3*hole_grid_edge_length (which 
+        results in a cell volume of 3^3 = 27 times larger cube volume).  This 
+        parameter yields a tradeoff between number of galaxies in a cell, and 
+        number of cells to search when growing a sphere.  Too large and many 
+        redundant galaxies may be searched, too small and too many cells will 
+        need to be searched.
+
+    verbose : int
+        Level of verbosity to print during running, 0 indicates off, 1 indicates 
+        to print after every 'print_after' cells have been processed, and 2 
+        indicates to print all debugging statements.  Default is 0.
         
         
     Returns
@@ -356,15 +401,186 @@ def calculate_grid(galaxy_coords_xyz,
     
     hole_grid_shape : tuple of ints (i,j,k)
         number of grid cells in each dimension
+
+    galaxy_map_grid_shape : tuple of ints (i,j,k)
+        number of galaxy map grid cells in each dimension
         
     coords_min : numpy.ndarray of shape (3,)
         the (min_x, min_y, min_z) point which is the (0,0,0) of the grid
-
[…]
