+++ python/vast/voidfinder/_voidfinder.py
@@ -243,16 +243,13 @@ def _hole_finder(galaxy_coords,
                  verbose=0,
                  print_after=10000,
                  num_cpus=None,
-               
+                 
                  CONFIG_PATH="/tmp/voidfinder_config.pickle",
                  SOCKET_PATH="/tmp/voidfinder.sock",
                  RESOURCE_DIR="/dev/shm",
                  DEBUG_DIR="/home/moose/VoidFinder/doc/debug_dir"
                  ):
     """
-    Description
-    ===========
-    
     See help(voidfinder.find_voids)
     
     Work-horse method for running VoidFinder with the Cython code in parallel
@@ -260,15 +257,15 @@ def _hole_finder(galaxy_coords,
     
     This method contains the logic for:
     
-    1). Sanity check the num_cpus to use
-    2). Open file handles and allocate memory for workers to memmap to
-    3). Build a few data structures for the workers to share
-    4). Register some cleanup helpers with the python interpreters for 
-        making sure the disk space gets reclaimed when we are done
-    5). Start the workers
-    6). Make sure workers connect to the comm socket
-    7). Checkpoint the progress if those parameters are enabled
-    8). Collect & print progress results from the workers
+    1) Sanity check the num_cpus to use
+    2) Open file handles and allocate memory for workers to memmap to
+    3) Build a few data structures for the workers to share
+    4) Register some cleanup helpers with the python interpreters for making 
+       sure the disk space gets reclaimed when we are done
+    5) Start the workers
+    6) Make sure workers connect to the comm socket
+    7) Checkpoint the progress if those parameters are enabled
+    8) Collect & print progress results from the workers
     
     This function is designed to be run on Linux on an SMP (Symmetric 
     Multi-Processing) architecture.  It takes advantage of 2 Linux-specific 
@@ -289,9 +286,10 @@ def _hole_finder(galaxy_coords,
     Parameters
     ==========
     
-    galaxy_coords : numpy.ndarray of shape (num_galaxies, 3)
-        coordinates of the galaxies in the survey, units of Mpc/h
-        (xyz space)
+    galaxy_coords : length-2 list of numpy.ndarrays of shape (num_galaxies, 3)
+        Cartesian coordinates of the galaxies in the survey, units of Mpc/h.  
+        The first element in the list is the wall galaxies, and the second is 
+        the field galaxies.
     
     hole_grid_edge_length : scalar float
         length of each cell in Mpc/h
@@ -391,7 +389,7 @@ def _hole_finder(galaxy_coords,
             max_dist is None:
             raise ValueError("Mask mode is 0 (ra-dec-z) but a required mask parameter is None")
         
-        #The cython requires some very specific types and shapes
+        # The cython requires some very specific types and shapes
         mask = mask.astype(np.uint8)
     
     if mask_mode == 1 and xyz_limits is None:
@@ -404,26 +402,24 @@ def _hole_finder(galaxy_coords,
     
     
     
-        
     ############################################################################
-    # Next, depending on the mask mode, calculate the transform origin and
-    # grid cell parameters for our Hole grid and our Galaxy Map grids.
+    # Next, depending on the mask mode, calculate the transform origin and grid
+    # cell parameters for our Hole grid and our Galaxy Map grids.
     #
-    # For 'ra-dec-z' - just use the min and max of the provided
-    #    coordinates of the survey for the transform 
+    # For 'ra-dec-z' - just use the min and max of the provided coordinates of 
+    #     the survey for the transform
     #
[…]
