### 2022-07-27 ca206fa
Cleaning up doc strings, increasing version number to reflect change in find_voids function call
