+++ python/vast/voidfinder/_voidfinder_cython_find_next.pyx
@@ -73,24 +73,24 @@ cdef FindNextReturnVal find_next_galaxy(DTYPE_F64_t[:,:] hole_center_memview,
     Description:
     ============
     
-    
     Function to locate the next nearest galaxy during hole center propagation 
     along direction defined by unit_vector_memview.
 
-    The algorithm needs to find 4 bounding galaxies per cell.  The 
-    first galaxy is found as the minimum of regular euclidean distance to the hole center.  The
-    second galaxy is determined using a ratio of distance with respect to projected distance
-    along the hole center search path.  To find galaxy 2, the hole center search path and 
-    distance ratios are calculated with respect to neighbor 1.  
+    The algorithm needs to find 4 bounding galaxies per cell.  The first galaxy 
+    is found as the minimum of regular euclidean distance to the hole center.  
+    The second galaxy is determined using a ratio of distance with respect to 
+    projected distance along the hole center search path.  To find galaxy 2, the 
+    hole center search path and distance ratios are calculated with respect to 
+    neighbor 1.  
 
     Implementation Detail:
-    For galaxies 3 and 4, the 
-    distance ratios are calculated with respect to the previous galaxies, so this code uses
-    an if-block to detect whether we're calculating for galaxy 2/B, 3/C, or 4/D, and runs slightly
-    different distance ratio calculations.  The if block looks for the number of elements
-    in the input nearest_gal_index_list, when its 1 it assumes we're finding for galaxy 2
-    and when its not 1 it assumes we have the first 2 galaxies and we're looking for galaxy
-    3 or 4.
+    For galaxies 3 and 4, the distance ratios are calculated with respect to the 
+    previous galaxies, so this code uses an if-block to detect whether we're 
+    calculating for galaxy 2/B, 3/C, or 4/D, and runs slightly different 
+    distance ratio calculations.  The if block looks for the number of elements 
+    in the input nearest_gal_index_list, when its 1 it assumes we're finding for 
+    galaxy 2 and when its not 1 it assumes we have the first 2 galaxies and 
+    we're looking for galaxy 3 or 4.
 
 
 
@@ -109,8 +109,8 @@ cdef FindNextReturnVal find_next_galaxy(DTYPE_F64_t[:,:] hole_center_memview,
         Incrememt value for hole propagation
         
     direction_mod : -1.0 or 1.0
-        basically a switch to go in the opposite direction of vector propagation for
-        finding galaxies 4a and 4b
+        basically a switch to go in the opposite direction of vector propagation 
+        for finding galaxies 4a and 4b
 
     unit_vector_memview : memview of shape (3)
         Unit vector indicating direction hole center will shift
@@ -148,7 +148,6 @@ cdef FindNextReturnVal find_next_galaxy(DTYPE_F64_t[:,:] hole_center_memview,
     Returns:
     ========
 
-
     retval : FindNextReturnVal
     
         struct with members:
@@ -160,33 +159,33 @@ cdef FindNextReturnVal find_next_galaxy(DTYPE_F64_t[:,:] hole_center_memview,
             ???
 
         in_mask : boolean
-            Flag indicating whether or not the temporary hole center is within the 
-            survey footprint.
+            Flag indicating whether or not the temporary hole center is within 
+            the survey footprint.
     '''
 
 
 
-    # We are going to shift the center of the hole by dr along 
-    # the direction of the vector pointing from the nearest 
-    # galaxy to the center of the empty cell.  From there, we 
-    # will search within a radius of length the distance between 
-    # the center of the hole and the first galaxy from the 
-    # center of the hole to find the next nearest neighbors.  
-    # From there, we will minimize top/bottom to find which one 
-    # is the next nearest galaxy that bounds the hole.
[…]
