+++ python/vast/vsquared/zobov.py
@@ -350,8 +350,8 @@ class Zobov:
         zverts = self.zones.zverts
         znorms = self.zones.znorms
         z2v = self.zvoid.T[1]
-        z2v2 = np.array([np.where(z2v==z2)[0] for z2 in np.unique(z2v[z2v!=-1])])
         z2v3 = np.unique(z2v[z2v!=-1])
+        z2v2 = np.array([np.where(z2v==z2)[0] for z2 in z2v3])
         zcut = [np.product([np.product(self.tesselation.vecut[self.zones.zcell[z]])>0 for z in z2])>0 for z2 in z2v2]
 
         tri1 = []
@@ -373,13 +373,13 @@ class Zobov:
                         tri2.append(verc[t[1]])
                         tri3.append(verc[t[2]])
                         norm.append(n)
-                        vid.append(z2v3[k])
-                    g2v[gids[p[0]]] = z2v3[k]
+                        vid.append(z2v3[zcut][k])
+                    g2v[gids[p[0]]] = z2v3[zcut][k]
         for k,v in enumerate(z2v2[zcut]):
             for z in v:
                 for i in range(len(znorms[z])):
                     if g2v[gids[p[1]]] != -1:
-                        g2v2[gids[p[1]]] = z2v3[k]
+                        g2v2[gids[p[1]]] = z2v3[zcut][k]
 
         if len(vid)==0:
             print("Error: largest void found encompasses entire survey (try using a method other than 1 or 2)")
