### 2022-08-05 6518a90
Remove old Vsquared directory
