### 2022-08-05 ee7754e
feat: HEALPix mask for VoidFinder
This introduces a healpix premask to VoidFinder to account for spherical geometry and resolve patchiness near the poles.
