+++ python/vast/voidfinder/multizmask.py
@@ -3,6 +3,7 @@ from astropy.table import Table
 
 import numpy as np
 import time
+import healpy as hp
 
 from vast.voidfinder.distance import z_to_comoving_dist
 from vast.voidfinder.constants import c #speed of light
@@ -139,39 +140,74 @@ def generate_mask(gal_data,
     # will fall into an integer bucket, so we use .astype(int), and that value
     # will represent that the mask[ra_scaled, dec_scaled] should be set to True.
     #
-    # We take the unique rows of the scaled coordinates, since as we converted
-    # to integers, many ra-dec pairs will fall into the same integer bucket and 
-    # we only need to set that integer bucket to "True" once.
     #---------------------------------------------------------------------------
     scaled_converted_ang = (mask_resolution*ang).astype(int)
     
-    pre_mask = np.unique(scaled_converted_ang, axis=0)
+    ############################################################################
+
+
+
+    ############################################################################
+    # Now we create a healpix based boolean pre-mask with approximately the same
+    # number of bins as in the mask: (360*N)*(180*N). The actual dimensionality 
+    # of the pre-mask is the closest squared multiple of 12, as required by the
+    # healpix formalism. This multiple, termed as NSIDE in the healpix
+    # algorithm, is used with healpix to select and mark bins in the pre-mask 
+    # that contain galaxies.
+    #
+    # We optionally smooth the premask to reduce patchiness.
+    #  
+    #---------------------------------------------------------------------------
+    
+    num_px = maskra * maskdec * mask_resolution ** 2
+    hpscale=1.
+    nside = int(hpscale*np.sqrt(num_px / 12)) #test scale by 4
+    healpix_mask = np.zeros(hp.nside2npix(nside), dtype = bool)
+    pxids = hp.ang2pix(nside, ra, dec, lonlat = True)
+    pxids = np.unique(pxids, axis=0)
+    healpix_mask[pxids] = 1
+    
+    if smooth_mask:
+        
+        correct_idxs = []
+        
+        for (i, curr_val) in enumerate(healpix_mask):
+            
+            neighbors=hp.get_all_neighbours(nside,i)[::2]
+            
+            if curr_val == 0 and np.sum(healpix_mask[neighbors]) >= 3:
+                correct_idxs.append(i)
+            
+        healpix_mask[correct_idxs] = 1
+    
     ############################################################################
 
 
 
     ############################################################################
     # Now we create the actual boolean mask by allocating an array of shape
-    # (360*N, 180*N), and iterating through all the unique galaxy ra-dec
-    # integer buckets and setting those locations to True to represent they are 
-    # valid locations.
+    # (360*N, 180*N), and iterating through all ra-dec positions on this grid,
+    # and using the corresponding location on pre-mask to determine if the bin
+    # is in the mask.
     #
     # Since declination is actually measured from the equator, we need to 
-    # subtract 90 degrees from the scaled value in order to convert from 
-    # [-90,90) space into [0,180) space.
+    # subtract 90 degrees from the mask index in order to convert from 
+    # [0,180) space into [-90,90) space.
     #---------------------------------------------------------------------------
-    mask = np.zeros((mask_resolution*maskra, mask_resolution*maskdec), 
-                    dtype=bool)
-    '''
-    for j in range(len(pre_mask[0])):
 
-        mask[ maskfile[0,j], maskfile[1,j] - mask_resolution*dec_offset] = True
-    '''
-    for row in pre_mask:
+    mask = np.zeros((mask_resolution*maskra, mask_resolution*maskdec), 
+                    dtype=bool)  
         
-        mask[row[0], row[1] - mask_resolution*dec_offset] = True
-        
-        
+
[…]
