### 2022-08-17 2ac4fe4
Commented out early return for next galaxy locator
