### 2022-08-25 d7478f9
Creating debug file, doc string cleanup
