+++ python/vast/voidfinder/_voidfinder_cython.pyx
@@ -43,6 +43,9 @@ from ._voidfinder_cython_find_next cimport find_next_galaxy, \
 import time
 
 
+#Debugging
+from .viz import VoidRender
+
 #Create a typedef for the type of pointer which can point
 #to the mask check function
 #ctypedef DTYPE_B_t (*MASK_CHECK_FUNCTION_t)(DTYPE_F64_t[:,:], DTYPE_B_t[:,:], DTYPE_INT32_t, DTYPE_F64_t, DTYPE_F64_t)
@@ -311,9 +314,6 @@ cpdef void main_algorithm(DTYPE_INT64_t[:,:] i_j_k_array,
                           int verbose,
                           ):
     '''
-    Description
-    ===========
-    
     Given a potential void cell center denoted by i,j,k, find the 4 bounding 
     galaxies that maximize the interior dimensions of the hole sphere at this 
     location.
@@ -324,13 +324,13 @@ cpdef void main_algorithm(DTYPE_INT64_t[:,:] i_j_k_array,
     The code below will utilize the naming scheme galaxy A, B, C, and D to 
     denote the 1st, 2nd, 3rd, and 4th "neighbor" bounding galaxies found during 
     the running of this algorithm.  I tried to use 1/A, 2/B, 3/C and 4/D to be 
-    clear on the numbers and letters are together.
+    clear on the numbers and letters that are together.
     
     The distance metrics are somewhat special.  Galaxy A is found by normal 
-    minimzation of euclidean distance between the cell center and itself and the 
+    minimzation of Euclidean distance between the cell center and itself and the 
     other neighbors.  Galaxies B, C, and D are found by propogating a hole 
-    center in specific directions and minimizing a ratio of two other 
-    distance-metric-like values.  This needs more detail on how and why.
+    center in specific directions and minimizing the distance the center needs 
+    to move to be bounded by each set of galaxies.
     
     
     Parameters:
@@ -342,9 +342,8 @@ cpdef void main_algorithm(DTYPE_INT64_t[:,:] i_j_k_array,
       
     galaxy_map : _voidfinder_cython_find_next.GalaxyMap
         An interface to searching, accessing, and retrieving the galaxies
-        in the survey efficiently.  handles logic like regular vs. periodic
+        in the survey efficiently.  Handles logic like regular vs. periodic
         access, and memory mapping the arrays for the workers
-        
         
     dl : float
         edge length in Mpc/h of the hole grid cell size
@@ -389,6 +388,7 @@ cpdef void main_algorithm(DTYPE_INT64_t[:,:] i_j_k_array,
     verbose : int
         values >= 1 indicate to make a lot more print statements
     
+
     Returns:
     ========
     
@@ -403,10 +403,8 @@ cpdef void main_algorithm(DTYPE_INT64_t[:,:] i_j_k_array,
     ############################################################################
     # re-used helper index variables
     # re-used helper computation variables
-    ############################################################################
-    
-    
-    #DEBUG_OUTFILE = open("VF_DEBUG.txt", 'a')
+    #---------------------------------------------------------------------------
+    DEBUG_OUTFILE = open("VF_DEBUG.txt", 'a')
     
     
     cdef ITYPE_t working_idx
@@ -423,47 +421,75 @@ cpdef void main_algorithm(DTYPE_INT64_t[:,:] i_j_k_array,
     cdef DTYPE_F64_t temp_f64_accum2
     
     cdef DTYPE_F64_t temp_f64_val
+    ############################################################################
+
+
     
     ############################################################################
     # Hole center vector and propogate hole center memory.  We can re-use the 
     # same memory for finding galaxies 2/B and 3/C but not for galaxy 4/D
[…]
