+++ python/vast/voidfinder/hole_combine.py
@@ -146,6 +146,7 @@ def combine_holes(spheres_table, frac=0.1):
 
 def combine_holes_2(x_y_z_r_array,
                     boundary_holes, 
+                    mask_checker,
                     dup_tol=0.1,
                     maximal_overlap_frac=0.1,
                     min_maximal_radius=10.0,
@@ -167,6 +168,8 @@ def combine_holes_2(x_y_z_r_array,
 
     boundary_holes : ndarray of shape (N,)
         whether or not the hole is on the survey boundary
+
+    mask_checker : 
         
     dup_tol : float 
         the tolerance in units of distance to check 2 hole centers
@@ -292,6 +295,59 @@ def combine_holes_2(x_y_z_r_array,
         if np.any(boundary_voids[void_hole_indices]):
 
             maximals_table["edge"][i] = 1
+
+        else:
+            #-------------------------------------------------------------------
+            # Also mark those voids with at least one hole within 10 Mpc/h of 
+            # the survey boundary
+            #-------------------------------------------------------------------
+            void_holes = holes_table[void_hole_indices]
+
+            for j in range(np.sum(void_hole_indices)):
+
+                # Find the points which are 10 Mpc/h in each direction from the 
+                # center
+                hole_x_min = void_holes['x'][j] - 10.
+                hole_y_min = void_holes['y'][j] - 10.
+                hole_z_min = void_holes['z'][j] - 10.
+                hole_x_max = void_holes['x'][j] + 10.
+                hole_y_max = void_holes['y'][j] + 10.
+                hole_z_max = void_holes['z'][j] + 10.
+
+                # Coordinates to check
+                x_coords = [hole_x_min, 
+                            hole_x_max, 
+                            void_holes['x'][j], 
+                            void_holes['x'][j], 
+                            void_holes['x'][j], 
+                            void_holes['x'][j]]
+
+                y_coords = [void_holes['y'][j], 
+                            void_holes['y'][j], 
+                            hole_y_min, 
+                            hole_y_max, 
+                            void_holes['y'][j], 
+                            void_holes['y'][j]]
+
+                z_coords = [void_holes['z'][j], 
+                            void_holes['z'][j], 
+                            void_holes['z'][j], 
+                            void_holes['z'][j], 
+                            hole_z_min, 
+                            hole_z_max]
+
+                extreme_coords = np.array([x_coords, y_coords, z_coords])
+                
+                # Check to see if any of these are outside the survey
+                for k in range(6):
+
+                    if mask_checker.not_in_mask(extreme_coords[:,k].reshape(1,3)):
+
+                        # Hole center is within 10 Mpc/h of the survey edge
+                        maximals_table["edge"][i] = 2
+
+                        break
+            #-------------------------------------------------------------------
     ############################################################################
     
     return maximals_table, holes_table

[…]
