### 2022-08-25 868f561
Switched _query_first variables from ijk to pqr label
