+++ python/vast/voidfinder/_voidfinder_cython_find_next.pyx
@@ -259,8 +259,8 @@ cdef FindNextReturnVal find_next_galaxy(DTYPE_F64_t[:,:] hole_center_memview,
             #curr_pqr[idx] = <CELL_ID_t>((temp_hole_center_memview[0,0] - galaxy_tree.coord_min[0,0])/galaxy_tree.dl)
         
         
-        #Note if we keep this code, potential bug whereby the dr *= 1.1
-        #makes us jump more than 1 level at a time
+        # Note if we keep this code, potential bug whereby the dr *= 1.1
+        # makes us jump more than 1 level at a time
         #if curr_pqr[0] != last_pqr[0] or \
         #   curr_pqr[1] != last_pqr[1] or \
         #   curr_pqr[2] != last_pqr[2]:
@@ -2608,7 +2608,7 @@ cdef class Cell_ID_Memory:
 @cython.wraparound(False)
 @cython.cdivision(True)
 @cython.profile(False)
-cdef DistIdxPair _query_first(CELL_ID_t[:,:] reference_point_pqr,
+cpdef DistIdxPair _query_first(CELL_ID_t[:,:] reference_point_pqr,
                               DTYPE_F64_t[:,:] coord_min,
                               DTYPE_F64_t dl,
                               DTYPE_F64_t[:,:] shell_boundaries_xyz,
@@ -3049,33 +3049,36 @@ cdef (DTYPE_INT64_t, DTYPE_INT64_t) _gen_shell(CELL_ID_t[:,:] center_ijk,
     Description
     ===========
     
-    Generate all the possible locations in the "shell" defined by the level parameter.
+    Generate all the possible locations in the "shell" defined by the level 
+    parameter.
     
-    Given a cubic grid structure, level 0 is a single cell, level 1 is a shell of
-    3x3x3 cells minus the 1 interior, level 2 is 5x5x5 minus the 3x3x3 interior, etc.
+    Given a cubic grid structure, level 0 is a single cell, level 1 is a shell 
+    of 3x3x3 cells minus the 1 interior, level 2 is 5x5x5 minus the 3x3x3 
+    interior, etc.
     
     
     Notes
     =====
     
-    Only called once in _query_first() (_voidfinder_cython_find_next) in main_algorithm()
-    (_voidfinder_cython), and now also by _gen_cube() in _query_shell_radius()
+    Only called once in _query_first() (_voidfinder_cython_find_next) in 
+    main_algorithm() (_voidfinder_cython), and now also by _gen_cube() 
+    in _query_shell_radius()
     
-    This means in main_algorithm, we come upon an ijk cell.  We call _query_first on that ijk
-    cell, and within _query_first we stay on that ijk cell and grow outward shells until we find
-    the first neighbor.  Then _gen_shell won't be called again until the next _ijk shell, and instead
-    we will use _gen_cube to find the 2nd 3rd and 4a/b-th neighbors
+    This means in main_algorithm, we come upon an ijk cell.  We call 
+    _query_first on that ijk cell, and within _query_first we stay on that ijk 
+    cell and grow outward shells until we find the first neighbor.  Then 
+    _gen_shell won't be called again until the next _ijk shell, and instead we 
+    will use _gen_cube to find the 2nd 3rd and 4a/b-th neighbors
     
-    ASSUMES THIS WILL ONLY BE CALLED SEQUENTIALLY AND IN-ORDER FOR A SINGLE IJK, WHICH BASED ON ABOVE
-    USAGE WOULD SEEM TO BE TRUE (except for level 0, which should never be called, level 1, then 2, 3, 
-    4,5, etc).  VIOLATING THIS WILL RESULT IN THE cell_ID_mem.max_level_available possibly being 
-    set incorrectly
-    
[…]
