+++ python/vast/voidfinder/hole_combine.py
@@ -1,8 +1,10 @@
+
+
 from astropy.table import Table
 
 import numpy as np
 
-from .table_functions import to_array,to_vector
+from .table_functions import to_array, to_vector
 
 import time
 
