+++ python/vast/test/test_VF.py
@@ -8,8 +8,21 @@ from astropy.table import Table, setdiff, vstack
 
 from vast.voidfinder.constants import c
 from vast.voidfinder import find_voids, filter_galaxies
+from vast.voidfinder.table_functions import to_array
 from vast.voidfinder.multizmask import generate_mask
 from vast.voidfinder.preprocessing import file_preprocess
+from vast.voidfinder._voidfinder_cython import main_algorithm
+from vast.voidfinder._voidfinder_cython_find_next import GalaxyMap, \
+                                                         Cell_ID_Memory, \
+                                                         GalaxyMapCustomDict, \
+                                                         HoleGridCustomDict, \
+                                                         NeighborMemory, \
+                                                         MaskChecker, \
+                                                         find_next_prime, \
+                                                         _query_first
+
+
+
 
 class TestVoidFinder(unittest.TestCase):
 
@@ -21,7 +34,9 @@ class TestVoidFinder(unittest.TestCase):
         TestVoidFinder.wall = None
         TestVoidFinder.dist_limits = None
         TestVoidFinder.mask = None
+        TestVoidFinder.mask_resolution = None
         TestVoidFinder.grid_shape = None
+        TestVoidFinder.galaxy_map = None
 
     def setUp(self):
         '''
@@ -30,7 +45,7 @@ class TestVoidFinder(unittest.TestCase):
         '''
         self.ra_range = np.arange(10, 30, 0.5)
         self.dec_range = np.arange(-10, 10, 0.5)
-        self.redshift_range = np.arange(0, 0.011, 0.0005) # 0.0001
+        self.redshift_range = np.arange(0.0005, 0.011, 0.0005) # 0.0001
 
         RA, DEC, REDSHIFT = np.meshgrid(self.ra_range, 
                                         self.dec_range, 
@@ -40,29 +55,81 @@ class TestVoidFinder(unittest.TestCase):
         self.galaxies_table['ra'] = np.ravel(RA)
         self.galaxies_table['dec'] = np.ravel(DEC)
         self.galaxies_table['redshift'] = np.ravel(REDSHIFT)
+        self.galaxies_table.add_row([20., 0., 0.])
 
         # Shuffle the table (so that the KDtree does not die)
-        rng = np.random.default_rng()
-        self.galaxies_shuffled = Table(rng.permutation(self.galaxies_table))
+        self.rng = np.random.default_rng()
+        self.galaxies_shuffled = Table(self.rng.permutation(self.galaxies_table))
         self.galaxies_shuffled['Rgal'] = c*self.galaxies_shuffled['redshift']/100.
         N_galaxies = len(self.galaxies_shuffled)
 
         # All galaxies will be brighter than the magnitude limit, so that none
         # of them are removed
-        self.galaxies_shuffled['rabsmag'] = 5*np.random.rand(N_galaxies) - 25.1
+        self.galaxies_shuffled['rabsmag'] = -21.0
 
+        # Create test_galaxies file
         self.galaxies_filename = 'test_galaxies.txt'
         self.galaxies_shuffled.write(self.galaxies_filename,
                                      format='ascii.commented_header',
                                      overwrite=True)
 
+        # Convert galaxy coordinates to Cartesian
         self.gal = np.zeros((N_galaxies,3))
         self.gal[:,0] = self.galaxies_shuffled['Rgal']*np.cos(self.galaxies_shuffled['ra']*np.pi/180.)*np.cos(self.galaxies_shuffled['dec']*np.pi/180.)
         self.gal[:,1] = self.galaxies_shuffled['Rgal']*np.sin(self.galaxies_shuffled['ra']*np.pi/180.)*np.cos(self.galaxies_shuffled['dec']*np.pi/180.)
[…]

[…]
