### 2022-09-05 022b6d5
Potential fix for the BufferError causing the GitHub Unit Tests to fail
