### 2022-09-05 3fdc047
Minor adjustments
