### 2022-09-05 72ad155
Minor adjustments, found secondary instance of the BufferError, using same fix
