### 2022-09-05 0afbdf0
Minor adjustments, found 3rd instance of the BufferError, just commenting out the close() calls this time, screw you Python devs
