### 2022-09-05 dfdc805
Minor adjustments, found 4th instance of the BufferError, just commenting out the close() calls this time, screw you Python devs
