+++ python/vast/voidfinder/_voidfinder_cython_find_next.pyx
@@ -2207,10 +2207,18 @@ cdef class GalaxyMap:
     def close(self):
         
         #self.wall_galaxy_buffer.close()
+        # Commented this out due to:
+        # BufferError: cannot close exported pointers exist.
+        # https://stackoverflow.com/questions/53339931/properly-discarding-ctypes-pointers-to-mmap-memory-in-python
+        # https://github.com/ercius/openNCEM/issues/39
         
         os.close(self.wall_galaxies_coords_fd)
         
         #self.galaxy_map_array_buffer.close()
+        # Commented this out due to:
+        # BufferError: cannot close exported pointers exist.
+        # https://stackoverflow.com/questions/53339931/properly-discarding-ctypes-pointers-to-mmap-memory-in-python
+        # https://github.com/ercius/openNCEM/issues/39
         
         os.close(self.gma_fd)
         
