### 2022-09-14 d382bb3
Minor paper changes, added vispy reference
