+++ doc/paper.bib
@@ -116,3 +116,12 @@ archivePrefix = {arXiv},
       url            = "http://www.cosmicvoids.net/"
 }
 
+@article{vispy, 
+    author    = {{Campagnola}, Luke and {Larson}, Eric and {Klein}, Almar and {Hoese}, David and {Siddharth} and {Rossant}, Cyrille and {Griffiths}, Adam and {Rougier}, Nicolas P. and {van Dijk}, Lucas and {Mühlbauer}, Kai and {Taylor}, Alexander and {MSS} and {Lambert, Talley} and {Champandard}, Alex J. and {Hunter}, Max and {Robitaille}, Thomas and {Kaptan}, Mustafa Furkan and {Sales de Andrade}, Elliott and {Czajkowski}, Karl and {Bacchini}, Alessandro and {Favelier}, Guillaume and {Combrisson}, Etienne and {Aye}, Michael and {Harfouche}, Mark and {van Elteren}, Casper and {Gestes}, Cedric and {Bokota}, Grzegorz}, 
+    year      = "2020", 
+    title     = {vispy/vispy: Version 0.11.0}, 
+    doi       = {10.5281/zenodo.4321173}, 
+    publisher = {Zenodo}, 
+    month     = {Dec}
+}
+
