+++ example_scripts/DR7_config.ini
@@ -1,6 +1,5 @@
 [Paths]
 #Currently only accepts FITS input
-<<<<<<< HEAD:Vsquared/data/DR7_config.ini
 Input Catalog    = ../../data/vollim_dr7_cbp_102709.fits
 Survey Name      = DR7
 Output Directory = ./
