### 2022-09-15 a76642d
Miscellaneous V2 fixes (#81)
* Add mask input option, galaxy ID tracking

* Fix bug allowing zero-volume cells in zones

* Add third ZOBOV pruning method

* Fix inclusion of isolated zones

* Update Vsquared documentation

* Fix triangle void IDs

* Remove old Vsquared directory

* Update V2 unit tests

* Fix typo

* Remove merge conflict message
