+++ example_scripts/DR7_config.ini
@@ -1,6 +1,6 @@
 [Paths]
 #Currently only accepts FITS input
-Input Catalog    = vollim_dr7_cbp_102709.fits
+Input Catalog    = ../../data/vollim_dr7_cbp_102709.fits
 Survey Name      = DR7
 Output Directory = ./
 

[…]
