### 2022-09-16 395fa1f
Changed V<sup>2</sup> to V$^2$
