### 2022-09-16 273c6c3
Minor changes to figure captions
