### 2022-09-16 #4033 @kadglass
@dfm Working on it!  The paper is ready to go, and I just tagged the code: the current version is v1.2.0.  We had previously linked the repo to Zenodo (DOI:10.5281/zenodo.7086886), so now we just need to figure out how to update the metadata to match the paper.
