### 2022-09-17 #4033 @dfm
@kadglass — Great! When I navigate to a Zenodo archive that I own, I see the following buttons in the top right of the page:

Clicking on "Edit" let's me change the metadata. Are you able to access this?
